### 2025-03-13 #235 @Jammy2211
PyAutoGalaxy
Submitting Author: (@Jammy2211)
Package Name: PyAutoGalaxy
One-Line Description of Package: Astronomy software for analysing the morphologies and structures of galaxies
Repository Link:  https://github.com/Jammy2211/PyAutoGalaxy
Version submitted:  [2025.1.18.7](https://github.com/Jammy2211/PyAutoGalaxy/tree/2025.1.18.7)
Version accepted: TBD

---

Nearly a century ago, Edwin Hubble famously classified galaxies into three distinct groups: ellipticals, spirals and irregulars [@Hubble1926]. Today, by analysing millions of galaxies with advanced image processing techniques Astronomers have expanded on this picture and revealed the rich diversity of galaxy morphology in both the nearby and distant Universe [@Kormendy2015a; @Vulcani2014; @VanDerWel2012]. PyAutoGalaxy is an open-source Python 3.8 - 3.11 package for analysing the morphologies and structures of large multiwavelength galaxy samples, with core features including fully automated Bayesian model-fitting of galaxy two-dimensional surface brightness profiles, support for dataset and interferometer datasets and comprehensive tools for simulating galaxy images. The software places a focus on big data analysis, including support for hierarchical models that simultaneously fit thousands of galaxies, massively parallel model-fitting and an SQLite3 database that allows large suites of modeling results to be loaded, queried and analysed.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Astronomers studying the morphology of galaxies, and undergraduates / people interested in learning how to do science with galaxies.  The package analyses galaxy images (e.g. from the James Webb Space Telescope) and extracts information on them.

Many: pysersic, GALFIT, ProFit, to name a few.

PyAutoGalaxy supports a more diverse range of ways to analysis galaxy images and has support for Bayesian inference on big data (e.g. sqlite database, graphical modeling) other packages do not.

This package was first submited to astropy years ago, then was to be moved to PyOpenSci, this PR discussion shows that: https://github.com/astropy/astropy.github.com/pull/491

- [ ] Do you wish to automatically submit to the [Journal of Open Source Software][JournalOfOpenSourceSoftware]?

Already Published in JOSS: https://joss.theoj.org/papers/10.21105/joss.04475
