### 2025-03-14 e57f637
fix test -using new fits rather than hdu
