### 2025-03-18 d21d260
added fit fits enum
