### 2025-03-18 4fbd006
extend aggregator examples
