+++ autogalaxy/aggregator/subplot.py
@@ -1,7 +1,7 @@
 from enum import Enum
 
 
-class FitFITS(Enum):
+class FITSFit(Enum):
     """
     The HDUs that can be extracted from the fit.fits file.
     """
