### 2025-03-20 bc8aa3f
hdu_for_output now used in analysis
