### 2025-03-21 54aaf2d
use hdu_list output in autoconf fitsable
