+++ autogalaxy/imaging/model/analysis.py
@@ -1,8 +1,7 @@
-from astropy.io import fits
 import numpy as np
 from typing import Dict, Optional, Tuple
 
-from autoconf.fitsable import hdu_for_output_from
+from autoconf.fitsable import hdu_list_for_output_from
 
 import autofit as af
 import autoarray as aa
@@ -234,15 +233,19 @@ class AnalysisImaging(AnalysisDataset):
 
         paths.save_fits(
             name="dataset",
-            fits=fits.HDUList(hdus=[
-                hdu_for_output_from(arr=self.dataset.mask.astype("float"), ext_name="mask", return_as_primary=True),
-                hdu_for_output_from(arr=self.dataset.data, ext_name="data"),
-                hdu_for_output_from(arr=self.dataset.noise_map, ext_name="noise_map"),
-                hdu_for_output_from(arr=self.dataset.psf, ext_name="psf"),
-                hdu_for_output_from(arr=self.dataset.grids.lp.over_sample_size.native, ext_name="over_sample_size_lp"),
-                hdu_for_output_from(arr=self.dataset.grids.pixelization.over_sample_size.native, ext_name="over_sample_size_pixelization"),
-            ]),
+            fits=hdu_list_for_output_from(
+                values_list=[
+                    self.dataset.mask.astype("float"),
+                    self.dataset.data.native,
+                    self.dataset.noise_map.native,
+                    self.dataset.psf.native,
+                    self.dataset.grids.lp.over_sample_size.native,
+                    self.dataset.grids.pixelization.over_sample_size.native,
+                ],
+                ext_name_list=["mask", "data", "noise_map", "psf", "over_sample_size_lp", "over_sample_size_pixelization"],
+            ),
         )
+
 
     def profile_log_likelihood_function(
         self, instance: af.ModelInstance, paths: Optional[af.DirectoryPaths] = None
