### 2025-03-21 5f06073
interferometer now  uses fitsable
