### 2025-03-21 33959cf
add adapt image output
