+++ autogalaxy/aggregator/ellipse/multipoles.py
@@ -17,7 +17,8 @@ def _multipoles_from(
     """
     Returns a list of `EllipseMultipole` objects from a `PyAutoFit` sqlite database `Fit` object.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
     - The model and its best fit parameters (e.g. `model.json`).
 
@@ -65,7 +66,8 @@ class MultipolesAgg(af.AggBase):
     Interfaces with an `PyAutoFit` aggregator object to create instances of `EllipseMultipole` objects from the results
     of a model-fit.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
     - The model and its best fit parameters (e.g. `model.json`).
 
