+++ autogalaxy/aggregator/galaxies.py
@@ -15,7 +15,8 @@ def _galaxies_from(fit: af.Fit, instance: af.ModelInstance) -> List[Galaxy]:
     """
     Returns a list of `Galaxy` objects from a `PyAutoFit` sqlite database `Fit` object.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
     - The model and its best fit parameters (e.g. `model.json`).
     - The adapt images associated with adaptive galaxy features (`adapt` folder).
@@ -73,7 +74,8 @@ class GalaxiesAgg(af.AggBase):
     Interfaces with an `PyAutoFit` aggregator object to create instances of `Galaxy` objects from the results
     of a model-fit.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
     - The model and its best fit parameters (e.g. `model.json`).
     - The adapt images associated with adaptive galaxy features (`adapt` folder).
