+++ autogalaxy/aggregator/interferometer/interferometer.py
@@ -12,10 +12,11 @@ def _interferometer_from(
     """
     Returns a list of `Interferometer` objects from a `PyAutoFit` sqlite database `Fit` object.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
-    - The interferometer visibilities data as a .fits file (`dataset/data.fits`).
-    - The visibilities noise-map as a .fits file (`dataset/noise_map.fits`).
+    - The interferometer visibilities data as a .fits file (`dataset.fits[hdu=1]`).
+    - The visibilities noise-map as a .fits file (`dataset.fits[hdu=2]`).
     - The uv wavelengths as a .fits file (`dataset/uv_wavelengths.fits`).
     - The real space mask defining the grid of the interferometer for the FFT (`dataset/real_space_mask.fits`).
     - The settings of the `Interferometer` data structure used in the fit (`dataset/settings.json`).
@@ -74,37 +75,38 @@ def _interferometer_from(
 class InterferometerAgg:
     def __init__(self, aggregator: af.Aggregator):
         """
-        Interfaces with an `PyAutoFit` aggregator object to create instances of `Interferometer` objects from the results
-        of a model-fit.
+            Interfaces with an `PyAutoFit` aggregator object to create instances of `Interferometer` objects from the results
+            of a model-fit.
 
-        The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+            The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+        attributes of the fit:
 
-        - The interferometer visibilities data as a .fits file (`dataset/data.fits`).
-        - The visibilities noise-map as a .fits file (`dataset/noise_map.fits`).
-        - The uv wavelengths as a .fits file (`dataset/uv_wavelengths.fits`).
-        - The real space mask defining the grid of the interferometer for the FFT (`dataset/real_space_mask.fits`).
-        - The settings of the `Interferometer` data structure used in the fit (`dataset/settings.json`).
+            - The interferometer visibilities data as a .fits file (`dataset.fits[hdu=1]`).
+            - The visibilities noise-map as a .fits file (`dataset.fits[hdu=2]`).
+            - The uv wavelengths as a .fits file (`dataset/uv_wavelengths.fits`).
+            - The real space mask defining the grid of the interferometer for the FFT (`dataset/real_space_mask.fits`).
+            - The settings of the `Interferometer` data structure used in the fit (`dataset/settings.json`).
 
-        The `aggregator` contains the path to each of these files, and they can be loaded individually. This class
-        can load them all at once and create an `Interferometer` object via the `_interferometer_from` method.
+            The `aggregator` contains the path to each of these files, and they can be loaded individually. This class
+            can load them all at once and create an `Interferometer` object via the `_interferometer_from` method.
 
-        This class's methods returns generators which create the instances of the `Interferometer` objects. This ensures
-        that large sets of results can be efficiently loaded from the hard-disk and do not require storing all
-        `Interferometer` instances in the memory at once.
+            This class's methods returns generators which create the instances of the `Interferometer` objects. This ensures
+            that large sets of results can be efficiently loaded from the hard-disk and do not require storing all
+            `Interferometer` instances in the memory at once.
 
-        For example, if the `aggregator` contains 3 model-fits, this class can be used to create a generator which
-        creates instances of the corresponding 3 `Interferometer` objects.
+            For example, if the `aggregator` contains 3 model-fits, this class can be used to create a generator which
+            creates instances of the corresponding 3 `Interferometer` objects.
 
[…]
