+++ autogalaxy/aggregator/imaging/imaging.py
@@ -1,6 +1,8 @@
 from functools import partial
 from typing import List
 
+from autoconf.fitsable import ndarray_via_hdu_from
+
 import autofit as af
 import autoarray as aa
 
@@ -11,13 +13,14 @@ def _imaging_from(
     """
     Returns a list of `Imaging` objects from a `PyAutoFit` sqlite database `Fit` object.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
-    - The imaging data as a .fits file (`dataset/data.fits`).
-    - The noise-map as a .fits file (`dataset/noise_map.fits`).
-    - The point spread function as a .fits file (`dataset/psf.fits`).
+    - The mask used to mask the `Imaging` data structure in the fit (`dataset.fits[hdu=0]`).
+    - The imaging data as a .fits file (`dataset.fits[hdu=1]`).
+    - The noise-map as a .fits file (`dataset.fits[hdu=2]`).
+    - The point spread function as a .fits file (`dataset.fits[hdu=3]`).
     - The settings of the `Imaging` data structure used in the fit (`dataset/settings.json`).
-    - The mask used to mask the `Imaging` data structure in the fit (`dataset/mask.fits`).
 
     Each individual attribute can be loaded from the database via the `fit.value()` method.
 
@@ -39,16 +42,24 @@ def _imaging_from(
     dataset_list = []
 
     for fit in fit_list:
-        data = aa.Array2D.from_primary_hdu(
-            primary_hdu=fit.value(name="dataset.data")[0]
+        header = aa.Header(header_sci_obj=fit.value(name="dataset")[0].header)
+        pixel_scales = (
+            header.header_sci_obj["PIXSCAY"],
+            header.header_sci_obj["PIXSCAX"],
         )
-        noise_map = aa.Array2D.from_primary_hdu(
-            primary_hdu=fit.value(name="dataset.noise_map")[0]
-        )
-        try:
-            psf = aa.Kernel2D.from_primary_hdu(
-                primary_hdu=fit.value(name="dataset.psf")[0]
+
+        def values_from(hdu: int, cls):
+            return cls.no_mask(
+                values=ndarray_via_hdu_from(fit.value(name="dataset")[hdu]),
+                pixel_scales=pixel_scales,
+                header=header,
             )
+
+        data = values_from(hdu=1, cls=aa.Array2D)
+        noise_map = values_from(hdu=2, cls=aa.Array2D)
+
+        try:
+            psf = values_from(hdu=3, cls=aa.Kernel2D)
         except TypeError:
             psf = None
 
@@ -59,22 +70,21 @@ def _imaging_from(
             check_noise_map=False,
         )
 
-        mask = aa.Mask2D.from_primary_hdu(primary_hdu=fit.value(name="dataset.mask")[0])
+        mask = aa.Mask2D(
+            mask=ndarray_via_hdu_from(fit.value(name="dataset")[0]),
+            pixel_scales=pixel_scales,
+        )
 
         dataset = dataset.apply_mask(mask=mask)
 
         try:
-            over_sample_size_lp = aa.Array2D.from_primary_hdu(
-                primary_hdu=fit.value(name="dataset.over_sample_size_lp")[0]
-            ).native
+            over_sample_size_lp = values_from(hdu=4, cls=aa.Array2D).native
             over_sample_size_lp = over_sample_size_lp.apply_mask(mask=mask)
         except TypeError:
             over_sample_size_lp = 1
 
         try:
-            over_sample_size_pixelization = aa.Array2D.from_primary_hdu(
[…]
