+++ autogalaxy/aggregator/dataset_model.py
@@ -14,7 +14,8 @@ def _dataset_model_from(
     """
     Returns a `DatasetModel` object from a `PyAutoFit` sqlite database `Fit` object.
 
-    The results of a model-fit can be stored in a sqlite database, including the following attributes of the fit:
+    The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+    attributes of the fit:
 
     - The model and its best fit parameters (e.g. `model.json`).
 

[…]
