+++ autogalaxy/interferometer/model/analysis.py
@@ -245,14 +245,16 @@ class AnalysisInterferometer(AnalysisDataset):
                     self.dataset.uv_wavelengths,
                 ],
                 ext_name_list=["mask", "data", "noise_map", "uv_wavelengths"],
+                header_dict={
+                    "PIXSCAY": self.dataset.pixel_scales[0],
+                    "PIXSCAX": self.dataset.pixel_scales[1],
+                },
             ),
         )
 
         paths.save_json(
             "transformer_class", to_dict(self.dataset.transformer.__class__), "dataset"
         )
-
-        ggg
 
     def profile_log_likelihood_function(
         self, instance: af.ModelInstance, paths: Optional[af.DirectoryPaths] = None
