### 2025-03-21 1803580
pixel scale header uses mask property
