+++ autogalaxy/analysis/analysis/dataset.py
@@ -194,10 +194,7 @@ class AnalysisDataset(Analysis):
                     + values_list,
                     ext_name_list=["mask"]
                     + list(self.adapt_images.galaxy_name_image_dict.keys()),
-                    header_dict={
-                        "PIXSCAY": self.dataset.pixel_scales[0],
-                        "PIXSCAX": self.dataset.pixel_scales[1],
-                    },
+                    header_dict=self.dataset.mask.pixel_scale_header
                 ),
             )
 
