### 2025-03-21 999179c
moved some aggregaotr loading to agg_util
