+++ autogalaxy/aggregator/galaxies.py
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 def _galaxies_from(fit: af.Fit, instance: af.ModelInstance) -> List[Galaxy]:
     """
-    Returns a list of `Galaxy` objects from a `PyAutoFit` sqlite database `Fit` object.
+    Returns a list of `Galaxy` objects from a `PyAutoFit` loaded directory `Fit` or sqlite database `Fit` object.
 
     The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
     attributes of the fit:
@@ -33,7 +33,8 @@ def _galaxies_from(fit: af.Fit, instance: af.ModelInstance) -> List[Galaxy]:
     Parameters
     ----------
     fit
-        A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry in a sqlite database.
+        A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry which has been loaded from 
+        an output directory or from an sqlite database..
     instance
         A manual instance that overwrites the max log likelihood instance in fit (e.g. for drawing the instance
         randomly from the PDF).
@@ -113,7 +114,8 @@ class GalaxiesAgg(af.AggBase):
         Parameters
         ----------
         fit
-            A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry in a sqlite database.
+            A `PyAutoFit` `Fit` object which contains the results of a model-fit as an entry which has been loaded from 
+        an output directory or from an sqlite database..
         instance
             A manual instance that overwrites the max log likelihood instance in fit (e.g. for drawing the instance
             randomly from the PDF).
