### 2025-03-21 9ccf85a
update ellipse analysis
