### 2025-03-21 fde289f
fix aggregator via IndexError Exception
