+++ autogalaxy/imaging/model/analysis.py
@@ -204,19 +204,17 @@ class AnalysisImaging(AnalysisDataset):
          Before the non-linear search begins, this routine saves attributes of the `Analysis` object to the `files`
          folder such that they can be loaded after the analysis using PyAutoFit's database and aggregator tools.
 
+         It outputs the following attributes of the dataset:
+
+         - The mask applied to the dataset, in the `PrimaryHDU` of `dataset.fits`.
+         - The imaging dataset as `dataset.fits` (data / noise-map / psf / over sampler / etc.).
+
          For this analysis, it uses the `AnalysisDataset` object's method to output the following:
 
-         - The imaging dataset (data / noise-map / psf / over sampler / etc.).
-         - The mask applied to the dataset.
          - The settings associated with the inversion.
          - The settings associated with the pixelization.
          - The Cosmology.
-         - The adapt image's model image and galaxy images, if used.
-
-         This function also outputs attributes specific to an imaging dataset:
-
-        - Its PSF.
-        - Its mask.
+         - The adapt image's model image and galaxy images, as `adapt_images.fits`, if used.
 
          It is common for these attributes to be loaded by many of the template aggregator functions given in the
          `aggregator` modules. For example, when using the database tools to perform a fit, the default behaviour is for
