### 2025-03-21 46f480d
update imaging plotter interface
