+++ test_autogalaxy/imaging/model/test_plotter_interface_imaging.py
@@ -53,17 +53,17 @@ def test__fit_imaging(
 
     assert path.join(plot_path, "subplot_fit.png") in plot_patch.paths
 
-    image = ag.util.array_2d.numpy_array_2d_via_fits_from(
+    image = ag.ndarray_via_fits_from(
         file_path=path.join(plot_path, "fit.fits"), hdu=0
     )
 
-    assert image.shape == (5, 5)
+    assert image.shape == (7, 7)
 
-    image = ag.util.array_2d.numpy_array_2d_via_fits_from(
+    image = ag.ndarray_via_fits_from(
         file_path=path.join(plot_path, "model_galaxy_images.fits"), hdu=0
     )
 
-    assert image.shape == (5, 5)
+    assert image.shape == (7, 7)
 
 
 def test__fit_imaging_combined(
