+++ autogalaxy/interferometer/model/analysis.py
@@ -205,31 +205,31 @@ class AnalysisInterferometer(AnalysisDataset):
 
     def save_attributes(self, paths: af.DirectoryPaths):
         """
-         Before the model-fit begins, this routine saves attributes of the `Analysis` object to the `files` folder
-         such that they can be loaded after the analysis using PyAutoFit's database and aggregator tools.
+        Before the model-fit begins, this routine saves attributes of the `Analysis` object to the `files` folder
+        such that they can be loaded after the analysis using PyAutoFit's database and aggregator tools.
 
-         It outputs the following attributes of the dataset:
+        It outputs the following attributes of the dataset:
 
-         - The real space mask applied to the dataset, in the `PrimaryHDU` of `dataset.fits`.
-         - The interferometer dataset as `dataset.fits` (data / noise-map / uv_wavelengths).
+        - The real space mask applied to the dataset, in the `PrimaryHDU` of `dataset.fits`.
+        - The interferometer dataset as `dataset.fits` (data / noise-map / uv_wavelengths).
 
-         For this analysis, it uses the `AnalysisDataset` object's method to output the following:
+        For this analysis, it uses the `AnalysisDataset` object's method to output the following:
 
-         - The settings associated with the inversion.
-         - The settings associated with the pixelization.
-         - The Cosmology.
-         - The adapt image's model image and galaxy images, as `adapt_images.fits`, if used.
+        - The settings associated with the inversion.
+        - The settings associated with the pixelization.
+        - The Cosmology.
+        - The adapt image's model image and galaxy images, as `adapt_images.fits`, if used.
 
-         It is common for these attributes to be loaded by many of the template aggregator functions given in the
-         `aggregator` modules. For example, when using the database tools to perform a fit, the default behaviour is for
-         the dataset, settings and other attributes necessary to perform the fit to be loaded via the pickle files
-         output by this function.
+        It is common for these attributes to be loaded by many of the template aggregator functions given in the
+        `aggregator` modules. For example, when using the database tools to perform a fit, the default behaviour is for
+        the dataset, settings and other attributes necessary to perform the fit to be loaded via the pickle files
+        output by this function.
 
-         Parameters
-         ----------
-         paths
-             The paths object which manages all paths, e.g. where the non-linear search outputs are stored, visualization,
-             and the pickled objects used by the aggregator output by this function.
+        Parameters
+        ----------
+        paths
+            The paths object which manages all paths, e.g. where the non-linear search outputs are stored, visualization,
+            and the pickled objects used by the aggregator output by this function.
         """
         super().save_attributes(paths=paths)
 
@@ -243,7 +243,7 @@ class AnalysisInterferometer(AnalysisDataset):
                     self.dataset.uv_wavelengths,
                 ],
                 ext_name_list=["mask", "data", "noise_map", "uv_wavelengths"],
-                header_dict=self.dataset.real_space_mask.pixel_scale_header
+                header_dict=self.dataset.real_space_mask.pixel_scale_header,
             ),
         )
 
