### 2025-03-21 825544b
move save attributes stuff to plotter interface
