+++ autogalaxy/analysis/plotter_interface.py
@@ -184,7 +184,7 @@ class PlotterInterface:
                 values_list=[grid.mask.astype("float")]
                 + [galaxy.image_2d_from(grid=grid) for galaxy in galaxies],
                 ext_name_list=["mask"] + [f"galaxy_{i}" for i in range(len(galaxies))],
-                header_dict=grid.mask.pixel_scale_header,
+                header_dict=grid.mask.header_dict,
             )
 
             hdu_list.writeto(self.image_path / "galaxy_images.fits", overwrite=True)
@@ -258,3 +258,20 @@ class PlotterInterface:
             adapt_plotter.subplot_adapt_images(
                 adapt_galaxy_name_image_dict=adapt_images.galaxy_image_dict
             )
+            
+        values_list = [
+            adapt_images.galaxy_name_image_dict[name].native
+            for name in adapt_images.galaxy_name_image_dict.keys()
+        ]
+
+        hdu_list = hdu_list_for_output_from(
+                values_list=[
+                                adapt_images.mask.astype("float"),
+                            ]
+                            + values_list,
+                ext_name_list=["mask"]
+                              + list(adapt_images.galaxy_name_image_dict.keys()),
+                header_dict=adapt_images.mask.header_dict,
+            )
+    
+        hdu_list.writeto(self.image_path / "adapt_images.fits", overwrite=True)
