+++ test_autogalaxy/interferometer/model/test_plotter_interface_interferometer.py
@@ -22,6 +22,12 @@ def test__interferometer(interferometer_7, include_2d_all, plot_path, plot_patch
 
     assert path.join(plot_path, "subplot_dataset.png") in plot_patch.paths
 
+    image = ag.ndarray_via_fits_from(
+        file_path=path.join(plot_path, "dataset.fits"), hdu=0
+    )
+
+    assert image.shape == (7, 7)
+
 
 def test__fit_interferometer(
     interferometer_7,
