+++ autogalaxy/ellipse/model/analysis.py
@@ -197,7 +197,7 @@ class AnalysisEllipse(af.Analysis):
                     "data",
                     "noise_map",
                 ],
-                header_dict=self.dataset.mask.pixel_scale_header,
+                header_dict=self.dataset.mask.header_dict,
             ),
         )
 
