+++ autogalaxy/ellipse/model/plotter_interface.py
@@ -1,5 +1,6 @@
 from typing import List
-from os import path
+
+from autoconf.fitsable import hdu_list_for_output_from
 
 import autoarray as aa
 import autoarray.plot as aplt
@@ -42,6 +43,22 @@ class PlotterInterfaceEllipse(PlotterInterface):
 
         if should_plot("subplot_dataset"):
             dataset_plotter.subplot_dataset()
+
+        hdu_list = hdu_list_for_output_from(
+                values_list=[
+                    dataset.mask.astype("float"),
+                    dataset.data.native,
+                    dataset.noise_map.native,
+                ],
+                ext_name_list=[
+                    "mask",
+                    "data",
+                    "noise_map",
+                ],
+                header_dict=dataset.mask.header_dict,
+            )
+
+        hdu_list.writeto(self.image_path / "dataset.fits", overwrite=True)
 
     def fit_ellipse(
         self,
