### 2025-03-24 f149cc9
fix imaging aggregator tests
