+++ autogalaxy/aggregator/interferometer/interferometer.py
@@ -43,9 +43,7 @@ def _interferometer_from(
     dataset_list = []
 
     for fit in fit_list:
-        real_space_mask, header = agg_util.mask_header_pixel_scales_origin_from(
-            fit=fit
-        )
+        real_space_mask, header = agg_util.mask_header_from(fit=fit)
 
         data = aa.Visibilities(
             visibilities=fit.value(name="dataset")[1].data.astype("float")
@@ -73,38 +71,38 @@ def _interferometer_from(
 class InterferometerAgg:
     def __init__(self, aggregator: af.Aggregator):
         """
-            Interfaces with an `PyAutoFit` aggregator object to create instances of `Interferometer` objects from the results
-            of a model-fit.
+        Interfaces with an `PyAutoFit` aggregator object to create instances of `Interferometer` objects from the results
+        of a model-fit.
 
-            The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
+        The results of a model-fit can be loaded from hard-disk or stored in a sqlite database, including the following
         attributes of the fit:
 
-            - The interferometer visibilities data as a .fits file (`dataset.fits[hdu=1]`).
-            - The visibilities noise-map as a .fits file (`dataset.fits[hdu=2]`).
-            - The uv wavelengths as a .fits file (`dataset/uv_wavelengths.fits`).
-            - The real space mask defining the grid of the interferometer for the FFT (`dataset/real_space_mask.fits`).
-            - The settings of the `Interferometer` data structure used in the fit (`dataset/settings.json`).
+        - The interferometer visibilities data as a .fits file (`dataset.fits[hdu=1]`).
+        - The visibilities noise-map as a .fits file (`dataset.fits[hdu=2]`).
+        - The uv wavelengths as a .fits file (`dataset/uv_wavelengths.fits`).
+        - The real space mask defining the grid of the interferometer for the FFT (`dataset/real_space_mask.fits`).
+        - The settings of the `Interferometer` data structure used in the fit (`dataset/settings.json`).
 
-            The `aggregator` contains the path to each of these files, and they can be loaded individually. This class
-            can load them all at once and create an `Interferometer` object via the `_interferometer_from` method.
+        The `aggregator` contains the path to each of these files, and they can be loaded individually. This class
+        can load them all at once and create an `Interferometer` object via the `_interferometer_from` method.
 
-            This class's methods returns generators which create the instances of the `Interferometer` objects. This ensures
-            that large sets of results can be efficiently loaded from the hard-disk and do not require storing all
-            `Interferometer` instances in the memory at once.
+        This class's methods returns generators which create the instances of the `Interferometer` objects. This ensures
+        that large sets of results can be efficiently loaded from the hard-disk and do not require storing all
+        `Interferometer` instances in the memory at once.
 
-            For example, if the `aggregator` contains 3 model-fits, this class can be used to create a generator which
-            creates instances of the corresponding 3 `Interferometer` objects.
+        For example, if the `aggregator` contains 3 model-fits, this class can be used to create a generator which
+        creates instances of the corresponding 3 `Interferometer` objects.
 
-            If multiple `Interferometer` objects were fitted simultaneously via analysis summing, the `fit.child_values()`
-            method is instead used to load lists of the data, noise-map, PSF and mask and combine them into a list of
-            `Interferometer` objects.
+        If multiple `Interferometer` objects were fitted simultaneously via analysis summing, the `fit.child_values()`
+        method is instead used to load lists of the data, noise-map, PSF and mask and combine them into a list of
+        `Interferometer` objects.
 
-            This can be done manually, but this object provides a more concise API.
+        This can be done manually, but this object provides a more concise API.
 
-            Parameters
[…]
