### 2025-03-24 4726e75
fix adapt image test
