### 2025-04-02 4d9af45
aggregator uses enum
