+++ test_autogalaxy/operate/test_image.py
@@ -9,14 +9,14 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 def test__blurred_image_2d_from(
-    grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3, convolver_7x7
+    grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3,
 ):
     lp = ag.lp.Sersic(intensity=1.0)
 
     image_2d = lp.image_2d_from(grid=grid_2d_7x7)
     blurring_image_2d = lp.image_2d_from(grid=blurring_grid_2d_7x7)
 
-    blurred_image_2d_manual = convolver_7x7.convolve_image(
+    blurred_image_2d_manual = psf_3x3.convolve_image(
         image=image_2d, blurring_image=blurring_image_2d
     )
 
@@ -31,7 +31,7 @@ def test__blurred_image_2d_from(
     lp_blurred_image_2d = lp.blurred_image_2d_from(
         grid=grid_2d_7x7,
         blurring_grid=blurring_grid_2d_7x7,
-        convolver=convolver_7x7,
+        psf=psf_3x3
     )
 
     assert blurred_image_2d_manual.native == pytest.approx(
@@ -55,7 +55,7 @@ def test__blurred_image_2d_from(
         grid=grid_2d_7x7, psf=psf_3x3, blurring_grid=blurring_grid_2d_7x7
     )
 
-    blurred_image_2d_manual_not_operated = convolver_7x7.convolve_image(
+    blurred_image_2d_manual_not_operated = psf_3x3.convolve_image(
         image=image_2d_not_operated,
         blurring_image=blurring_image_2d_not_operated,
     )
@@ -167,7 +167,7 @@ def test__visibilities_from_grid_and_transformer(grid_2d_7x7, transformer_7x7_7)
 
 
 def test__blurred_image_2d_list_from(
-    grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3, convolver_7x7
+    grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3,
 ):
     lp_0 = ag.lp.Gaussian(intensity=1.0)
     lp_1 = ag.lp.Gaussian(intensity=2.0)
@@ -196,7 +196,7 @@ def test__blurred_image_2d_list_from(
     blurred_image_2d_list = gal.blurred_image_2d_list_from(
         grid=grid_2d_7x7,
         blurring_grid=blurring_grid_2d_7x7,
-        convolver=convolver_7x7,
+        psf=psf_3x3
     )
 
     assert blurred_image_2d_list[0].native == pytest.approx(
@@ -226,7 +226,7 @@ def test__blurred_image_2d_list_from(
     blurred_image_2d_list = gal.blurred_image_2d_list_from(
         grid=grid_2d_7x7,
         blurring_grid=blurring_grid_2d_7x7,
-        convolver=convolver_7x7,
+        psf=psf_3x3
     )
 
     assert blurred_image_2d_list[0].native == pytest.approx(
@@ -297,7 +297,7 @@ def test__visibilities_list_from(grid_2d_7x7, transformer_7x7_7):
 
 
 def test__galaxy_blurred_image_2d_dict_from(
-    grid_2d_7x7, blurring_grid_2d_7x7, convolver_7x7
+    grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3
 ):
     lp_0 = ag.lp.Sersic(intensity=1.0)
 
@@ -312,13 +312,13 @@ def test__galaxy_blurred_image_2d_dict_from(
 
     blurred_image_2d_list = galaxies.blurred_image_2d_list_from(
[…]
