+++ autogalaxy/operate/deflections.py
@@ -99,16 +99,10 @@ def one_step(r, _, theta, fun, fun_dr):
 @partial(jit, static_argnums=(4,))
 def step_r(r, theta, fun, fun_dr, N=20):
     one_step_partial = jax.tree_util.Partial(
-        one_step,
-        theta=theta,
-        fun=fun,
-        fun_dr=fun_dr
+        one_step, theta=theta, fun=fun, fun_dr=fun_dr
     )
     new_r = jax.lax.scan(one_step_partial, r, xs=np.arange(N))[0]
-    return np.stack([
-        new_r * np.sin(theta),
-        new_r * np.cos(theta)
-    ]).T
+    return np.stack([new_r * np.sin(theta), new_r * np.cos(theta)]).T
 
 
 class OperateDeflections:
@@ -129,7 +123,7 @@ class OperateDeflections:
 
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs):
         raise NotImplementedError
-    
+
     def deflections_yx_scalar(self, y, x, pixel_scales):
         if not use_jax:
             return
@@ -140,13 +134,13 @@ class OperateDeflections:
                 y=y.reshape(1),
                 x=x.reshape(1),
                 shape_native=(1, 1),
-                pixel_scales=pixel_scales
+                pixel_scales=pixel_scales,
             )
             return self.deflections_yx_2d_from(g).squeeze()
 
     def __eq__(self, other):
         return self.__dict__ == other.__dict__ and self.__class__ is other.__class__
-    
+
     def __hash__(self):
         return hash(repr(self))
 
@@ -754,85 +748,73 @@ class OperateDeflections:
             return
         else:
             return np.stack(
-                jax.jacfwd(
-                    self.deflections_yx_scalar,
-                    argnums=(0, 1)
-                )(y, x, pixel_scales)
+                jax.jacfwd(self.deflections_yx_scalar, argnums=(0, 1))(
+                    y, x, pixel_scales
+                )
             )
-    
+
     def jacobian_stack_vector(self, y, x, pixel_scales):
         if not use_jax:
             return
         else:
             return np.vectorize(
-                jax.tree_util.Partial(
-                    self.jacobian_stack,
-                    pixel_scales=pixel_scales
-                ),
-                signature='(),()->(i,i)'
+                jax.tree_util.Partial(self.jacobian_stack, pixel_scales=pixel_scales),
+                signature="(),()->(i,i)",
             )(y, x)
-        
+
     def convergence_mag_shear_yx(self, y, x):
         J = self.jacobian_stack_vector(y, x, 0.05)
         K = 0.5 * (J[..., 0, 0] + J[..., 1, 1])
         mag_shear = 0.5 * np.sqrt(
-            (J[..., 0, 1] + J[..., 1, 0])**2 + (J[..., 0, 0] - J[..., 1, 1])**2
+            (J[..., 0, 1] + J[..., 1, 0]) ** 2 + (J[..., 0, 0] - J[..., 1, 1]) ** 2
         )
         return K, mag_shear
-    
+
     @partial(jit, static_argnums=(0,))
     def tangential_eigen_value_yx(self, y, x):
         K, mag_shear = self.convergence_mag_shear_yx(y, x)
         return 1 - K - mag_shear
-    
+
     @partial(jit, static_argnums=(0, 3))
     def tangential_eigen_value_rt(self, r, theta, centre=(0.0, 0.0)):
         y = r * np.sin(theta) + centre[0]
         x = r * np.cos(theta) + centre[1]
         return self.tangential_eigen_value_yx(y, x)
-    
+
     @partial(jit, static_argnums=(0, 3))
[…]
