+++ autogalaxy/profiles/light/snr/abstract.py
@@ -84,7 +84,9 @@ class LightProfileSNR:
 
         image_2d = self.image_2d_from(grid=grid)
         if psf is not None:
-            image_2d = psf.convolve_image_no_blurring(image=image_2d)
+            image_2d = psf.convolve_image_no_blurring(
+                image=image_2d, mask=image_2d.mask
+            )
 
         brightest_value = np.max(image_2d)
 
