+++ test_autogalaxy/profiles/light/linear/test_abstract.py
@@ -33,15 +33,11 @@ def test__mapping_matrix_from(grid_2d_7x7, blurring_grid_2d_7x7, psf_3x3):
     )
 
     lp_0_blurred_image = lp_0.blurred_image_2d_from(
-        grid=grid_2d_7x7,
-        blurring_grid=blurring_grid_2d_7x7,
-        psf=psf_3x3
+        grid=grid_2d_7x7, blurring_grid=blurring_grid_2d_7x7, psf=psf_3x3
     )
 
     lp_1_blurred_image = lp_1.blurred_image_2d_from(
-        grid=grid_2d_7x7,
-        blurring_grid=blurring_grid_2d_7x7,
-        psf=psf_3x3
+        grid=grid_2d_7x7, blurring_grid=blurring_grid_2d_7x7, psf=psf_3x3
     )
 
     assert lp_linear_obj_func_list.operated_mapping_matrix_override[
