+++ autogalaxy/profiles/mass/total/power_law.py
@@ -61,7 +61,6 @@ class PowerLaw(PowerLawCore):
 
     @aa.grid_dec.to_vector_yx
     @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs):
         """
         Calculate the deflection angles on a grid of (y,x) arc-second coordinates.
@@ -165,7 +164,6 @@ class PowerLawSph(PowerLaw):
 
     @aa.grid_dec.to_vector_yx
     @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs):
         eta = self.radial_grid_from(grid=grid, **kwargs)
         deflection_r = (
