+++ autogalaxy/profiles/light/linear/shapelets/exponential.py
@@ -53,7 +53,6 @@ class ShapeletExponential(lp.ShapeletExponential, LightProfileLinear):
     @aa.grid_dec.to_array
     @check_operated_only
     @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
     def image_2d_from(
         self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None, **kwargs
     ) -> np.ndarray:
