+++ autogalaxy/profiles/light/standard/shapelets/polar.py
@@ -64,7 +64,6 @@ class ShapeletPolar(AbstractShapelet):
     @aa.grid_dec.to_array
     @check_operated_only
     @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
     def image_2d_from(
         self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None, **kwargs
     ) -> np.ndarray:
