### 2025-04-05 a6cb5b8
fixes to geometry profiles when testing light profiles
