### 2025-04-05 ac4878e
fix test_sersic_core
