+++ autogalaxy/profiles/mass/stellar/sersic_core.py
@@ -78,7 +78,9 @@ class SersicCore(Sersic):
                 jnp.power(
                     jnp.add(
                         1,
-                        jnp.power(jnp.divide(self.radius_break, grid_radii), self.alpha),
+                        jnp.power(
+                            jnp.divide(self.radius_break, grid_radii), self.alpha
+                        ),
                     ),
                     (self.gamma / self.alpha),
                 ),
