+++ autogalaxy/profiles/mass/stellar/sersic.py
@@ -1,13 +1,6 @@
-import os
+import jax.numpy as jnp
+import numpy as np
 
-if os.environ.get("USE_JAX", "0") == "1":
-    USING_JAX = True
-    import jax.numpy as np
-else:
-    USING_JAX = False
-    import numpy as np
-
-import copy
 from scipy.integrate import quad
 from typing import List, Tuple
 
