+++ test_autogalaxy/profiles/light/standard/test_eff.py
@@ -59,7 +59,7 @@ def test__image_2d_from():
     image_elliptical = elliptical.image_2d_from(grid=grid)
     image_spherical = spherical.image_2d_from(grid=grid)
 
-    assert image_elliptical == pytest.approx(image_spherical, 1.0e-4)
+    assert image_elliptical.array == pytest.approx(image_spherical.array, 1.0e-4)
 
 
 def test__half_light_radius():
