+++ autogalaxy/profiles/light/standard/eff.py
@@ -1,3 +1,4 @@
+import jax.numpy as jnp
 import numpy as np
 from typing import Optional, Tuple
 
