+++ autogalaxy/profiles/light/standard/eff.py
@@ -1,4 +1,3 @@
-import jax.numpy as jnp
 import numpy as np
 from typing import Optional, Tuple
 
