### 2025-04-05 109c166
fix a standard light profile test by altering autoarray
