+++ test_autogalaxy/config/non_linear.yaml
@@ -105,4 +105,3 @@ nest:
     updates:
       iterations_per_update: 2500
       remove_state_files_at_end: true
->>>>>>> cb2f01c19d3b3421c74df3425ffa32cdff82aebc
