+++ autogalaxy/profiles/light/standard/shapelets/polar.py
@@ -1,5 +1,7 @@
+import jax.numpy as jnp
 import numpy as np
-from scipy.special import factorial, genlaguerre
+from scipy.special import genlaguerre
+from jax.scipy.special import factorial
 from typing import Optional, Tuple
 
 import autoarray as aa
@@ -84,28 +86,28 @@ class ShapeletPolar(AbstractShapelet):
             The image of the Polar Shapelet evaluated at every (y,x) coordinate on the transformed grid.
         """
 
-        laguerre = genlaguerre(n=(self.n - np.abs(self.m)) / 2.0, alpha=np.abs(self.m))
+        laguerre = genlaguerre(n=(self.n - jnp.abs(self.m)) / 2.0, alpha=jnp.abs(self.m))
 
         const = (
-            ((-1) ** ((self.n - np.abs(self.m)) // 2))
-            * np.sqrt(
-                factorial((self.n - np.abs(self.m)) // 2)
-                / factorial((self.n + np.abs(self.m)) // 2)
+            ((-1) ** ((self.n - jnp.abs(self.m)) // 2))
+            * jnp.sqrt(
+                factorial((self.n - jnp.abs(self.m)) // 2)
+                / factorial((self.n + jnp.abs(self.m)) // 2)
             )
             / self.beta
-            / np.sqrt(np.pi)
+            / jnp.sqrt(jnp.pi)
         )
 
-        rsq = (grid[:, 0] ** 2 + grid[:, 1] ** 2) / self.beta**2
-        theta = np.arctan2(grid[:, 1], grid[:, 0])
-        radial = rsq ** (abs(self.m / 2.0)) * np.exp(-rsq / 2.0) * laguerre(rsq)
+        rsq = (grid.array[:, 0] ** 2 + grid.array[:, 1] ** 2) / self.beta**2
+        theta = jnp.arctan2(grid.array[:, 1], grid.array[:, 0])
+        radial = rsq ** (abs(self.m / 2.0)) * jnp.exp(-rsq / 2.0) * laguerre(rsq)
 
         if self.m == 0:
             azimuthal = 1
         elif self.m > 0:
-            azimuthal = np.sin((-1) * self.m * theta)
+            azimuthal = jnp.sin((-1) * self.m * theta)
         else:
-            azimuthal = np.cos((-1) * self.m * theta)
+            azimuthal = jnp.cos((-1) * self.m * theta)
 
         return const * radial * azimuthal
 
