+++ test_autogalaxy/profiles/light/shapelets/test_polar.py
@@ -26,7 +26,7 @@ def test__elliptical__image_2d_from():
 
     image = shapelet.image_2d_from(grid=ag.Grid2DIrregular([[0.0, 1.0], [0.5, 0.25]]))
 
-    assert image == pytest.approx(np.array([0.0, -0.33177]), 1e-4)
+    assert image == pytest.approx(np.array([0.0, -0.33177]), abs=1e-4)
 
     shapelet = ag.lp_linear.ShapeletPolar(
         n=2, m=0, centre=(0.0, 0.0), ell_comps=(0.5, 0.7), beta=1.0
@@ -34,4 +34,4 @@ def test__elliptical__image_2d_from():
 
     image = shapelet.image_2d_from(grid=ag.Grid2DIrregular([[0.0, 1.0], [0.5, 0.25]]))
 
-    assert image == pytest.approx(np.array([0.0, -0.33177]), 1e-4)
+    assert image == pytest.approx(np.array([0.0, -0.33177]), abs=1e-4)
