### 2025-04-05 2e661b3
fix exponmential shjapelet
