+++ autogalaxy/profiles/light/standard/shapelets/exponential.py
@@ -1,6 +1,7 @@
 import jax.numpy as jnp
 import numpy as np
-from scipy.special import factorial, genlaguerre
+from scipy.special import genlaguerre
+from jax.scipy.special import factorial
 from typing import Optional, Tuple
 
 import autoarray as aa
@@ -85,8 +86,8 @@ class ShapeletExponential(AbstractShapelet):
             The image of the Exponential Shapelet evaluated at every (y,x) coordinate on the transformed grid.
         """
 
-        radial = (grid[:, 0] ** 2 + grid[:, 1] ** 2) / self.beta
-        theta = jnp.arctan(grid[:, 1] / grid[:, 0])
+        radial = (grid.array[:, 0] ** 2 + grid.array[:, 1] ** 2) / self.beta
+        theta = jnp.arctan(grid.array[:, 1] / grid.array[:, 0])
 
         prefactor = (
             1.0
