### 2025-04-05 a563bd0
linear light profile tests
