### 2025-04-05 fe9d3ae
fix power law core
