+++ test_autogalaxy/profiles/mass/total/test_isothermal.py
@@ -40,7 +40,7 @@ def test__deflections_yx_2d_from():
     spherical = ag.mp.IsothermalSph(centre=(1.1, 1.1), einstein_radius=3.0)
 
     assert elliptical.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        spherical.deflections_yx_2d_from(grid=grid), 1e-4
+        spherical.deflections_yx_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -107,7 +107,7 @@ def test__potential_2d_from():
     spherical = ag.mp.IsothermalSph(centre=(1.1, 1.1), einstein_radius=3.0)
 
     assert elliptical.potential_2d_from(grid=grid) == pytest.approx(
-        spherical.potential_2d_from(grid=grid), 1e-4
+        spherical.potential_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -119,19 +119,19 @@ def test__shear_yx_2d_from():
     shear = mp.shear_yx_2d_from(grid=ag.Grid2DIrregular([[0.0, 1.0]]))
 
     assert shear[0, 0] == pytest.approx(0.0, 1e-4)
-    assert shear[0, 1] == pytest.approx(-convergence, 1e-4)
+    assert shear[0, 1] == pytest.approx(-convergence[0], 1e-4)
 
     convergence = mp.convergence_2d_from(grid=ag.Grid2DIrregular([[2.0, 1.0]]))
     shear = mp.shear_yx_2d_from(grid=ag.Grid2DIrregular([[2.0, 1.0]]))
 
-    assert shear[0, 0] == pytest.approx(-(4.0 / 5.0) * convergence, 1e-4)
-    assert shear[0, 1] == pytest.approx((3.0 / 5.0) * convergence, 1e-4)
+    assert shear[0, 0] == pytest.approx(-(4.0 / 5.0) * convergence[0], 1e-4)
+    assert shear[0, 1] == pytest.approx((3.0 / 5.0) * convergence[0], 1e-4)
 
     convergence = mp.convergence_2d_from(grid=ag.Grid2DIrregular([[3.0, 5.0]]))
     shear = mp.shear_yx_2d_from(grid=ag.Grid2DIrregular([[3.0, 5.0]]))
 
-    assert shear[0, 0] == pytest.approx(-(30.0 / 34.0) * convergence, 1e-4)
-    assert shear[0, 1] == pytest.approx(-(16.0 / 34.0) * convergence, 1e-4)
+    assert shear[0, 0] == pytest.approx(-(30.0 / 34.0) * convergence[0], 1e-4)
+    assert shear[0, 1] == pytest.approx(-(16.0 / 34.0) * convergence[0], 1e-4)
 
     mp = ag.mp.Isothermal(centre=(0.0, 0.0), ell_comps=(0.0, 0.0), einstein_radius=2.0)
 
@@ -140,13 +140,13 @@ def test__shear_yx_2d_from():
     shear = mp.shear_yx_2d_from(grid=ag.Grid2DIrregular([[0.0, 1.0]]))
 
     assert shear[0, 0] == pytest.approx(0.0, 1e-4)
-    assert shear[0, 1] == pytest.approx(-convergence, 1e-4)
+    assert shear[0, 1] == pytest.approx(-convergence[0], 1e-4)
 
     mp = ag.mp.Isothermal(centre=(0.0, 0.0), ell_comps=(0.3, 0.4), einstein_radius=2.0)
 
     shear = mp.shear_yx_2d_from(grid=ag.Grid2DIrregular([[0.0, 1.0]]))
 
-    assert shear[0, 0] == pytest.approx(0.0, 1e-4)
[…]
