### 2025-04-05 9959e74
test_isothermal_cored
