+++ test_autogalaxy/profiles/mass/total/test_isothermal_cored.py
@@ -60,7 +60,7 @@ def test__deflections_yx_2d_from():
     )
 
     assert elliptical.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        spherical.deflections_yx_2d_from(grid=grid), 1e-4
+        spherical.deflections_yx_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -205,7 +205,7 @@ def test__potential_2d_from():
     )
 
     assert elliptical.potential_2d_from(grid=grid) == pytest.approx(
-        spherical.potential_2d_from(grid=grid), 1e-4
+        spherical.potential_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -226,14 +226,14 @@ def test__compare_to_cored_power_law():
     )
 
     assert power_law.potential_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.potential_2d_from(grid=grid), 1e-3
+        cored_power_law.potential_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.potential_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.potential_2d_from(grid=grid), 1e-3
+        cored_power_law.potential_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.deflections_yx_2d_from(grid=grid), 1e-3
+        cored_power_law.deflections_yx_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.deflections_yx_2d_from(grid=grid), 1e-3
+        cored_power_law.deflections_yx_2d_from(grid=grid).array, 1e-3
     )
