### 2025-04-05 db07da1
test_power_law
