+++ autogalaxy/profiles/mass/total/power_law.py
@@ -1,12 +1,6 @@
-import os
-
-if os.environ.get("USE_JAX", "0") == "1":
-    USING_JAX = True
-    import jax.numpy as np
-    from .jax_utils import omega
-else:
-    USING_JAX = False
-    import numpy as np
+import jax.numpy as jnp
+import numpy as np
+from .jax_utils import omega
 
 from scipy import special
 from typing import Tuple
@@ -54,8 +48,8 @@ class PowerLaw(PowerLawCore):
         alpha_x = alpha[:, 1]
         alpha_y = alpha[:, 0]
 
-        x = grid[:, 1] - self.centre[1]
-        y = grid[:, 0] - self.centre[0]
+        x = grid.array[:, 1] - self.centre[1]
+        y = grid.array[:, 0] - self.centre[0]
 
         return (x * alpha_x + y * alpha_y) / (3 - self.slope)
 
@@ -82,26 +76,17 @@ class PowerLaw(PowerLawCore):
             2.0 / (self.axis_ratio**-0.5 + self.axis_ratio**0.5)
         ) * self.einstein_radius
 
-        factor = np.divide(1.0 - self.axis_ratio, 1.0 + self.axis_ratio)
-        b = np.multiply(einstein_radius, np.sqrt(self.axis_ratio))
-        angle = np.arctan2(
-            grid[:, 0], np.multiply(self.axis_ratio, grid[:, 1])
+        factor = jnp.divide(1.0 - self.axis_ratio, 1.0 + self.axis_ratio)
+        b = jnp.multiply(einstein_radius, jnp.sqrt(self.axis_ratio))
+        angle = jnp.arctan2(
+            grid.array[:, 0], jnp.multiply(self.axis_ratio, grid.array[:, 1])
         )  # Note, this angle is not the position angle
-        z = np.add(
-            np.multiply(np.cos(angle), 1 + 0j), np.multiply(np.sin(angle), 0 + 1j)
+        z = jnp.add(
+            jnp.multiply(jnp.cos(angle), 1 + 0j), jnp.multiply(jnp.sin(angle), 0 + 1j)
         )
 
-        if USING_JAX:
-            # offset radius so calculation is finite at (0, 0)
-            R = np.sqrt((self.axis_ratio * grid[:, 1]) ** 2 + grid[:, 0] ** 2 + 1e-16)
-            zh = omega(z, slope, factor, n_terms=20)
-        else:
-            R = np.sqrt(
-                np.add(
-                    np.multiply(self.axis_ratio**2, grid[:, 1] ** 2), grid[:, 0] ** 2
-                )
-            )
-            zh = z * special.hyp2f1(1.0, 0.5 * slope, 2.0 - 0.5 * slope, -factor * z**2)
+        R = jnp.sqrt((self.axis_ratio * grid.array[:, 1]) ** 2 + grid.array[:, 0] ** 2 + 1e-16)
+        zh = omega(z, slope, factor, n_terms=20)
 
         complex_angle = (
             2.0 * b / (1.0 + self.axis_ratio) * (b / R) ** (slope - 1.0) * zh
@@ -116,17 +101,17 @@ class PowerLaw(PowerLawCore):
         deflection_x *= rescale_factor
 
         return self.rotated_grid_from_reference_frame_from(
-            grid=np.vstack((deflection_y, deflection_x)).T
+            grid=jnp.vstack((deflection_y, deflection_x)).T
         )
 
     def convergence_func(self, grid_radius: float) -> float:
         if grid_radius > 0.0:
             return self.einstein_radius_rescaled * grid_radius ** (-(self.slope - 1))
-        return np.inf
+        return jnp.inf
 
     @staticmethod
[…]
