+++ test_autogalaxy/profiles/mass/total/test_power_law.py
@@ -86,7 +86,7 @@ def test__deflections_yx_2d_from():
     spherical = ag.mp.PowerLawSph(centre=(1.1, 1.1), einstein_radius=3.0, slope=2.4)
 
     assert elliptical.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        spherical.deflections_yx_2d_from(grid=grid), 1e-4
+        spherical.deflections_yx_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -190,7 +190,7 @@ def test__potential_2d_from():
     spherical = ag.mp.PowerLawSph(centre=(1.1, 1.1), einstein_radius=3.0, slope=2.4)
 
     assert elliptical.potential_2d_from(grid=grid) == pytest.approx(
-        spherical.potential_2d_from(grid=grid), 1e-4
+        spherical.potential_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -211,14 +211,14 @@ def test__compare_to_cored_power_law():
     )
 
     assert power_law.potential_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.potential_2d_from(grid=grid), 1e-3
+        cored_power_law.potential_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.potential_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.potential_2d_from(grid=grid), 1e-3
+        cored_power_law.potential_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.deflections_yx_2d_from(grid=grid), 1e-3
+        cored_power_law.deflections_yx_2d_from(grid=grid).array, 1e-3
     )
     assert power_law.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        cored_power_law.deflections_yx_2d_from(grid=grid), 1e-3
+        cored_power_law.deflections_yx_2d_from(grid=grid).array, 1e-3
     )
