### 2025-04-05 016647f
test_power_law_broken
