+++ autogalaxy/profiles/mass/total/power_law_broken.py
@@ -1,4 +1,4 @@
-import copy
+import jax.numpy as jnp
 import numpy as np
 from typing import Tuple
 
@@ -31,7 +31,7 @@ class PowerLawBroken(MassProfile):
         super().__init__(centre=centre, ell_comps=ell_comps)
 
         self.einstein_radius = einstein_radius
-        self.einstein_radius_elliptical = np.sqrt(self.axis_ratio) * einstein_radius
+        self.einstein_radius_elliptical = jnp.sqrt(self.axis_ratio) * einstein_radius
         self.break_radius = break_radius
         self.inner_slope = inner_slope
         self.outer_slope = outer_slope
@@ -58,7 +58,7 @@ class PowerLawBroken(MassProfile):
         """
 
         # Ell radius
-        radius = np.hypot(grid[:, 1] * self.axis_ratio, grid[:, 0])
+        radius = jnp.hypot(grid.array[:, 1] * self.axis_ratio, grid.array[:, 0])
 
         # Inside break radius
         kappa_inner = self.kB * (self.break_radius / radius) ** self.inner_slope
@@ -72,7 +72,7 @@ class PowerLawBroken(MassProfile):
 
     @aa.grid_dec.to_array
     def potential_2d_from(self, grid: aa.type.Grid2DLike, **kwargs):
-        return np.zeros(shape=grid.shape[0])
+        return jnp.zeros(shape=grid.shape[0])
 
     @aa.grid_dec.to_vector_yx
     @aa.grid_dec.transform
@@ -81,10 +81,10 @@ class PowerLawBroken(MassProfile):
         Returns the complex deflection angle from eq. 18 and 19
         """
         # Rotate coordinates
-        z = grid[:, 1] + 1j * grid[:, 0]
+        z = grid.array[:, 1] + 1j * grid.array[:, 0]
 
         # Ell radius
-        R = np.hypot(z.real * self.axis_ratio, z.imag)
+        R = jnp.hypot(z.real * self.axis_ratio, z.imag)
 
         # Factors common to eq. 18 and 19
         factors = (
@@ -126,8 +126,8 @@ class PowerLawBroken(MassProfile):
         ).conjugate()
 
         return self.rotated_grid_from_reference_frame_from(
-            grid=np.multiply(
-                1.0, np.vstack((np.imag(deflections), np.real(deflections))).T
+            grid=jnp.multiply(
+                1.0, jnp.vstack((jnp.imag(deflections), jnp.real(deflections))).T
             )
         )
 
@@ -140,13 +140,13 @@ class PowerLawBroken(MassProfile):
 
         # u from eq. 25
         q_ = (1 - q**2) / (q**2)
-        u = 0.5 * (1 - np.sqrt(1 - q_ * (r / z) ** 2))
+        u = 0.5 * (1 - jnp.sqrt(1 - q_ * (r / z) ** 2))
 
         # First coefficient
         a_n = 1.0
 
         # Storage for sum
-        F = np.zeros_like(z, dtype="complex64")
+        F = jnp.zeros_like(z, dtype="complex64")
 
         for n in range(max_terms):
             F += a_n * (u**n)
