### 2025-04-05 c786653
power_law_multopole
