+++ test_autogalaxy/profiles/mass/total/test_power_law_cored.py
@@ -66,7 +66,7 @@ def test__deflections_yx_2d_from():
     )
 
     assert elliptical.deflections_yx_2d_from(grid=grid) == pytest.approx(
-        spherical.deflections_yx_2d_from(grid=grid), 1e-4
+        spherical.deflections_yx_2d_from(grid=grid).array, 1e-4
     )
 
 
@@ -176,5 +176,5 @@ def test__potential_2d_from():
     )
 
     assert elliptical.potential_2d_from(grid=grid) == pytest.approx(
-        spherical.potential_2d_from(grid=grid), 1e-4
+        spherical.potential_2d_from(grid=grid).array, 1e-4
     )
