+++ autogalaxy/profiles/mass/total/power_law_multipole.py
@@ -1,4 +1,5 @@
 from astropy import units
+import jax.numpy as jnp
 import numpy as np
 from typing import Tuple
 
@@ -26,14 +27,14 @@ def radial_and_angle_grid_from(
     -------
     The radial and polar coordinate grids of the input (y,x) Cartesian grid.
     """
-    y, x = grid.T
+    y, x = grid.array.T
 
-    x_shifted = np.subtract(x, centre[1])
-    y_shifted = np.subtract(y, centre[0])
+    x_shifted = jnp.subtract(x, centre[1])
+    y_shifted = jnp.subtract(y, centre[0])
 
-    radial_grid = np.sqrt(x_shifted**2 + y_shifted**2)
+    radial_grid = jnp.sqrt(x_shifted**2 + y_shifted**2)
 
-    angle_grid = np.arctan2(y_shifted, x_shifted)
+    angle_grid = jnp.arctan2(y_shifted, x_shifted)
 
     return radial_grid, angle_grid
 
@@ -142,8 +143,8 @@ class PowerLawMultipole(MassProfile):
             The polar angle coordinates of the input (y,x) Cartesian grid of coordinates.
         """
         return (
-            a_r * np.sin(polar_angle_grid) + a_angle * np.cos(polar_angle_grid),
-            a_r * np.cos(polar_angle_grid) - a_angle * np.sin(polar_angle_grid),
+            a_r * jnp.sin(polar_angle_grid) + a_angle * jnp.cos(polar_angle_grid),
+            a_r * jnp.cos(polar_angle_grid) - a_angle * jnp.sin(polar_angle_grid),
         )
 
     @aa.grid_dec.to_vector_yx
@@ -172,7 +173,7 @@ class PowerLawMultipole(MassProfile):
             )
             / (self.m**2.0 - (3.0 - self.slope) ** 2.0)
             * self.k_m
-            * np.cos(self.m * (polar_angle_grid - self.angle_m))
+            * jnp.cos(self.m * (polar_angle_grid - self.angle_m))
         )
 
         a_angle = (
@@ -183,10 +184,10 @@ class PowerLawMultipole(MassProfile):
             )
             / (self.m**2.0 - (3.0 - self.slope) ** 2.0)
             * self.k_m
-            * np.sin(self.m * (polar_angle_grid - self.angle_m))
+            * jnp.sin(self.m * (polar_angle_grid - self.angle_m))
         )
 
-        return np.stack(
+        return jnp.stack(
             self.jacobian(a_r=a_r, a_angle=a_angle, polar_angle_grid=polar_angle_grid),
             axis=-1,
         )
@@ -210,7 +211,7 @@ class PowerLawMultipole(MassProfile):
             / 2.0
             * (self.einstein_radius / r) ** (self.slope - 1)
             * self.k_m
-            * np.cos(self.m * (angle - self.angle_m))
+            * jnp.cos(self.m * (angle - self.angle_m))
         )
 
     @aa.grid_dec.to_array
@@ -223,4 +224,4 @@ class PowerLawMultipole(MassProfile):
         grid
             The grid of (y,x) arc-second coordinates the deflection angles are computed on.
         """
-        return np.zeros(shape=grid.shape[0])
+        return jnp.zeros(shape=grid.shape[0])
