### 2025-04-05 b9afe00
test abstract mass
