+++ autogalaxy/profiles/mass/sheets/input_deflections.py
@@ -1,3 +1,4 @@
+import jax.numpy as jnp
 import numpy as np
 from scipy.interpolate import griddata
 
