+++ autogalaxy/operate/deflections.py
@@ -1,10 +1,10 @@
+import jax
+from jax import jit
+import jax.numpy as jnp
 from functools import wraps, partial
 import logging
-from autofit.jax_wrapper import numpy as np, use_jax, jit
+import numpy as np
 from typing import List, Tuple, Union
-
-if use_jax:
-    import jax
 
 from autoconf import conf
 
@@ -92,7 +92,7 @@ def evaluation_grid(func):
 
 
 def one_step(r, _, theta, fun, fun_dr):
-    r = np.abs(r - fun(r, theta) / fun_dr(r, theta))
+    r = jnp.abs(r - fun(r, theta) / fun_dr(r, theta))
     return r, None
 
 
@@ -101,8 +101,8 @@ def step_r(r, theta, fun, fun_dr, N=20):
     one_step_partial = jax.tree_util.Partial(
         one_step, theta=theta, fun=fun, fun_dr=fun_dr
     )
-    new_r = jax.lax.scan(one_step_partial, r, xs=np.arange(N))[0]
-    return np.stack([new_r * np.sin(theta), new_r * np.cos(theta)]).T
+    new_r = jax.lax.scan(one_step_partial, r, xs=jnp.arange(N))[0]
+    return jnp.stack([new_r * jnp.sin(theta), new_r * jnp.cos(theta)]).T
 
 
 class OperateDeflections:
@@ -125,18 +125,21 @@ class OperateDeflections:
         raise NotImplementedError
 
     def deflections_yx_scalar(self, y, x, pixel_scales):
-        if not use_jax:
-            return
-        else:
-            # A version of the deflection function that takes in two scalars
-            # and outputs a 2D vector.  Needed for JAX auto differentiation.
-            g = aa.Grid2D.from_yx_1d(
-                y=y.reshape(1),
-                x=x.reshape(1),
-                shape_native=(1, 1),
-                pixel_scales=pixel_scales,
-            )
-            return self.deflections_yx_2d_from(g).squeeze()
+
+        # A version of the deflection function that takes in two scalars
+        # and outputs a 2D vector.  Needed for JAX auto differentiation.
+
+        mask = aa.Mask2D.all_false(
+            shape_native=(1,1),
+            pixel_scales=pixel_scales,
+        )
+
+        g = aa.Grid2D(
+            values=jnp.stack((y.reshape(1), x.reshape(1)), axis=-1),
+            mask=mask
+        )
+
+        return self.deflections_yx_2d_from(g).squeeze()
 
     def __eq__(self, other):
         return self.__dict__ == other.__dict__ and self.__class__ is other.__class__
@@ -328,7 +331,7 @@ class OperateDeflections:
         gamma_1 = 0.5 * (hessian_xx - hessian_yy)
         gamma_2 = hessian_xy
 
-        shear_yx_2d = np.zeros(shape=(grid.shape_slim, 2))
+        shear_yx_2d = jnp.zeros(shape=(grid.shape_slim, 2))
 
         shear_yx_2d[:, 0] = gamma_2
         shear_yx_2d[:, 1] = gamma_1
@@ -577,7 +580,7 @@ class OperateDeflections:
 
         for curve in curve_list:
             x, y = curve[:, 0], curve[:, 1]
-            area = np.abs(0.5 * np.sum(y[:-1] * np.diff(x) - x[:-1] * np.diff(y)))
+            area = jnp.abs(0.5 * jnp.sum(y[:-1] * jnp.diff(x) - x[:-1] * jnp.diff(y)))
             area_within_each_curve_list.append(area)
 
         return area_within_each_curve_list
@@ -614,7 +617,7 @@ class OperateDeflections:
             area_list = self.tangential_critical_curve_area_list_from(
[…]
