### 2025-04-06 8468808
more operate tests pass
