+++ autogalaxy/operate/deflections.py
@@ -47,50 +47,6 @@ def precompute_jacobian(func):
 
     return wrapper
 
-
-def evaluation_grid(func):
-    @wraps(func)
-    def wrapper(
-        lensing_obj, grid, pixel_scale: Union[Tuple[float, float], float] = 0.05
-    ):
-        if hasattr(grid, "is_evaluation_grid"):
-            if grid.is_evaluation_grid:
-                return func(lensing_obj, grid, pixel_scale)
-
-        pixel_scale_ratio = grid.pixel_scale / pixel_scale
-
-        zoom_shape_native = grid.mask.zoom_shape_native
-        shape_native = (
-            int(pixel_scale_ratio * zoom_shape_native[0]),
-            int(pixel_scale_ratio * zoom_shape_native[1]),
-        )
-
-        max_evaluation_grid_size = conf.instance["general"]["grid"][
-            "max_evaluation_grid_size"
-        ]
-
-        # This is a hack to prevent the evaluation gird going beyond 1000 x 1000 pixels, which slows the code
-        # down a lot. Need a better moe robust way to set this up for any general lens.
-
-        if shape_native[0] > max_evaluation_grid_size:
-            pixel_scale = pixel_scale_ratio / (
-                shape_native[0] / float(max_evaluation_grid_size)
-            )
-            shape_native = (max_evaluation_grid_size, max_evaluation_grid_size)
-
-        grid = aa.Grid2D.uniform(
-            shape_native=shape_native,
-            pixel_scales=(pixel_scale, pixel_scale),
-            origin=grid.mask.zoom_offset_scaled,
-        )
-
-        grid.is_evaluation_grid = True
-
-        return func(lensing_obj, grid, pixel_scale)
-
-    return wrapper
-
-
 def one_step(r, _, theta, fun, fun_dr):
     r = jnp.abs(r - fun(r, theta) / fun_dr(r, theta))
     return r, None
@@ -378,9 +334,8 @@ class OperateDeflections:
 
         return grid_contour.contour_list
 
-    @evaluation_grid
     def tangential_critical_curve_list_from(
-        self, grid, pixel_scale: Union[Tuple[float, float], float] = 0.05
+        self, grid, 
     ) -> List[aa.Grid2DIrregular]:
         """
         Returns all tangential critical curves of the lensing system, which are computed as follows:
@@ -404,9 +359,8 @@ class OperateDeflections:
 
         return self.contour_list_from(grid=grid, contour_array=tangential_eigen_values)
 
-    @evaluation_grid
     def radial_critical_curve_list_from(
-        self, grid, pixel_scale: Union[Tuple[float, float], float] = 0.05
+        self, grid, 
     ) -> List[aa.Grid2DIrregular]:
         """
         Returns all radial critical curves of the lensing system, which are computed as follows:
@@ -430,9 +384,8 @@ class OperateDeflections:
 
         return self.contour_list_from(grid=grid, contour_array=radial_eigen_values)
 
-    @evaluation_grid
     def tangential_caustic_list_from(
-        self, grid, pixel_scale: Union[Tuple[float, float], float] = 0.05
+        self, grid, 
     ) -> List[aa.Grid2DIrregular]:
         """
         Returns all tangential caustics of the lensing system, which are computed as follows:
@@ -456,7 +409,7 @@ class OperateDeflections:
         """
 
         tangential_critical_curve_list = self.tangential_critical_curve_list_from(
-            grid=grid, pixel_scale=pixel_scale
+            grid=grid
         )
 
         tangential_caustic_list = []
@@ -472,9 +425,8 @@ class OperateDeflections:
 
         return tangential_caustic_list
 
-    @evaluation_grid
[…]
