### 2025-04-06 a9f1a26
remove input delections
