+++ autogalaxy/profiles/mass/__init__.py
@@ -48,4 +48,4 @@ from .stellar import (
     Chameleon,
     ChameleonSph,
 )
-from .sheets import ExternalShear, MassSheet, InputDeflections
+from .sheets import ExternalShear, MassSheet
