+++ autogalaxy/profiles/mass/sheets/__init__.py
@@ -1,3 +1,2 @@
 from .mass_sheet import MassSheet
 from .external_shear import ExternalShear
-from .input_deflections import InputDeflections
