+++ test_autogalaxy/profiles/mass/stellar/test_sersic_gradient.py
@@ -59,7 +59,7 @@ def test__deflections_2d_via_mge_from():
         grid=ag.Grid2DIrregular([[0.1625, 0.1625]])
     )
 
-    assert deflections_via_integral == pytest.approx(deflections_via_mge, 1.0e-3)
+    assert deflections_via_integral == pytest.approx(deflections_via_mge.array, 1.0e-3)
 
 
 def test__deflections_2d_via_cse_from():
@@ -80,7 +80,7 @@ def test__deflections_2d_via_cse_from():
         grid=ag.Grid2DIrregular([[0.1625, 0.1625]])
     )
 
-    assert deflections_via_integral == pytest.approx(deflections_via_cse, 1.0e-4)
+    assert deflections_via_integral == pytest.approx(deflections_via_cse.array, 1.0e-4)
 
     mp = ag.mp.SersicGradient(
         centre=(-0.4, -0.2),
@@ -99,7 +99,7 @@ def test__deflections_2d_via_cse_from():
         grid=ag.Grid2DIrregular([[0.1625, 0.1625]])
     )
 
-    assert deflections_via_integral == pytest.approx(deflections_via_cse, 1.0e-4)
+    assert deflections_via_integral == pytest.approx(deflections_via_cse.array, 1.0e-4)
 
 
 def test__deflections_yx_2d_from():
@@ -110,7 +110,7 @@ def test__deflections_yx_2d_from():
         grid=ag.Grid2DIrregular([[1.0, 0.0]])
     )
 
-    assert deflections == pytest.approx(deflections_via_integral, 1.0e-4)
+    assert deflections == pytest.approx(deflections_via_integral.array, 1.0e-4)
 
     mp = ag.mp.SersicGradientSph()
 
@@ -119,7 +119,7 @@ def test__deflections_yx_2d_from():
         grid=ag.Grid2DIrregular([[1.0, 0.0]])
     )
 
-    assert deflections == pytest.approx(deflections_via_integral, 1.0e-4)
+    assert deflections == pytest.approx(deflections_via_integral.array, 1.0e-4)
 
     elliptical = ag.mp.SersicGradient(
         centre=(0.0, 0.0),
@@ -143,7 +143,7 @@ def test__deflections_yx_2d_from():
     ell_deflections_yx_2d = elliptical.deflections_yx_2d_from(grid=grid)
     sph_deflections_yx_2d = spherical.deflections_yx_2d_from(grid=grid)
 
-    ell_deflections_yx_2d == pytest.approx(sph_deflections_yx_2d, 1.0e-4)
+    ell_deflections_yx_2d == pytest.approx(sph_deflections_yx_2d.array, 1.0e-4)
 
 
 def test__convergence_2d_from():
@@ -239,7 +239,7 @@ def test__convergence_2d_from():
     ell_convergence_2d = elliptical.convergence_2d_from(grid=grid)
     sph_convergence_2d = spherical.convergence_2d_from(grid=grid)
 
-    assert ell_convergence_2d == pytest.approx(sph_convergence_2d, 1.0e-4)
+    assert ell_convergence_2d == pytest.approx(sph_convergence_2d.array, 1.0e-4)
 
 
 def test__compare_to_sersic():
