+++ autogalaxy/profiles/mass/stellar/gaussian.py
@@ -104,7 +104,7 @@ class Gaussian(MassProfile, StellarProfile):
         """
 
         def calculate_deflection_component(npow, index):
-            deflection_grid = self.axis_ratio * grid[:, index]
+            deflection_grid = np.array(self.axis_ratio * grid.array[:, index])
 
             for i in range(grid.shape[0]):
                 deflection_grid[i] *= (
@@ -115,8 +115,8 @@ class Gaussian(MassProfile, StellarProfile):
                         a=0.0,
                         b=1.0,
                         args=(
-                            grid[i, 0],
-                            grid[i, 1],
+                            grid.array[i, 0],
+                            grid.array[i, 1],
                             npow,
                             self.axis_ratio,
                             self.sigma / np.sqrt(self.axis_ratio),
@@ -181,7 +181,7 @@ class Gaussian(MassProfile, StellarProfile):
             np.exp(
                 -0.5
                 * np.square(
-                    np.divide(grid_radii, self.sigma / np.sqrt(self.axis_ratio))
+                    np.divide(grid_radii.array, self.sigma / np.sqrt(self.axis_ratio))
                 )
             ),
         )
@@ -196,15 +196,15 @@ class Gaussian(MassProfile, StellarProfile):
 
     def zeta_from(self, grid: aa.type.Grid2DLike):
         q2 = self.axis_ratio**2.0
-        ind_pos_y = grid[:, 0] >= 0
+        ind_pos_y = grid.array[:, 0] >= 0
         shape_grid = np.shape(grid)
         output_grid = np.zeros((shape_grid[0]), dtype=np.complex128)
         scale_factor = self.axis_ratio / (self.sigma * np.sqrt(2.0 * (1.0 - q2)))
 
-        xs_0 = grid[:, 1][ind_pos_y] * scale_factor
-        ys_0 = grid[:, 0][ind_pos_y] * scale_factor
-        xs_1 = grid[:, 1][~ind_pos_y] * scale_factor
-        ys_1 = -grid[:, 0][~ind_pos_y] * scale_factor
+        xs_0 = grid.array[:, 1][ind_pos_y] * scale_factor
+        ys_0 = grid.array[:, 0][ind_pos_y] * scale_factor
+        xs_1 = grid.array[:, 1][~ind_pos_y] * scale_factor
+        ys_1 = -grid.array[:, 0][~ind_pos_y] * scale_factor
 
         output_grid[ind_pos_y] = -1j * (
             wofz(xs_0 + 1j * ys_0)
