+++ autogalaxy/profiles/mass/stellar/chameleon.py
@@ -1,4 +1,4 @@
-import copy
+import jax.numpy as jnp
 import numpy as np
 from typing import Tuple
 
@@ -71,13 +71,13 @@ class Chameleon(MassProfile, StellarProfile):
             * self.intensity
             / (1 + self.axis_ratio)
             * self.axis_ratio
-            / np.sqrt(1.0 - self.axis_ratio**2.0)
+            / jnp.sqrt(1.0 - self.axis_ratio**2.0)
         )
 
-        core_radius_0 = np.sqrt(
+        core_radius_0 = jnp.sqrt(
             (4.0 * self.core_radius_0**2.0) / (1.0 + self.axis_ratio) ** 2
         )
-        core_radius_1 = np.sqrt(
+        core_radius_1 = jnp.sqrt(
             (4.0 * self.core_radius_1**2.0) / (1.0 + self.axis_ratio) ** 2
         )
 
@@ -88,39 +88,39 @@ class Chameleon(MassProfile, StellarProfile):
             grid=grid, axis_ratio=self.axis_ratio, core_radius=core_radius_1
         )
 
-        deflection_y0 = np.arctanh(
-            np.divide(
-                np.multiply(np.sqrt(1.0 - self.axis_ratio**2.0), grid[:, 0]),
-                np.add(psi0, self.axis_ratio**2.0 * core_radius_0),
+        deflection_y0 = jnp.arctanh(
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1.0 - self.axis_ratio**2.0), grid.array[:, 0]),
+                jnp.add(psi0, self.axis_ratio**2.0 * core_radius_0),
             )
         )
 
-        deflection_x0 = np.arctan(
-            np.divide(
-                np.multiply(np.sqrt(1.0 - self.axis_ratio**2.0), grid[:, 1]),
-                np.add(psi0, core_radius_0),
+        deflection_x0 = jnp.arctan(
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1.0 - self.axis_ratio**2.0), grid.array[:, 1]),
+                jnp.add(psi0, core_radius_0),
             )
         )
 
-        deflection_y1 = np.arctanh(
-            np.divide(
-                np.multiply(np.sqrt(1.0 - self.axis_ratio**2.0), grid[:, 0]),
-                np.add(psi1, self.axis_ratio**2.0 * core_radius_1),
+        deflection_y1 = jnp.arctanh(
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1.0 - self.axis_ratio**2.0), grid.array[:, 0]),
+                jnp.add(psi1, self.axis_ratio**2.0 * core_radius_1),
             )
         )
 
-        deflection_x1 = np.arctan(
-            np.divide(
-                np.multiply(np.sqrt(1.0 - self.axis_ratio**2.0), grid[:, 1]),
-                np.add(psi1, core_radius_1),
+        deflection_x1 = jnp.arctan(
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1.0 - self.axis_ratio**2.0), grid.array[:, 1]),
+                jnp.add(psi1, core_radius_1),
             )
         )
 
-        deflection_y = np.subtract(deflection_y0, deflection_y1)
-        deflection_x = np.subtract(deflection_x0, deflection_x1)
+        deflection_y = jnp.subtract(deflection_y0, deflection_y1)
+        deflection_x = jnp.subtract(deflection_x0, deflection_x1)
 
         return self.rotated_grid_from_reference_frame_from(
-            np.multiply(factor, np.vstack((deflection_y, deflection_x)).T)
[…]
