### 2025-04-07 c8fca55
added fits_dataset and fits_adapt_images to plots.yaml
