+++ autogalaxy/config/visualize/plots.yaml
@@ -1,12 +1,20 @@
 # The `plots` section customizes every image that is output to hard-disk during a model-fit.
 
-# For example, if `plots: fit: subplot_fit=True``, the ``fit_dataset.png`` subplot file will
+# For example, if `plots: fit: subplot_fit=True``, the ``subplot_fit.png`` subplot file will
 # be plotted every time visualization is performed.
+
+# There are two settings which are important for inspecting results via the dataset after a fit is complete which are:
+
+# - `fits_dataset`: This outputs `dataset.fits` which the database functionality may use to reperform fits.
+# -`fits_adapt_images`, This outputs `adapt_images.fits` which the database functionality may use to reperform fits.
+
+# These can be disabled to save on hard-disk space but will lead to certain database functionality being disabled.
 
 subplot_format: [png]                      # Output format of all subplots, can be png, pdf or both (e.g. [png, pdf])
 
 dataset:                                   # Settings for plots of all datasets (e.g. ImagingPlotter, InterferometerPlotter).
   subplot_dataset: true                    # Plot subplot containing all dataset quantities (e.g. the data, noise-map, etc.)?
+  fits_dataset : true                      # Output a .fits file containing the dataset data, noise-map and other quantities?
 
 fit:                                       # Settings for plots of all fits (e.g. FitImagingPlotter, FitInterferometerPlotter).
   subplot_fit: true                        # Plot subplot of all fit quantities for any dataset (e.g. the model data, residual-map, etc.)?
@@ -32,6 +40,7 @@ inversion:                                 # Settings for plots of inversions (e
 
 adapt:                                     # Settings for plots of adapt images used by adaptive pixelizations.
   subplot_adapt_images: true               # Plot subplot showing each adapt image used for adaptive pixelization?
+  fits_adapt_images: true                  # Output a .fits file containing the adapt images used for adaptive pixelization?
 
 fit_interferometer:                        # Settings for plots of fits to interferometer datasets (e.g. FitInterferometerPlotter).
   subplot_fit_dirty_images: false          # Plot subplot of the dirty-images of all interferometer datasets?
