+++ autogalaxy/imaging/model/plotter_interface.py
@@ -103,27 +103,29 @@ class PlotterInterfaceImaging(PlotterInterface):
         if should_plot("subplot_dataset"):
             dataset_plotter.subplot_dataset()
 
-        hdu_list = hdu_list_for_output_from(
-            values_list=[
-                dataset.mask.astype("float"),
-                dataset.data.native,
-                dataset.noise_map.native,
-                dataset.psf.native,
-                dataset.grids.lp.over_sample_size.native,
-                dataset.grids.pixelization.over_sample_size.native,
-            ],
-            ext_name_list=[
-                "mask",
-                "data",
-                "noise_map",
-                "psf",
-                "over_sample_size_lp",
-                "over_sample_size_pixelization",
-            ],
-            header_dict=dataset.mask.header_dict,
-        )
+        if should_plot("fits_dataset"):
 
-        hdu_list.writeto(self.image_path / "dataset.fits", overwrite=True)
+            hdu_list = hdu_list_for_output_from(
+                values_list=[
+                    dataset.mask.astype("float"),
+                    dataset.data.native,
+                    dataset.noise_map.native,
+                    dataset.psf.native,
+                    dataset.grids.lp.over_sample_size.native,
+                    dataset.grids.pixelization.over_sample_size.native,
+                ],
+                ext_name_list=[
+                    "mask",
+                    "data",
+                    "noise_map",
+                    "psf",
+                    "over_sample_size_lp",
+                    "over_sample_size_pixelization",
+                ],
+                header_dict=dataset.mask.header_dict,
+            )
+
+            hdu_list.writeto(self.image_path / "dataset.fits", overwrite=True)
 
     def fit_imaging(
         self,
