### 2025-04-07 3e2ce85
also use on fits_adapt_dataset
