+++ autogalaxy/interferometer/model/plotter_interface.py
@@ -109,7 +109,6 @@ class PlotterInterfaceInterferometer(PlotterInterface):
             dataset_plotter.subplot_dataset()
 
         if should_plot("fits_dataset"):
-
             hdu_list = hdu_list_for_output_from(
                 values_list=[
                     dataset.real_space_mask.astype("float"),
