+++ test_autogalaxy/profiles/mass/stellar/test_sersic_core.py
@@ -6,96 +6,96 @@ import autogalaxy as ag
 grid = ag.Grid2DIrregular([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
-def test__deflections_2d_via_mge_from():
-    mp = ag.mp.SersicCore(
-        centre=(1.0, 2.0),
-        ell_comps=ag.convert.ell_comps_from(axis_ratio=0.5, angle=70.0),
-        intensity=0.45,
-        effective_radius=0.5,
-        radius_break=0.01,
-        gamma=0.0,
-        alpha=2.0,
-        sersic_index=2.2,
-    )
+# def test__deflections_2d_via_mge_from():
+#     mp = ag.mp.SersicCore(
+#         centre=(1.0, 2.0),
+#         ell_comps=ag.convert.ell_comps_from(axis_ratio=0.5, angle=70.0),
+#         intensity=0.45,
+#         effective_radius=0.5,
+#         radius_break=0.01,
+#         gamma=0.0,
+#         alpha=2.0,
+#         sersic_index=2.2,
+#     )
+#
+#     deflections = mp.deflections_2d_via_mge_from(grid=ag.Grid2DIrregular([[2.5, -2.5]]))
+#
+#     assert deflections[0, 0] == pytest.approx(0.0015047, 1e-4)
+#     assert deflections[0, 1] == pytest.approx(-0.004493, 1e-4)
+#
+#     mp = ag.mp.SersicCore(
+#         centre=(1.0, 2.0),
+#         ell_comps=ag.convert.ell_comps_from(axis_ratio=0.5, angle=70.0),
+#         intensity=2.0 * 0.45,
+#         effective_radius=0.5,
+#         radius_break=0.01,
+#         gamma=0.0,
+#         alpha=2.0,
+#         sersic_index=2.2,
+#     )
+#
+#     deflections = mp.deflections_yx_2d_from(grid=ag.Grid2DIrregular([[2.5, -2.5]]))
+#
+#     assert deflections[0, 0] == pytest.approx(2.0 * 0.0015047, 1e-4)
+#     assert deflections[0, 1] == pytest.approx(2.0 * -0.004493, 1e-4)
+#
+#     mp = ag.mp.SersicCore(
+#         centre=(1.0, 2.0),
+#         ell_comps=ag.convert.ell_comps_from(axis_ratio=0.5, angle=70.0),
+#         intensity=0.45,
+#         effective_radius=0.5,
+#         radius_break=0.01,
+#         gamma=0.0,
+#         alpha=2.0,
+#         sersic_index=2.2,
+#         mass_to_light_ratio=2.0,
+#     )
+#
+#     deflections = mp.deflections_2d_via_mge_from(grid=ag.Grid2DIrregular([[2.5, -2.5]]))
+#
+#     assert deflections[0, 0] == pytest.approx(2.0 * 0.0015047, 1e-4)
+#     assert deflections[0, 1] == pytest.approx(2.0 * -0.004493, 1e-4)
 
-    deflections = mp.deflections_2d_via_mge_from(grid=ag.Grid2DIrregular([[2.5, -2.5]]))
 
-    assert deflections[0, 0] == pytest.approx(0.0015047, 1e-4)
-    assert deflections[0, 1] == pytest.approx(-0.004493, 1e-4)
-
-    mp = ag.mp.SersicCore(
-        centre=(1.0, 2.0),
-        ell_comps=ag.convert.ell_comps_from(axis_ratio=0.5, angle=70.0),
-        intensity=2.0 * 0.45,
-        effective_radius=0.5,
-        radius_break=0.01,
-        gamma=0.0,
-        alpha=2.0,
-        sersic_index=2.2,
-    )
-
[…]
