### 2025-04-07 a967fbb
NFW tests pass
