### 2025-04-07 fd8b550
test_light_and_mass_profiles
