+++ test_autogalaxy/profiles/test_light_and_mass_profiles.py
@@ -20,14 +20,11 @@ def test__gaussian():
         mass_to_light_ratio=2.0,
     )
 
-    assert (lp.image_2d_from(grid=grid) == lmp.image_2d_from(grid=grid)).all()
-    assert (
-        mp.convergence_2d_from(grid=grid) == lmp.convergence_2d_from(grid=grid)
-    ).all()
+    assert lp.image_2d_from(grid=grid) == pytest.approx(lmp.image_2d_from(grid=grid).array, 1.0e-4)
+    assert mp.convergence_2d_from(grid=grid) == pytest.approx(lmp.convergence_2d_from(grid=grid).array, 1.0e-4)
     #    assert (mp.potential_2d_from(grid=grid) == lmp.potential_2d_from(grid=grid)).all()
-    assert (
-        mp.deflections_yx_2d_from(grid=grid) == lmp.deflections_yx_2d_from(grid=grid)
-    ).all()
+    assert mp.deflections_yx_2d_from(grid=grid) == pytest.approx(lmp.deflections_yx_2d_from(grid=grid).array, 1.0e-4)
+
 
 
 def test__gaussian_gradient():
@@ -49,14 +46,11 @@ def test__gaussian_gradient():
         mass_to_light_radius=1.0,
     )
 
-    assert (lp.image_2d_from(grid=grid) == lmp.image_2d_from(grid=grid)).all()
-    assert (
-        mp.convergence_2d_from(grid=grid) == lmp.convergence_2d_from(grid=grid)
-    ).all()
+    assert lp.image_2d_from(grid=grid) == pytest.approx(lmp.image_2d_from(grid=grid).array, 1.0e-4)
+    assert mp.convergence_2d_from(grid=grid) == pytest.approx(lmp.convergence_2d_from(grid=grid).array, 1.0e-4)
     #    assert (mp.potential_2d_from(grid=grid) == lmp.potential_2d_from(grid=grid)).all()
-    assert (
-        mp.deflections_yx_2d_from(grid=grid) == lmp.deflections_yx_2d_from(grid=grid)
-    ).all()
+    assert mp.deflections_yx_2d_from(grid=grid) == pytest.approx(lmp.deflections_yx_2d_from(grid=grid).array, 1.0e-4)
+
 
 
 def test__sersic():
@@ -81,14 +75,11 @@ def test__sersic():
         mass_to_light_ratio=2.0,
     )
 
-    assert (lp.image_2d_from(grid=grid) == lmp.image_2d_from(grid=grid)).all()
-    assert (
-        mp.convergence_2d_from(grid=grid) == lmp.convergence_2d_from(grid=grid)
-    ).all()
+    assert lp.image_2d_from(grid=grid) == pytest.approx(lmp.image_2d_from(grid=grid).array, 1.0e-4)
+    assert mp.convergence_2d_from(grid=grid) == pytest.approx(lmp.convergence_2d_from(grid=grid).array, 1.0e-4)
     #    assert (mp.potential_2d_from(grid=grid) == lmp.potential_2d_from(grid=grid)).all()
-    assert (
-        mp.deflections_yx_2d_from(grid=grid) == lmp.deflections_yx_2d_from(grid=grid)
-    ).all()
+    assert mp.deflections_yx_2d_from(grid=grid) == pytest.approx(lmp.deflections_yx_2d_from(grid=grid).array, 1.0e-4)
+
 
 
 def test__exponential():
@@ -106,14 +97,10 @@ def test__exponential():
         mass_to_light_ratio=2.0,
     )
 
-    assert (lp.image_2d_from(grid=grid) == lmp.image_2d_from(grid=grid)).all()
-    assert (
[…]
