+++ test_autogalaxy/operate/test_image.py
@@ -27,7 +27,7 @@ def test__blurred_image_2d_from(
     )
 
     assert blurred_image_2d_manual.native == pytest.approx(
-        lp_blurred_image_2d.native, 1.0e-4
+        lp_blurred_image_2d.native.array, 1.0e-4
     )
 
     lp_blurred_image_2d = lp.blurred_image_2d_from(
@@ -35,7 +35,7 @@ def test__blurred_image_2d_from(
     )
 
     assert blurred_image_2d_manual.native == pytest.approx(
-        lp_blurred_image_2d.native, 1.0e-4
+        lp_blurred_image_2d.native.array, 1.0e-4
     )
 
     light_not_operated = ag.lp.Sersic(intensity=1.0)
@@ -62,7 +62,7 @@ def test__blurred_image_2d_from(
 
     assert (
         blurred_image_2d
-        == pytest.approx(blurred_image_2d_manual_not_operated + image_2d_operated),
+        == pytest.approx(blurred_image_2d_manual_not_operated.array + image_2d_operated.array),
         1.0e-4,
     )
 
@@ -88,7 +88,7 @@ def test__x1_galaxies__padded_image__compare_to_galaxy_images_using_padded_grid_
 
     assert padded_image.shape_native == (9, 9)
     assert padded_image == pytest.approx(
-        padded_g0_image + padded_g1_image + padded_g2_image, 1.0e-4
+        padded_g0_image.array + padded_g1_image.array + padded_g2_image.array, 1.0e-4
     )
 
 
@@ -101,6 +101,7 @@ def test__unmasked_blurred_image_2d_from():
     mask = ag.Mask2D(
         mask=[[True, True, True], [True, False, True], [True, True, True]],
         pixel_scales=1.0,
+        origin=(0.3, 0.3),
     )
 
     grid = ag.Grid2D.from_mask(mask=mask, over_sample_size=1)
@@ -112,9 +113,9 @@ def test__unmasked_blurred_image_2d_from():
     assert unmasked_blurred_image_2d.native == pytest.approx(
         np.array(
             [
-                [1.31618566e-01, 2.48460648e02, 1.91885830e-01],
-                [9.66709359e-02, 8.29737297e01, 1.65678804e02],
-                [4.12176698e-02, 8.74484826e-02, 1.01484934e-01],
+                [0.21305245, 0.6141556, 0.10010613],
+                [0.18998845, 0.50114327, 0.43959065],
+                [0.07793524, 0.16386363, 0.15628968],
             ]
         ),
         1.0e-4,
@@ -189,10 +190,10 @@ def test__blurred_image_2d_list_from(
     )
 
     assert blurred_image_2d_list[0].native == pytest.approx(
-        lp_0_blurred_image_2d.native, 1.0e-4
+        lp_0_blurred_image_2d.native.array, 1.0e-4
     )
     assert blurred_image_2d_list[1].native == pytest.approx(
-        lp_1_blurred_image_2d.native, 1.0e-4
+        lp_1_blurred_image_2d.native.array, 1.0e-4
     )
 
     blurred_image_2d_list = gal.blurred_image_2d_list_from(
@@ -200,10 +201,10 @@ def test__blurred_image_2d_list_from(
     )
 
     assert blurred_image_2d_list[0].native == pytest.approx(
-        lp_0_blurred_image_2d.native, 1.0e-4
+        lp_0_blurred_image_2d.native.array, 1.0e-4
     )
     assert blurred_image_2d_list[1].native == pytest.approx(
-        lp_1_blurred_image_2d.native, 1.0e-4
+        lp_1_blurred_image_2d.native.array, 1.0e-4
     )
 
[…]
