+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -53,7 +53,7 @@ def test__image_2d_from(grid_2d_7x7, gal_x2_lp):
 
     gal_image = gal_x2_lp.image_2d_from(grid=grid_2d_7x7)
 
-    assert gal_image == pytest.approx(lp_image, 1.0e-4)
+    assert gal_image == pytest.approx(lp_image.array, 1.0e-4)
 
 
 def test__image_2d_from__operated_only_input(grid_2d_7x7, lp_0, lp_operated_0):
@@ -298,13 +298,13 @@ def test__light_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric
     assert gal_x2_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[0.0, 0.0]])
     ) == pytest.approx(
-        gal_x2_lp.image_2d_from(grid=ag.Grid2DIrregular([[100.0, 0.0]])), 1.0e-4
+        gal_x2_lp.image_2d_from(grid=ag.Grid2DIrregular([[100.0, 0.0]])).array, 1.0e-4
     )
 
     assert gal_x2_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[49.0, 0.0]])
     ) == pytest.approx(
-        gal_x2_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 0.0]])), 1.0e-4
+        gal_x2_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 0.0]])).array, 1.0e-4
     )
 
     lp_0 = ag.lp.Sersic(
@@ -349,25 +349,25 @@ def test__light_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric
     assert gal_x4_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[49.0, 0.0]])
     ) == pytest.approx(
-        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 0.0]])), 1e-5
+        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 0.0]])).array, 1e-5
     )
 
     assert gal_x4_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[0.0, 49.0]])
     ) == pytest.approx(
-        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[0.0, 51.0]])), 1e-5
+        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[0.0, 51.0]])).array, 1e-5
     )
 
     assert gal_x4_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[100.0, 49.0]])
     ) == pytest.approx(
-        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[100.0, 51.0]])), 1e-5
+        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[100.0, 51.0]])).array, 1e-5
     )
 
     assert gal_x4_lp.image_2d_from(
         grid=ag.Grid2DIrregular([[49.0, 49.0]])
     ) == pytest.approx(
-        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 51.0]])), 1e-5
+        gal_x4_lp.image_2d_from(grid=ag.Grid2DIrregular([[51.0, 51.0]])).array, 1e-5
     )
 
 
@@ -395,25 +395,25 @@ def test__mass_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric_
     assert gal_x4_mp.potential_2d_from(
         grid=ag.Grid2DIrregular([[1.0, 0.0]])
[…]
