+++ test_autogalaxy/galaxy/test_to_inversion.py
@@ -151,7 +151,7 @@ def test__mapper_galaxy_dict(masked_imaging_7x7):
 
 
 def test__inversion_imaging_from(grid_2d_7x7, masked_imaging_7x7):
-    g_linear = ag.Galaxy(redshift=0.5, light_linear=ag.lp_linear.Sersic())
+    g_linear = ag.Galaxy(redshift=0.5, light_linear=ag.lp_linear.Sersic(centre=(0.05, 0.05)))
 
     to_inversion = ag.GalaxiesToInversion(
         dataset=masked_imaging_7x7,
@@ -161,7 +161,7 @@ def test__inversion_imaging_from(grid_2d_7x7, masked_imaging_7x7):
 
     inversion = to_inversion.inversion
 
-    assert inversion.reconstruction[0] == pytest.approx(0.00543437, 1.0e-2)
+    assert inversion.reconstruction[0] == pytest.approx(0.186868464426, 1.0e-2)
 
     pixelization = ag.Pixelization(
         mesh=ag.mesh.Rectangular(shape=(3, 3)),
@@ -184,7 +184,7 @@ def test__inversion_imaging_from(grid_2d_7x7, masked_imaging_7x7):
 
 
 def test__inversion_interferometer_from(grid_2d_7x7, interferometer_7):
-    g_linear = ag.Galaxy(redshift=0.5, light_linear=ag.lp_linear.Sersic())
+    g_linear = ag.Galaxy(redshift=0.5, light_linear=ag.lp_linear.Sersic(centre=(0.05, 0.05)))
 
     to_inversion = ag.GalaxiesToInversion(
         dataset=interferometer_7,
@@ -196,7 +196,7 @@ def test__inversion_interferometer_from(grid_2d_7x7, interferometer_7):
 
     inversion = to_inversion.inversion
 
-    assert inversion.reconstruction[0] == pytest.approx(0.0012073, 1.0e-2)
+    assert inversion.reconstruction[0] == pytest.approx(0.04124846952, 1.0e-2)
 
     interferometer_7.data = ag.Visibilities.ones(shape_slim=(7,))
 
