### 2025-04-07 e1aa7bf
fix test_fit_imaging
