+++ test_autogalaxy/imaging/test_fit_imaging.py
@@ -32,18 +32,18 @@ def test__fit_figure_of_merit(
     masked_imaging_7x7,
     masked_imaging_covariance_7x7,
 ):
-    g0 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
-    g1 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
+    g0 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
+    g1 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, galaxies=[g0, g1])
 
     assert fit.perform_inversion is False
-    assert fit.figure_of_merit == pytest.approx(-75938.05, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-71.46911964, 1.0e-4)
 
     basis = ag.lp_basis.Basis(
         profile_list=[
-            ag.lp.Sersic(intensity=1.0),
-            ag.lp.Sersic(intensity=1.0),
+            ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)),
+            ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)),
         ]
     )
 
@@ -52,7 +52,7 @@ def test__fit_figure_of_merit(
     fit = ag.FitImaging(dataset=masked_imaging_7x7, galaxies=[g0])
 
     assert fit.perform_inversion is False
-    assert fit.figure_of_merit == pytest.approx(-75938.05, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-71.46911964, 1.0e-4)
 
     pixelization = ag.Pixelization(
         mesh=ag.mesh.Rectangular(shape=(3, 3)),
@@ -68,19 +68,19 @@ def test__fit_figure_of_merit(
     assert fit.perform_inversion is True
     assert fit.figure_of_merit == pytest.approx(-22.9005, 1.0e-4)
 
-    galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
+    galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, galaxies=[galaxy_light, galaxy_pix])
 
     assert fit.perform_inversion is True
-    assert fit.figure_of_merit == pytest.approx(-6840.5851, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-29.201919365, 1.0e-4)
 
     g0_linear_light = ag.Galaxy(
-        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=1.0)
+        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=1.0, centre=(0.05, 0.05))
     )
 
     g1_linear_light = ag.Galaxy(
-        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=4.0)
+        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=4.0, centre=(0.05, 0.05))
     )
 
     fit = ag.FitImaging(
@@ -88,12 +88,12 @@ def test__fit_figure_of_merit(
     )
 
     assert fit.perform_inversion is True
-    assert fit.figure_of_merit == pytest.approx(-14.52327, 1.0e-4)
[…]
