+++ test_autogalaxy/interferometer/test_fit_interferometer.py
@@ -14,19 +14,19 @@ def test__model_visibilities(interferometer_7):
 
 
 def test__fit_figure_of_merit(interferometer_7):
-    g0 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
+    g0 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
 
-    g1 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
+    g1 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
 
     fit = ag.FitInterferometer(dataset=interferometer_7, galaxies=[g0, g1])
 
     assert fit.perform_inversion is False
-    assert fit.figure_of_merit == pytest.approx(-2398107.3849, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-1994.35383952, 1.0e-4)
 
     basis = ag.lp_basis.Basis(
         profile_list=[
-            ag.lp.Sersic(intensity=1.0),
-            ag.lp.Sersic(intensity=1.0),
+            ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)),
+            ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)),
         ]
     )
 
@@ -35,7 +35,7 @@ def test__fit_figure_of_merit(interferometer_7):
     fit = ag.FitInterferometer(dataset=interferometer_7, galaxies=[g0])
 
     assert fit.perform_inversion is False
-    assert fit.figure_of_merit == pytest.approx(-2398107.3849, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-1994.3538395, 1.0e-4)
 
     pixelization = ag.Pixelization(
         mesh=ag.mesh.Rectangular(shape=(3, 3)),
@@ -53,7 +53,7 @@ def test__fit_figure_of_merit(interferometer_7):
     assert fit.perform_inversion is True
     assert fit.figure_of_merit == pytest.approx(-66.90612, 1.0e-4)
 
-    galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
+    galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0, centre=(0.05, 0.05)))
 
     pixelization = ag.Pixelization(
         mesh=ag.mesh.Rectangular(shape=(3, 3)),
@@ -69,14 +69,14 @@ def test__fit_figure_of_merit(interferometer_7):
     )
 
     assert fit.perform_inversion is True
-    assert fit.figure_of_merit == pytest.approx(-283424.48941, 1.0e-4)
+    assert fit.figure_of_merit == pytest.approx(-250.2259451235, 1.0e-4)
 
     g0_linear_light = ag.Galaxy(
-        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=1.0)
+        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=1.0, centre=(0.05, 0.05))
     )
 
     g1_linear_light = ag.Galaxy(
-        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=4.0)
+        redshift=0.5, bulge=ag.lp_linear.Sersic(sersic_index=4.0, centre=(0.05, 0.05))
     )
 
     fit = ag.FitInterferometer(
@@ -88,58 +88,35 @@ def test__fit_figure_of_merit(interferometer_7):
 
     basis = ag.lp_basis.Basis(
         profile_list=[
[…]
