### 2025-04-08 0ff3798
test_simulate_and_fit_imaging.py
