### 2025-04-08 d17d1d3
fix unit test in quantity
