### 2025-04-08 bc9672d
fix geometry profiles
