+++ autogalaxy/profiles/geometry_profiles.py
@@ -160,7 +160,7 @@ class SphProfile(GeometryProfile):
         grid
             The (y, x) coordinates in the reference frame of the profile.
         """
-        return jnp.add(jnp.array(grid), jnp.array(self.centre))
+        return jnp.add(grid.array, jnp.array(self.centre))
 
 
 class EllProfile(SphProfile):
@@ -378,7 +378,7 @@ class EllProfile(SphProfile):
             return super().transformed_from_reference_frame_grid_from(grid=grid)
 
         return aa.util.geometry.transform_grid_2d_from_reference_frame(
-            grid_2d=jnp.array(grid), centre=self.centre, angle=self.angle
+            grid_2d=grid.array, centre=self.centre, angle=self.angle
         )
 
     def _eta_u(self, u, coordinates):
