+++ test_autogalaxy/imaging/model/test_result_imaging.py
@@ -48,7 +48,9 @@ def test___image_dict(analysis_imaging_7x7):
 
 def test___linear_light_profiles_in_result(analysis_imaging_7x7):
     galaxies = af.ModelInstance()
-    galaxies.galaxy = ag.Galaxy(redshift=0.5, bulge=ag.lp_linear.Sersic(centre=(0.05, 0.05)))
+    galaxies.galaxy = ag.Galaxy(
+        redshift=0.5, bulge=ag.lp_linear.Sersic(centre=(0.05, 0.05))
+    )
 
     instance = af.ModelInstance()
     instance.galaxies = galaxies
