+++ autogalaxy/profiles/mass/total/isothermal.py
@@ -38,6 +38,7 @@ def psi_from(grid, axis_ratio, core_radius):
         + 1e-16
     )
 
+
 class Isothermal(PowerLaw):
     def __init__(
         self,
@@ -96,10 +97,14 @@ class Isothermal(PowerLaw):
         psi = psi_from(grid=grid, axis_ratio=self.axis_ratio, core_radius=0.0)
 
         deflection_y = jnp.arctanh(
-            jnp.divide(jnp.multiply(jnp.sqrt(1 - self.axis_ratio**2), grid.array[:, 0]), psi)
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1 - self.axis_ratio**2), grid.array[:, 0]), psi
+            )
         )
         deflection_x = jnp.arctan(
-            jnp.divide(jnp.multiply(jnp.sqrt(1 - self.axis_ratio**2), grid.array[:, 1]), psi)
+            jnp.divide(
+                jnp.multiply(jnp.sqrt(1 - self.axis_ratio**2), grid.array[:, 1]), psi
+            )
         )
         return self.rotated_grid_from_reference_frame_from(
             grid=jnp.multiply(factor, jnp.vstack((deflection_y, deflection_x)).T),
@@ -130,12 +135,15 @@ class Isothermal(PowerLaw):
         gamma_2 = (
             -2
             * convergence.array
-            * jnp.divide(grid.array[:, 1] * grid.array[:, 0], grid.array[:, 1] ** 2 + grid.array[:, 0] ** 2)
+            * jnp.divide(
+                grid.array[:, 1] * grid.array[:, 0],
+                grid.array[:, 1] ** 2 + grid.array[:, 0] ** 2,
+            )
         )
         gamma_1 = -convergence.array * jnp.divide(
-            grid.array[:, 1] ** 2 - grid.array[:, 0] ** 2, grid.array[:, 1] ** 2 + grid.array[:, 0] ** 2
+            grid.array[:, 1] ** 2 - grid.array[:, 0] ** 2,
+            grid.array[:, 1] ** 2 + grid.array[:, 0] ** 2,
         )
-
 
         shear_field = self.rotated_grid_from_reference_frame_from(
             grid=jnp.vstack((gamma_2, gamma_1)).T, angle=self.angle * 2
