+++ test_autogalaxy/imaging/test_simulator.py
@@ -37,7 +37,9 @@ def test__from_fits__all_imaging_data_structures_are_flipped_for_ds9():
 
     create_fits(fits_path=image_path, array=[[1.0, 0.0], [0.0, 0.0]])
     create_fits(fits_path=noise_map_path, array=[[2.0, 1.0], [1.0, 1.0]])
-    create_fits(fits_path=psf_path, array=[[1.0, 1.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]])
+    create_fits(
+        fits_path=psf_path, array=[[1.0, 1.0, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]]
+    )
 
     dataset = ag.Imaging.from_fits(
         data_path=image_path,
@@ -48,7 +50,9 @@ def test__from_fits__all_imaging_data_structures_are_flipped_for_ds9():
 
     assert (dataset.data.native == np.array([[0.0, 0.0], [1.0, 0.0]])).all()
     assert (dataset.noise_map.native == np.array([[1.0, 1.0], [2.0, 1.0]])).all()
-    assert dataset.psf.native == pytest.approx(np.array([[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.5, 0.5, 0.0]]), 1.0e-4)
+    assert dataset.psf.native == pytest.approx(
+        np.array([[0.0, 0.0, 0.0], [0.0, 0.0, 0.0], [0.5, 0.5, 0.0]]), 1.0e-4
+    )
 
     dataset.output_to_fits(
         data_path=image_path,
@@ -67,7 +71,9 @@ def test__from_fits__all_imaging_data_structures_are_flipped_for_ds9():
 
     hdu_list = fits.open(psf_path)
     psf = np.array(hdu_list[0].data).astype("float64")
-    assert psf == pytest.approx(np.array([[0.5, 0.5, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]]), 1.0e-4)
+    assert psf == pytest.approx(
+        np.array([[0.5, 0.5, 0.0], [0.0, 0.0, 0.0], [0.0, 0.0, 0.0]]), 1.0e-4
+    )
 
     clean_fits(fits_path=fits_path)
 
