### 2025-04-09 e7c59d3
clean up some code to match autolens behaivour
