### 2025-04-09 ed5d025
fix more unit tests
