+++ test_autogalaxy/profiles/mass/dark/test_nfw_truncated_mcr.py
@@ -42,18 +42,10 @@ def test__duffy__mass_and_concentration_consistent_with_normal_truncated_nfw():
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
-
     assert mp.axis_ratio == 1.0
-    assert isinstance(mp.axis_ratio, float)
-
     assert mp.angle == 0.0
-    assert isinstance(mp.angle, float)
-
     assert mp.inner_slope == 1.0
-    assert isinstance(mp.inner_slope, float)
 
     assert mp.scale_radius == pytest.approx(0.273382, 1.0e-4)
 
@@ -94,18 +86,10 @@ def test__ludlow__mass_and_concentration_consistent_with_normal_truncated_nfw__s
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
-
     assert mp.axis_ratio == 1.0
-    assert isinstance(mp.axis_ratio, float)
-
     assert mp.angle == 0.0
-    assert isinstance(mp.angle, float)
-
     assert mp.inner_slope == 1.0
-    assert isinstance(mp.inner_slope, float)
 
     assert mp.scale_radius == pytest.approx(0.21157, 1.0e-4)
     assert mp.truncation_radius == pytest.approx(33.7134116, 1.0e-4)
