+++ test_autogalaxy/profiles/mass/dark/test_nfw_mcr.py
@@ -41,18 +41,13 @@ def test__mass_and_concentration_consistent_with_normal_nfw():
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
 
     assert mp.axis_ratio == 1.0
-    assert isinstance(mp.axis_ratio, float)
 
     assert mp.angle == 0.0
-    assert isinstance(mp.angle, float)
 
     assert mp.inner_slope == 1.0
-    assert isinstance(mp.inner_slope, float)
 
     assert mp.scale_radius == pytest.approx(0.273382, 1.0e-4)
 
@@ -92,19 +87,10 @@ def test__mass_and_concentration_consistent_with_normal_nfw__scatter_0():
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
-
     assert mp.axis_ratio == 1.0
-    assert isinstance(mp.axis_ratio, float)
-
     assert mp.angle == 0.0
-    assert isinstance(mp.angle, float)
-
     assert mp.inner_slope == 1.0
-    assert isinstance(mp.inner_slope, float)
-
     assert mp.scale_radius == pytest.approx(0.21157, 1.0e-4)
 
     deflections_ludlow = mp.deflections_yx_2d_from(grid=grid)
@@ -150,20 +136,13 @@ def test__same_as_above_but_elliptical():
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
 
     axis_ratio, angle = ag.convert.axis_ratio_and_angle_from(ell_comps=(0.1, 0.2))
 
     assert mp.axis_ratio == axis_ratio
-    assert isinstance(mp.axis_ratio, float)
-
     assert mp.angle == angle
-    assert isinstance(mp.angle, float)
-
     assert mp.inner_slope == 1.0
-    assert isinstance(mp.inner_slope, float)
 
     assert mp.scale_radius == pytest.approx(0.211578, 1.0e-4)
 
@@ -212,20 +191,13 @@ def test__same_as_above_but_generalized_elliptical():
     assert mass_at_200_via_kappa_s == mass_at_200_via_mass
     assert concentration_via_kappa_s == concentration_via_mass
 
-    assert isinstance(mp.kappa_s, float)
-
     assert mp.centre == (1.0, 2.0)
 
     axis_ratio, angle = ag.convert.axis_ratio_and_angle_from(ell_comps=(0.1, 0.2))
 
     assert mp.axis_ratio == axis_ratio
-    assert isinstance(mp.axis_ratio, float)
-
     assert mp.angle == angle
-    assert isinstance(mp.angle, float)
-
     assert mp.inner_slope == 2.0
-    assert isinstance(mp.inner_slope, float)
 
     assert mp.scale_radius == pytest.approx(0.21157, 1.0e-4)
 
