### 2025-04-13 cd2d5ba
now outputs inversion results as a fits table
