### 2025-04-14 #235 @coatless
---
---

Thanks for submitting `PyAutoGalaxy` to pyOpenSci! The package is in ship shape to move forward with our review process. The documentation and functionality is excellent! Well done on creating actionable streams to get users immediately able to analyzing morphoologies and structures of galaxies!

The only minor nit is a section in the `README.md` clearly labeled as "Installation Instructions". That said, this is largely covered by "[Installation Guide](https://pyautogalaxy.readthedocs.io/en/latest/installation/overview.html)" link. However, it would be great if you could include a brief _separate_ snippet. Though, I'm going to go ahead and move the process along by assigning to @hamogu.
