### 2025-04-21 ef6a6f3
added fits_are_Zoomed to config
