### 2025-04-21 1e4cbb3
fits output uses native_for_fits
