+++ autogalaxy/quantity/model/plotter_interface.py
@@ -27,12 +27,17 @@ class PlotterInterfaceQuantity(PlotterInterface):
         dataset
             The imaging dataset which is visualized.
         """
+
+        image_list = [
+            dataset.data.native_for_fits,
+            dataset.noise_map.native_for_fits,
+        ]
+
         hdu_list = hdu_list_for_output_from(
             values_list=[
-                dataset.mask.astype("float"),
-                dataset.data.native,
-                dataset.noise_map.native,
-            ],
+                image_list[0].mask.astype("float"),
+            ]
+            + image_list,
             ext_name_list=[
                 "mask",
                 "data",
