### 2025-04-29 1f1c919
fix one galaxy images
