+++ autogalaxy/galaxy/galaxy.py
@@ -230,7 +230,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
         ):
             return sum(self.image_2d_list_from(grid=grid, operated_only=operated_only))
 
-        return jnp.zeros((grid.shape[0],))
+        return np.zeros((grid.shape[0],))
+#        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def image_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -262,8 +263,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 image_1d_list.append(light_profile.image_1d_from(grid=grid_radial))
 
             return sum(image_1d_list)
-
-        return jnp.zeros((grid.shape[0],))
+        return np.zeros((grid.shape[0],))
+     #   return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_vector_yx
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -289,9 +290,9 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 map(
                     lambda p: p.deflections_yx_2d_from(grid=grid),
                     self.cls_list_from(cls=MassProfile),
-                )
-            )
-        return jnp.zeros((grid.shape[0], 2))
+                ))
+        return np.zeros((grid.shape[0], 2))
+#        return jnp.zeros((grid.shape[0], 2))
 
     @aa.grid_dec.to_array
     def convergence_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -318,7 +319,9 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                     self.cls_list_from(cls=MassProfile),
                 )
             )
-        return jnp.zeros((grid.shape[0],))
+
+        return np.zeros((grid.shape[0],))
+#        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_grid
     def traced_grid_2d_from(self, grid: aa.type.Grid2DLike) -> aa.type.Grid2DLike:
@@ -368,7 +371,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
 
             return sum(convergence_1d_list)
 
-        return jnp.zeros((grid.shape[0],))
+        return np.zeros((grid.shape[0],))
+  #     return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_array
     def potential_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -395,7 +399,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                     self.cls_list_from(cls=MassProfile),
                 )
             )
-        return jnp.zeros((grid.shape[0],))
+        return np.zeros((grid.shape[0],))
+#        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def potential_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -427,8 +432,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 )
 
             return sum(potential_1d_list)
-
-        return jnp.zeros((grid.shape[0],))
+        return np.zeros((grid.shape[0],))
+#        return jnp.zeros((grid.shape[0],))
 
     @property
     def half_light_radius(self):
