### 2025-05-06 2a32c64
aggregator works with FactorGraph
