+++ test_autogalaxy/aggregator/imaging/test_aggregator_fit_imaging.py
@@ -45,7 +45,11 @@ def test__fit_imaging__max_log_likelihood__analysis_has_single_dataset(
     fit_agg = ag.agg.FitImagingAgg(aggregator=agg_7x7)
     fit_max_lh_gen = fit_agg.max_log_likelihood_gen_from()
 
-    for fit_list in fit_max_lh_gen:  # Only Max LH sample so fit_list contains 1 lists of a single fit.
+    for (
+        fit_list
+    ) in (
+        fit_max_lh_gen
+    ):  # Only Max LH sample so fit_list contains 1 lists of a single fit.
 
         assert fit_list[0].galaxies[0].redshift == 0.5
         assert fit_list[0].galaxies[0].light.centre == (10.0, 10.0)
@@ -64,7 +68,11 @@ def test__fit_imaging__randomly_drawn_via_pdf_gen_from__analysis_has_single_data
     i = 0
 
     for fit_list_gen in fit_pdf_gen:  # 1 Dataset so just one fit
-        for fit_list in fit_list_gen:  # Iterate over each total_samples=3, each with two fits for each analysis.
+        for (
+            fit_list
+        ) in (
+            fit_list_gen
+        ):  # Iterate over each total_samples=3, each with two fits for each analysis.
 
             i += 1
 
@@ -72,7 +80,6 @@ def test__fit_imaging__randomly_drawn_via_pdf_gen_from__analysis_has_single_data
             assert fit_list[0].galaxies[0].redshift == 0.5
             assert fit_list[0].galaxies[0].light.centre == (10.0, 10.0)
             assert fit_list[0].dataset_model.background_sky_level == 10.0
-
 
     assert i == 3
 
