### 2025-05-07 85003ef
try / except on dataset model
