### 2025-05-07 6653e1c
'Updated version in __init__ to 2025.5.7.12
