+++ autogalaxy/__init__.py
@@ -117,4 +117,4 @@ from autoconf.fitsable import hdu_list_for_output_from
 
 conf.instance.register(__file__)
 
-__version__ = "2025.5.7.12"
+__version__ = "2025.5.7.15"
