+++ pyproject.toml
@@ -8,6 +8,7 @@ dynamic = ["version", "dependencies"]
 description="Open-Source Multi Wavelength Galaxy Structure & Morphology"
 readme = { file = "README.rst", content-type = "text/x-rst" }
 license = { text = "MIT License" }
+license-file = "LICENSE"
 requires-python = ">=3.9"
 authors = [
     { name = "James Nightingale", email = "James.Nightingale@newcastle.ac.uk" },
