+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "setuptools-scm", "wheel"]
+requires = ["setuptools>=79.0", "setuptools-scm", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -7,8 +7,7 @@ name = "autogalaxy"
 dynamic = ["version", "dependencies"]
 description="Open-Source Multi Wavelength Galaxy Structure & Morphology"
 readme = { file = "README.rst", content-type = "text/x-rst" }
-license = { text = "MIT License" }
-license-file = "LICENSE"
+license = { text = "MIT" }
 requires-python = ">=3.9"
 authors = [
     { name = "James Nightingale", email = "James.Nightingale@newcastle.ac.uk" },
