+++ pyproject.toml
@@ -7,7 +7,10 @@ name = "autogalaxy"
 dynamic = ["version", "dependencies"]
 description="Open-Source Multi Wavelength Galaxy Structure & Morphology"
 readme = { file = "README.rst", content-type = "text/x-rst" }
-license = { text = "MIT" }
+license = "MIT"
+license-files = [
+    "LICENSE",
+]
 requires-python = ">=3.9"
 authors = [
     { name = "James Nightingale", email = "James.Nightingale@newcastle.ac.uk" },
