+++ pyproject.toml
@@ -7,7 +7,6 @@ name = "autogalaxy"
 dynamic = ["version", "dependencies"]
 description="Open-Source Multi Wavelength Galaxy Structure & Morphology"
 readme = { file = "README.rst", content-type = "text/x-rst" }
-license = "MIT"
 license-files = [
     "LICENSE",
 ]
