+++ requirements.txt
@@ -1,3 +1,0 @@
-colossus==1.3.1
-astropy>=5.0,<=6.1.2
-nautilus-sampler==1.0.4
