+++ pyproject.toml
@@ -26,6 +26,13 @@ classifiers = [
   "Programming Language :: Python :: 3.12"
 ]
 keywords = ["cli"]
+dependencies = [
+    "autofit",
+    "autoarray",
+    "colossus==1.3.1",
+    "astropy>=5.0,<=6.1.2",
+    "nautilus-sampler==1.0.4"
+]
 
 [project.urls]
 Homepage = "https://github.com/Jammy2211/PyAutoGalaxy"
