### 2025-05-09 bb44a33
remove setup.py
