+++ pyproject.toml
@@ -43,9 +43,6 @@ include-package-data = true
 [tool.setuptools.packages.find]
 exclude = ["docs", "test_autogalaxy", "test_autogalaxy*"]
 
-[tool.setuptools.dynamic]
-dependencies = { file = "requirements.txt" }
-
 [tool.setuptools_scm]
 version_scheme = "post-release"
 local_scheme = "no-local-version"
