### 2025-05-09 7d7ba27
'Updated version in __init__ to 2025.5.9.36
