+++ optional_requirements.txt
@@ -1,8 +1,0 @@
-numba
-pylops>=1.10.0,<=2.3.1
-pynufft
-ultranest==3.6.2
-zeus-mcmc==2.5.4
-getdist==1.4
-#jax>=0.4.13
-#jaxlib>=0.4.13
