+++ pyproject.toml
@@ -34,17 +34,6 @@ dependencies = [
     "nautilus-sampler==1.0.4"
 ]
 
-[project.optional-dependencies]
-optional=[
-    "numba",
-    "pylops>=1.10.0,<=2.3.1",
-    "pynufft",
-    "ultranest==3.6.2",
-    "zeus-mcmc==2.5.4",
-    "getdist==1.4"
-]
-
-
 [project.urls]
 Homepage = "https://github.com/Jammy2211/PyAutoGalaxy"
 
@@ -60,6 +49,14 @@ local_scheme = "no-local-version"
 
 
 [project.optional-dependencies]
+optional=[
+    "numba",
+    "pylops>=1.10.0,<=2.3.1",
+    "pynufft",
+    "ultranest==3.6.2",
+    "zeus-mcmc==2.5.4",
+    "getdist==1.4"
+]
 test = ["pytest"]
 dev = ["pytest", "black"]
 
