### 2025-05-13 e5e5625
refactored dPIE to use autolens grid API
