### 2025-05-13 19e0ac5
refactored and implemented
