### 2025-05-14 #235 @hamogu
Hi @Jammy2211! I just want to let you know that I'm still looking for a second reviewer. I've contacted three people so far, one declined and two never responded. I know with summer approaching it's a busy time for many and I hope I'll find someone soon. I just need to ask for a little more patience; I'm sorry it takes to long and I promise I keep working on it.
