+++ docs/requirements.txt
@@ -1,12 +1,0 @@
-autoconf
-astropy>=5.0
-colossus==1.3.1
-numba
-pyprojroot==0.2.0
-sphinx
-furo
-myst-parser
-sphinx_copybutton
-sphinx_design
-sphinx_inline_tabs
-sphinx_autodoc_typehints
