+++ readthedocs.yaml
@@ -1,13 +1,16 @@
 version: 2
 
 build:
-  os: ubuntu-20.04
+  os: ubuntu-22.04
   tools:
     python: "3.11"
+  jobs:
+    pre_build: pip install .[docs]  # install your package + docs extras
 
 python:
   install:
-  - requirements: docs/requirements.txt
+    method: pip
+    path: .  # install from current directory
 
 sphinx:
   configuration: docs/conf.py
