+++ readthedocs.yaml
@@ -4,13 +4,13 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
-  jobs:
-    pre_build: pip install .[docs]  # install your package + docs extras
 
 python:
   install:
-    method: pip
-    path: .  # install from current directory
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
 
 sphinx:
   configuration: docs/conf.py
