+++ pyproject.toml
@@ -45,7 +45,7 @@ exclude = ["docs", "test_autogalaxy", "test_autogalaxy*"]
 
 [tool.setuptools_scm]
 version_scheme = "post-release"
-local_scheme = "no-local-version"
+local_scheme = "no-local-version" 
 
 
 [project.optional-dependencies]
@@ -56,7 +56,7 @@ optional=[
     "ultranest==3.6.2",
     "zeus-mcmc==2.5.4",
     "getdist==1.4"
-] 
+]
 docs=[
     "sphinx",
     "furo",
