+++ test_autogalaxy/operate/test_deflections.py
@@ -67,6 +67,20 @@ def test__time_delay_geometry_term_from():
 
     assert time_delay_geometry_term == pytest.approx(np.array([1.92815688, 1.97625436]), 1.0e-4)
 
+
+def test__fermat_potential_from():
+
+    grid = ag.Grid2DIrregular(values=[(0.7, 0.5), (1.0, 1.0)])
+
+    mp = ag.mp.Isothermal(
+        centre=(0.0, 0.0), ell_comps=(0.0, -0.111111), einstein_radius=2.0
+    )
+
+    fermat_potential = mp.fermat_potential_from(grid=grid)
+
+    assert fermat_potential == pytest.approx(np.array([0.24329033, -0.82766592]), 1.0e-4)
+
+
 def test__hessian_from():
     grid = ag.Grid2DIrregular(values=[(0.5, 0.5), (1.0, 1.0)])
 
