### 2025-06-10 83c9cde
fix dual isothermal
