+++ autogalaxy/analysis/plotter_interface.py
@@ -236,10 +236,15 @@ class PlotterInterface:
             mapper_list = inversion.cls_list_from(cls=aa.AbstractMapper)
 
             for i, mapper in enumerate(mapper_list):
-                y = mapper.mapper_grids.source_plane_mesh_grid[:, 0]
-                x = mapper.mapper_grids.source_plane_mesh_grid[:, 1]
+                y = mapper.mapper_grids.source_plane_mesh_grid[:, 0].array
+                x = mapper.mapper_grids.source_plane_mesh_grid[:, 1].array
                 reconstruction = inversion.reconstruction_dict[mapper]
                 noise_map = inversion.reconstruction_noise_map_dict[mapper]
+
+                print(y)
+                print(x)
+                print(reconstruction)
+                print(noise_map)
 
                 with open(
                     self.image_path / f"source_plane_reconstruction_{i}.csv",
