+++ test_autogalaxy/analysis/test_plotter_interface.py
@@ -71,7 +71,8 @@ def test__inversion(
 
         for row in reader:
             for key, value in zip(header_list, row):
-                reconstruction_dict[key].append(value.astype(float))
+                print(value)
+                reconstruction_dict[key].append(float(value))
 
         # Convert lists to NumPy arrays
         for key in reconstruction_dict:
