### 2025-06-10 898c05c
fixed operate image uni test
