### 2025-06-11 0f1514e
operate image uses JAX, fix test
