+++ autogalaxy/convert.py
@@ -1,10 +1,8 @@
 from astropy import units
-from autofit.jax_wrapper import numpy as np, use_jax
 from typing import Tuple
 
-if use_jax:
-    import jax
-
+import jax
+import jax.numpy as jnp
 
 def ell_comps_from(axis_ratio: float, angle: float) -> Tuple[float, float]:
     """
@@ -25,10 +23,10 @@ def ell_comps_from(axis_ratio: float, angle: float) -> Tuple[float, float]:
     angle
         Rotation angle of light profile counter-clockwise from positive x-axis.
     """
-    angle *= np.pi / 180.0
+    angle *= jnp.pi / 180.0
     fac = (1 - axis_ratio) / (1 + axis_ratio)
-    ellip_y = fac * np.sin(2 * angle)
-    ellip_x = fac * np.cos(2 * angle)
+    ellip_y = fac * jnp.sin(2 * angle)
+    ellip_x = fac * jnp.cos(2 * angle)
     return (ellip_y, ellip_x)
 
 
@@ -62,23 +60,14 @@ def axis_ratio_and_angle_from(ell_comps: Tuple[float, float]) -> Tuple[float, fl
     ell_comps
         The elliptical components of the light or mass profile which are converted to an angle.
     """
-    angle = np.arctan2(ell_comps[0], ell_comps[1]) / 2
-    angle *= 180.0 / np.pi
+    angle = jnp.arctan2(ell_comps[0], ell_comps[1]) / 2
+    angle *= 180.0 / jnp.pi
 
-    if use_jax:
-        angle = jax.lax.select(angle < -45, angle + 180, angle)
-    else:
-        if abs(angle) > 45 and angle < 0:
-            angle += 180
+    angle = jax.lax.select(angle < -45, angle + 180, angle)
 
-    fac = np.sqrt(ell_comps[1] ** 2 + ell_comps[0] ** 2)
-    if use_jax:
-        fac = jax.lax.min(fac, 0.999)
-    else:
-        fac = min(fac, 0.999)
-    # if fac > 0.999:
-    #     fac = 0.999  # avoid unphysical solution
-    # if fac > 1: print('unphysical e1,e2')
+    fac = jnp.sqrt(ell_comps[1] ** 2 + ell_comps[0] ** 2)
+    fac = jax.lax.min(fac, 0.999)
+
     axis_ratio = (1 - fac) / (1 + fac)
     return axis_ratio, angle
 
@@ -150,15 +139,15 @@ def shear_gamma_1_2_from(magnitude: float, angle: float) -> Tuple[float, float]:
 
     The gamma 1 and gamma 2 components of a shear are given by:
 
-    gamma_1 = magnitude * np.cos(2 * angle * np.pi / 180.0)
-    gamma_2 = magnitude * np.sin(2 * angle * np.pi / 180.0)
+    gamma_1 = magnitude * jnp.cos(2 * angle * jnp.pi / 180.0)
+    gamma_2 = magnitude * jnp.sin(2 * angle * jnp.pi / 180.0)
 
     Which are the values this function returns.
 
     Converting from gamma 1 and gamma 2 to magnitude and angle is given by:
 
-    magnitude = np.sqrt(gamma_1**2 + gamma_2**2)
-    angle = np.arctan2(gamma_2, gamma_1) / 2 * 180.0 / np.pi
+    magnitude = jnp.sqrt(gamma_1**2 + gamma_2**2)
+    angle = jnp.arctan2(gamma_2, gamma_1) / 2 * 180.0 / jnp.pi
 
     Parameters
     ----------
[…]
