+++ autogalaxy/plot/visuals/two_d.py
@@ -81,7 +81,9 @@ class Visuals2D(aplt.Visuals2D):
             )
 
         if self.multiple_images is not None:
-            plotter.multiple_images_scatter.scatter_grid(grid=self.multiple_images.array)
+            plotter.multiple_images_scatter.scatter_grid(
+                grid=self.multiple_images.array
+            )
 
         if self.tangential_critical_curves is not None:
             try:
