+++ autogalaxy/galaxy/galaxy.py
@@ -231,7 +231,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
             return sum(self.image_2d_list_from(grid=grid, operated_only=operated_only))
 
         return np.zeros((grid.shape[0],))
-#        return jnp.zeros((grid.shape[0],))
+
+    #        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def image_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -264,7 +265,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
 
             return sum(image_1d_list)
         return np.zeros((grid.shape[0],))
-     #   return jnp.zeros((grid.shape[0],))
+
+    #   return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_vector_yx
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -290,9 +292,11 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 map(
                     lambda p: p.deflections_yx_2d_from(grid=grid),
                     self.cls_list_from(cls=MassProfile),
-                ))
+                )
+            )
         return np.zeros((grid.shape[0], 2))
-#        return jnp.zeros((grid.shape[0], 2))
+
+    #        return jnp.zeros((grid.shape[0], 2))
 
     @aa.grid_dec.to_array
     def convergence_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -321,7 +325,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
             )
 
         return np.zeros((grid.shape[0],))
-#        return jnp.zeros((grid.shape[0],))
+
+    #        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_grid
     def traced_grid_2d_from(self, grid: aa.type.Grid2DLike) -> aa.type.Grid2DLike:
@@ -372,7 +377,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
             return sum(convergence_1d_list)
 
         return np.zeros((grid.shape[0],))
-  #     return jnp.zeros((grid.shape[0],))
+
+    #     return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_array
     def potential_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -400,7 +406,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 )
             )
         return np.zeros((grid.shape[0],))
-#        return jnp.zeros((grid.shape[0],))
+
+    #        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def potential_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -433,7 +440,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
 
             return sum(potential_1d_list)
         return np.zeros((grid.shape[0],))
-#        return jnp.zeros((grid.shape[0],))
+
+    #        return jnp.zeros((grid.shape[0],))
 
     @property
     def half_light_radius(self):
