+++ autogalaxy/profiles/mass/total/isothermal.py
@@ -107,7 +107,7 @@ class Isothermal(PowerLaw):
         )
         return self.rotated_grid_from_reference_frame_from(
             grid=jnp.multiply(factor, jnp.vstack((deflection_y, deflection_x)).T),
-            **kwargs
+            **kwargs,
         )
 
     @aa.grid_dec.to_vector_yx
@@ -203,5 +203,5 @@ class IsothermalSph(Isothermal):
         return self._cartesian_grid_via_radial_from(
             grid=grid,
             radius=jnp.full(grid.shape[0], 2.0 * self.einstein_radius_rescaled),
-            **kwargs
+            **kwargs,
         )
