### 2025-06-18 f2cf47a
operate image now supports JAX
