### 2025-06-18 296d716
fixed some unit tests due to JAX conversion
