### 2025-06-18 98ed1a4
add uses_imaging to linear absract light profile
