### 2025-06-23 cc8cdba
update log likelihood function to not use try / exception
