+++ autogalaxy/config/priors/mass/stellar/sersic_core.yaml
@@ -3,24 +3,20 @@ SersicCore:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   centre_1:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   effective_radius:
@@ -30,11 +26,11 @@ SersicCore:
     width_modifier:
       type: Relative
       value: 1.0
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   ell_comps_0:
-    type: Gaussian
+    type: TruncatedGaussian
     mean: 0.0
     sigma: 0.3
     lower_limit: -1.0
@@ -42,11 +38,11 @@ SersicCore:
     width_modifier:
       type: Absolute
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: -1.0
       upper: 1.0
   ell_comps_1:
-    type: Gaussian
+    type: TruncatedGaussian
     mean: 0.0
     sigma: 0.3
     lower_limit: -1.0
@@ -54,7 +50,7 @@ SersicCore:
     width_modifier:
       type: Absolute
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: -1.0
       upper: 1.0
   intensity:
@@ -64,7 +60,7 @@ SersicCore:
     width_modifier:
       type: Relative
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   alpha:
@@ -83,7 +79,7 @@ SersicCore:
     width_modifier:
       type: Relative
       value: 0.3
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   sersic_index:
@@ -93,7 +89,7 @@ SersicCore:
     width_modifier:
       type: Absolute
       value: 1.5
-    gaussian_limits:
+    limits:
       lower: 0.8
       upper: 5.0
 SersicCoreSph:
@@ -104,24 +100,20 @@ SersicCoreSph:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   centre_1:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   effective_radius:
@@ -131,7 +123,7 @@ SersicCoreSph:
     width_modifier:
       type: Relative
       value: 1.0
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   gamma:
@@ -144,7 +136,7 @@ SersicCoreSph:
     width_modifier:
       type: Relative
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   mass_to_light_ratio:
@@ -154,7 +146,7 @@ SersicCoreSph:
     width_modifier:
       type: Relative
       value: 0.3
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   radius_break:
@@ -167,6 +159,6 @@ SersicCoreSph:
     width_modifier:
[…]
