+++ autogalaxy/config/priors/light/operated/sersic.yaml
@@ -3,24 +3,20 @@ ersic:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   centre_1:
     type: Gaussian
     mean: 0.0
     sigma: 0.3
-    lower_limit: -inf
-    upper_limit: inf
     width_modifier:
       type: Absolute
       value: 0.05
-    gaussian_limits:
+    limits:
       lower: -inf
       upper: inf
   effective_radius:
@@ -30,11 +26,11 @@ ersic:
     width_modifier:
       type: Relative
       value: 1.0
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   ell_comps_0:
-    type: Gaussian
+    type: TruncatedGaussian
     mean: 0.0
     sigma: 0.3
     lower_limit: -1.0
@@ -42,11 +38,11 @@ ersic:
     width_modifier:
       type: Absolute
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: -1.0
       upper: 1.0
   ell_comps_1:
-    type: Gaussian
+    type: TruncatedGaussian
     mean: 0.0
     sigma: 0.3
     lower_limit: -1.0
@@ -54,7 +50,7 @@ ersic:
     width_modifier:
       type: Absolute
       value: 0.2
-    gaussian_limits:
+    limits:
       lower: -1.0
       upper: 1.0
   intensity:
@@ -64,7 +60,7 @@ ersic:
     width_modifier:
       type: Relative
       value: 0.5
-    gaussian_limits:
+    limits:
       lower: 0.0
       upper: inf
   sersic_index:
@@ -74,6 +70,6 @@ ersic:
     width_modifier:
       type: Absolute
       value: 1.5
-    gaussian_limits:
+    limits:
       lower: 0.8
       upper: 5.0

[…]
