+++ autogalaxy/imaging/fit_imaging.py
@@ -25,7 +25,6 @@ class FitImaging(aa.FitImaging, AbstractFitInversion):
         dataset_model: Optional[aa.DatasetModel] = None,
         adapt_images: Optional[AdaptImages] = None,
         settings_inversion: aa.SettingsInversion = aa.SettingsInversion(),
-        run_time_dict: Optional[Dict] = None,
     ):
         """
         Fits an imaging dataset using a list of galaxies.
@@ -64,17 +63,13 @@ class FitImaging(aa.FitImaging, AbstractFitInversion):
             reconstructed galaxy's morphology.
         settings_inversion
             Settings controlling how an inversion is fitted for example which linear algebra formalism is used.
-        run_time_dict
-            A dictionary which if passed to the fit records how long fucntion calls which have the `profile_func`
-            decorator take to run.
         """
 
-        self.galaxies = Galaxies(galaxies=galaxies, run_time_dict=run_time_dict)
+        self.galaxies = Galaxies(galaxies=galaxies)
 
         super().__init__(
             dataset=dataset,
             dataset_model=dataset_model,
-            run_time_dict=run_time_dict,
         )
         AbstractFitInversion.__init__(
             self=self,
@@ -130,7 +125,6 @@ class FitImaging(aa.FitImaging, AbstractFitInversion):
             galaxies=self.galaxies,
             adapt_images=self.adapt_images,
             settings_inversion=self.settings_inversion,
-            run_time_dict=self.run_time_dict,
         )
 
     @cached_property
