+++ autogalaxy/analysis/analysis/analysis.py
@@ -37,7 +37,7 @@ class Analysis(af.Analysis):
         self.cosmology = cosmology
 
     def galaxies_via_instance_from(
-        self, instance: af.ModelInstance, run_time_dict: Optional[Dict] = None
+        self, instance: af.ModelInstance,
     ) -> List[Galaxy]:
         """
         Create a list of galaxies from a model instance, which is used to fit the dataset.
@@ -60,10 +60,9 @@ class Analysis(af.Analysis):
             if getattr(instance, "extra_galaxies", None) is not None:
                 return Galaxies(
                     galaxies=instance.galaxies + instance.extra_galaxies,
-                    run_time_dict=run_time_dict,
                 )
 
-        return Galaxies(galaxies=instance.galaxies, run_time_dict=run_time_dict)
+        return Galaxies(galaxies=instance.galaxies)
 
     def dataset_model_via_instance_from(
         self, instance: af.ModelInstance
