+++ autogalaxy/interferometer/model/analysis.py
@@ -142,7 +142,6 @@ class AnalysisInterferometer(AnalysisDataset):
     def fit_from(
         self,
         instance: af.ModelInstance,
-        run_time_dict: Optional[Dict] = None,
     ) -> FitInterferometer:
         """
         Given a model instance create a `FitInterferometer` object.
@@ -157,8 +156,6 @@ class AnalysisInterferometer(AnalysisDataset):
             via a non-linear search).
         preload_overwrite
             If a `Preload` object is input this is used instead of the preloads stored as an attribute in the analysis.
-        run_time_dict
-            A dictionary which times functions called to fit the model to data, for profiling.
 
         Returns
         -------
@@ -166,7 +163,7 @@ class AnalysisInterferometer(AnalysisDataset):
             The fit of the galaxies to the interferometer dataset, which includes the log likelihood.
         """
         galaxies = self.galaxies_via_instance_from(
-            instance=instance, run_time_dict=run_time_dict
+            instance=instance,
         )
 
         adapt_images = self.adapt_images_via_instance_from(instance=instance)
@@ -176,7 +173,6 @@ class AnalysisInterferometer(AnalysisDataset):
             galaxies=galaxies,
             adapt_images=adapt_images,
             settings_inversion=self.settings_inversion,
-            run_time_dict=run_time_dict,
         )
 
     def save_attributes(self, paths: af.DirectoryPaths):
