### 2025-06-23 6ee3423
remove profile_log_likelihood_function
