+++ autogalaxy/analysis/analysis/analysis.py
@@ -125,135 +125,3 @@ class Analysis(af.Analysis):
             search_internal=search_internal,
             analysis=self,
         )
-
-    def profile_log_likelihood_function(
-        self, instance: af.ModelInstance, paths: Optional[af.DirectoryPaths] = None
-    ) -> Tuple[Dict, Dict]:
-        """
-        This function is optionally called throughout a model-fit to profile the log likelihood function.
-
-        All function calls inside the `log_likelihood_function` that are decorated with the `profile_func` are timed
-        with their times stored in a dictionary called the `run_time_dict`.
-
-        An `info_dict` is also created which stores information on aspects of the model and dataset that dictate
-        run times, so the profiled times can be interpreted with this context.
-
-        The results of this profiling are then output to hard-disk in the `profiling` folder of the model-fit results,
-        which they can be inspected to ensure run-times are as expected.
-
-        Parameters
-        ----------
-        instance
-            An instance of the model that is being fitted to the data by this analysis (whose parameters have been set
-            via a non-linear search).
-        paths
-            The paths object which manages all paths, e.g. where the non-linear search outputs are stored,
-            visualization and the pickled objects used by the aggregator output by this function.
-
-        Returns
-        -------
-        Two dictionaries, the profiling dictionary and info dictionary, which contain the profiling times of the
-        `log_likelihood_function` and information on the model and dataset used to perform the profiling.
-        """
-
-        if isinstance(paths, af.DatabasePaths):
-            return
-
-        run_time_dict = {}
-        info_dict = {}
-
-        repeats = conf.instance["general"]["profiling"]["repeats"]
-        info_dict["repeats"] = repeats
-
-        # Ensure numba functions are compiled before profiling begins.
-
-        fit = self.fit_from(instance=instance)
-        fit.figure_of_merit
-
-        start = time.time()
-
-        for _ in range(repeats):
-            try:
-                fit = self.fit_from(instance=instance)
-                fit.figure_of_merit
-            except Exception:
-                logger.info(
-                    "Profiling failed. Returning without outputting information."
-                )
-                return
-
-        fit_time = (time.time() - start) / repeats
-
-        run_time_dict["fit_time"] = fit_time
-
-        fit = self.fit_from(instance=instance, run_time_dict=run_time_dict)
-        fit.figure_of_merit
-
-        try:
-            info_dict["image_pixels"] = self.dataset.grids.lp.shape_slim
-            info_dict["sub_total_light_profiles"] = (
-                self.dataset.grids.lp.over_sampler.sub_total
-            )
-        except AttributeError:
-            pass
-
-        if fit.model_obj.has(cls=aa.Pixelization):
-            info_dict["use_w_tilde"] = fit.inversion.settings.use_w_tilde
-            try:
-                info_dict["sub_total_pixelization"] = (
-                    self.dataset.grids.pixelization.over_sampler.sub_total
-                )
-            except AttributeError:
-                pass
-            info_dict["use_positive_only_solver"] = (
-                fit.inversion.settings.use_positive_only_solver
-            )
-            info_dict["force_edge_pixels_to_zeros"] = (
-                fit.inversion.settings.force_edge_pixels_to_zeros
-            )
-            info_dict["use_w_tilde_numpy"] = fit.inversion.settings.use_w_tilde_numpy
-            info_dict["source_pixels"] = len(fit.inversion.reconstruction)
-
-            if hasattr(fit.inversion, "w_tilde"):
-                info_dict["w_tilde_curvature_preload_size"] = (
-                    fit.inversion.w_tilde.curvature_preload.shape[0]
-                )
-
-        self.output_profiling_info(
-            paths=paths, run_time_dict=run_time_dict, info_dict=info_dict
-        )
-
[…]
