### 2025-06-23 22c6d66
fix aggregator unit test
