### 2025-06-25 415cf0e
implemented
