+++ autogalaxy/profiles/mass/aymeric/epl.py
@@ -1,0 +1,399 @@
+"""
+This code a copy of the following lenstronomy submodules, with some functions removed:
+- https://github.com/lenstronomy/lenstronomy/blob/main/lenstronomy/LensModel/Profiles/epl.py
+- https://github.com/lenstronomy/lenstronomy/blob/main/lenstronomy/Util/param_util.py
+- https://github.com/lenstronomy/lenstronomy/blob/main/lenstronomy/Util/util.py
+
+The function derivatives_q_phi() has been added, for practicality.
+
+All credits go to Nicolas Tessore and the Lenstronomy contributors.
+"""
+
+__author__ = "ntessore, Lenstronomy contributors"
+
+import numpy as np
+from scipy.special import hyp2f1
+from typing import Tuple
+
+__all__ = ["EPL", "EPLMajorAxis"]
+
+
+class EPLEuclid:
+
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        q : float = 1.0,
+        phi: float = 0.0,
+        einstein_radius: float = 1.0,
+        gamma: float = 2.0,
+    ):
+
+        self.centre = centre
+        self.q = q
+        self.phi = phi
+        self.einstein_radius = einstein_radius
+        self.gamma = gamma
+
+    def convergence_2d_from(self, grid, **kwargs):
+        return np.zeros(shape=grid.shape[0])
+
+    def potential_2d_from(self, grid, **kwargs):
+        return np.zeros(shape=grid.shape[0])
+
+    def deflections_yx_2d_from(self, grid, **kwargs):
+        """
+        Calculate the deflection angles at a given set of arc-second gridded coordinates.
+
+        Parameters
+        ----------
+        grid
+            The grid of (y,x) arc-second coordinates the deflection angles are computed on.
+
+        """
+
+        profile = EPL()
+
+        x = grid[:, 1]
+        y = grid[:, 0]
+
+        deflections_x, deflections_y = profile.derivatives_q_phi(
+                x,
+            y,
+                self.einstein_radius,
+                self.gamma,
+                self.q,
+                self.phi,
+                center_x=self.centre[1],
+                center_y=self.centre[0],
+            )
+
+        return np.stack((deflections_y, deflections_x), axis=-1)
+
+
+
+
+def rotate(xcoords, ycoords, angle):
+    """
+
+    :param xcoords: x points
+    :param ycoords: y points
+    :param angle: angle in radians
+    :return: x points and y points rotated ccw by angle theta
+    """
+    return xcoords * np.cos(angle) + ycoords * np.sin(angle), -xcoords * np.sin(
+        angle
+    ) + ycoords * np.cos(angle)
+
+
+def ellipticity2phi_q(e1, e2):
+    """Transforms complex ellipticity moduli in orientation angle and axis ratio.
+
+    :param e1: eccentricity in x-direction
+    :param e2: eccentricity in xy-direction
+    :return: angle in radian, axis ratio (minor/major)
+    """
+    phi = np.arctan2(e2, e1) / 2
+    c = np.sqrt(e1 ** 2 + e2 ** 2)
+    c = np.minimum(c, 0.9999)
+    q = (1 - c) / (1 + c)
+    return phi, q
+
+
+def phi_q2_ellipticity(phi, q):
+    """Transforms orientation angle and axis ratio into complex ellipticity moduli e1,
+    e2.
+
+    :param phi: angle of orientation (in radian)
+    :param q: axis ratio minor axis / major axis
+    :return: eccentricities e1 and e2 in complex ellipticity moduli
+    """
[…]
