+++ autogalaxy/profiles/mass/aymeric/multipole.py
@@ -1,0 +1,176 @@
+"""
+This module contains functions to compute the lensing potential, deflection angles, and Hessian
+for a spherical multipole with radial power-law decay.
+
+Note that only the m=4 case is implemented for the deflection angles.
+
+Credits to Leon Ecker (LMU), based on Chu et al. (2013) and Nightingale et al. (2023).
+"""
+
+__author__ = "Leon Ecker"
+
+import numpy as np
+from typing import Tuple
+
+def cart2polar(x, y, center_x=0, center_y=0):
+    coord_shift_x = x - center_x
+    coord_shift_y = y - center_y
+    r = np.sqrt(coord_shift_x ** 2 + coord_shift_y ** 2)
+    phi = np.arctan2(coord_shift_y, coord_shift_x)
+    return r, phi
+
+
+
+class SphericalPowerlawMultipoleEuclid:
+
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        k_m : float = 1.0,
+        phi_m: float = 0.0,
+        einstein_radius: float = 1.0,
+        gamma: float = 2.0,
+    ):
+
+        self.centre = centre
+        self.k_m = k_m
+        self.phi_m = phi_m
+        self.einstein_radius = einstein_radius
+        self.gamma = gamma
+
+    def convergence_2d_from(self, grid, **kwargs):
+        return np.zeros(shape=grid.shape[0])
+
+    def potential_2d_from(self, grid, **kwargs):
+        return np.zeros(shape=grid.shape[0])
+
+    def deflections_yx_2d_from(self, grid, **kwargs):
+        """
+        Calculate the deflection angles at a given set of arc-second gridded coordinates.
+
+        Parameters
+        ----------
+        grid
+            The grid of (y,x) arc-second coordinates the deflection angles are computed on.
+
+        """
+
+        profile = SphericalPowerlawMultipole()
+
+        x = grid[:, 1]
+        y = grid[:, 0]
+
+        deflections_x, deflections_y = profile.deflection_angles_multipol_m4_o24(
+                x,
+            y,
+                self.einstein_radius,
+                self.gamma,
+                self.k_m,
+                self.phi_m,
+                center_x=self.centre[1],
+                center_y=self.centre[0],
+                m=4,
+            )
+
+        return np.stack((deflections_y, deflections_x), axis=-1)
+
+
+
+__all__ = ["SphericalPowerlawMultipole"]
+
+
+class SphericalPowerlawMultipole(object):
+
+    def function(self, x, y, theta_E, gamma, k_m, phi_m, center_x=0, center_y=0, m=4):
+        """
+        Lensing potential of PLmultipole contribution (for 1 component with m>=2).
+        The equation for PLMultipole is Eq. (8) from Nightingale et al. (2023) (https://arxiv.org/abs/2209.10566)
+        :param x: x-coordinate in image plane
+        :param y: y-coordinate in image plane
+        :param theta_E: Einstein radius
+        :param gamma: power law slope
+        :param k_m : multipole strength and sqrt (a_m^2+b_m^2) of old definition.
+        :param phi_m : multipole orientation in radian, arctan(b_m/a_m) of old definition
+        :param center_x: profile center
+        :param center_y: profile center
+        :return: lensing potential
+        """
+        r, phi = cart2polar(x, y, center_x=center_x, center_y=center_y)
+        r = np.maximum(r, 0.000001)
[…]
