+++ autogalaxy/profiles/mass/aymeric/__init__.py
@@ -1,0 +1,3 @@
+from .epl import EPLEuclid
+from .multipole import SphericalPowerlawMultipoleEuclid
+from .shear import ShearEuclid
