+++ autogalaxy/config/priors/mass/aymeric/epl.yaml
@@ -1,0 +1,65 @@
+EPLEuclid:
+  centre_0:
+    type: Gaussian
+    mean: 0.0
+    sigma: 0.1
+    lower_limit: -inf
+    upper_limit: inf
+    width_modifier:
+      type: Absolute
+      value: 0.05
+    gaussian_limits:
+      lower: -inf
+      upper: inf
+  centre_1:
+    type: Gaussian
+    mean: 0.0
+    sigma: 0.1
+    lower_limit: -inf
+    upper_limit: inf
+    width_modifier:
+      type: Absolute
+      value: 0.05
+    gaussian_limits:
+      lower: -inf
+      upper: inf
+  q:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 1.0
+    width_modifier:
+      type: Relative
+      value: 0.25
+    gaussian_limits:
+      lower: 0.0
+      upper: 1.0
+  phi:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 180.0
+    width_modifier:
+      type: Relative
+      value: 0.25
+    gaussian_limits:
+      lower: 0.0
+      upper: 180.0
+  einstein_radius:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 8.0
+    width_modifier:
+      type: Relative
+      value: 0.25
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+  slope:
+    type: Uniform
+    lower_limit: 1.5
+    upper_limit: 3.0
+    width_modifier:
+      type: Absolute
+      value: 0.2
+    gaussian_limits:
+      lower: 1.0
+      upper: 3.0
