+++ autogalaxy/config/priors/mass/aymeric/shear.yaml
@@ -1,0 +1,21 @@
+ShearEuclid:
+  gamma_ext:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 0.3
+    width_modifier:
+      type: Absolute
+      value: 0.05
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+  phi_ext:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 180.0
+    width_modifier:
+      type: Absolute
+      value: 0.05
+    gaussian_limits:
+      lower: 0.0
+      upper: 90.0

[…]
