### 2025-06-25 f127306
inheritance
