+++ test_autogalaxy/profiles/mass/dark/test_nfw_mcr.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 import autogalaxy as ag
@@ -7,7 +6,9 @@ grid = ag.Grid2DIrregular([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 def test__mass_and_concentration_consistent_with_normal_nfw():
-    cosmology = ag.cosmo.FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
+
+    from autogalaxy.cosmology.model import FlatLambdaCDMWrap
+    cosmology = FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
 
     mp = ag.mp.NFWMCRDuffySph(
         centre=(1.0, 2.0),
@@ -53,7 +54,9 @@ def test__mass_and_concentration_consistent_with_normal_nfw():
 
 
 def test__mass_and_concentration_consistent_with_normal_nfw__scatter_0():
-    cosmology = ag.cosmo.FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
+
+    from autogalaxy.cosmology.model import FlatLambdaCDMWrap
+    cosmology = FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
 
     mp = ag.mp.NFWMCRLudlowSph(
         centre=(1.0, 2.0),
@@ -100,7 +103,9 @@ def test__mass_and_concentration_consistent_with_normal_nfw__scatter_0():
 
 
 def test__same_as_above_but_elliptical():
-    cosmology = ag.cosmo.FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
+
+    from autogalaxy.cosmology.model import FlatLambdaCDMWrap
+    cosmology = FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
 
     mp = ag.mp.NFWMCRLudlow(
         centre=(1.0, 2.0),
@@ -153,7 +158,9 @@ def test__same_as_above_but_elliptical():
 
 
 def test__same_as_above_but_generalized_elliptical():
-    cosmology = ag.cosmo.FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
+
+    from autogalaxy.cosmology.model import FlatLambdaCDMWrap
+    cosmology = FlatLambdaCDMWrap(H0=70.0, Om0=0.3)
 
     mp = ag.mp.gNFWMCRLudlow(
         centre=(1.0, 2.0),
