+++ autogalaxy/util/mock/mock_cosmology.py
@@ -44,3 +44,140 @@ class MockCosmology:
 
     def critical_density(self, z):
         return Value(value=self.cosmic_average_density)
+
+    def cosmic_average_density_solar_mass_per_kpc3_from(self, redshift: float) -> float:
+        """
+        Critical density of the Universe at an input `redshift` in units of solar masses per kiloparsecs**3.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
+        This function therefore returns only the value of the astropy function it wraps, omitting the units instance.
+
+        Parameters
+        ----------
+        redshift
+            Redshift at which the critiical density in solMass/kpc^3 of the Universe is calculated.
+        """
+        cosmic_average_density_kpc = (
+            self.critical_density(z=redshift).to("solMass / kpc^3").value
+        )
+
+        return cosmic_average_density_kpc
+
+
+    def critical_surface_density_between_redshifts_solar_mass_per_kpc2_from(
+        self, redshift_0: float, redshift_1: float
+    ) -> float:
+        """
+        The critical surface density for lensing, often written as $\sigma_{cr}$, is given by:
+
+        critical_surface_density = (c^2 * D_s) / (4 * pi * G * D_ls * D_l)
+
+        c = speed of light
+        G = Newton's gravity constant
+        D_s = Angular diameter distance of source redshift to earth
+        D_ls = Angular diameter distance of lens redshift to source redshift
+        D_l = Angular diameter distance of lens redshift to earth
+
+        This function returns the critical surface density in units of solar masses / kpc^2, which are convenient
+        units for converting the inferred masses of a model from angular units (e.g. dimensionless units inferred
+        from data in arcseconds) to solar masses.
+
+        Parameters
+        ----------
+        redshift_0
+            The redshift of the first strong lens galaxy (E.g. the lens galaxy) for which the critical surface
+            density is calculated.
+        redshift_1
+            The redshift of the second strong lens galaxy (E.g. the lens galaxy) for which the critical surface
+            density is calculated.
+        """
+        from astropy import constants
+
+        const = constants.c.to("kpc / s") ** 2.0 / (
+            4 * math.pi * constants.G.to("kpc3 / (solMass s2)")
+        )
+
+        angular_diameter_distance_of_redshift_0_to_earth_kpc = (
+            self.angular_diameter_distance_to_earth_in_kpc_from(redshift=redshift_0)
+        )
+
+        angular_diameter_distance_of_redshift_1_to_earth_kpc = (
+            self.angular_diameter_distance_to_earth_in_kpc_from(redshift=redshift_1)
+        )
+
+        angular_diameter_distance_between_redshifts_kpc = (
+            self.angular_diameter_distance_between_redshifts_in_kpc_from(
+                redshift_0=redshift_0, redshift_1=redshift_1
+            )
+        )
+
+        return (
+            const
+            * angular_diameter_distance_of_redshift_1_to_earth_kpc
+            / (
+                angular_diameter_distance_between_redshifts_kpc
+                * angular_diameter_distance_of_redshift_0_to_earth_kpc
+            )
+        ).value
+
+    def angular_diameter_distance_to_earth_in_kpc_from(self, redshift: float) -> float:
+        """
+        Angular diameter distance from the input `redshift` to redshift zero (e.g. us, the observer on earth) in
+        kiloparsecs.
+
+        This gives the proper (sometimes called 'physical') transverse distance corresponding to an angle of 1 radian
+        for an object at redshift `z`.
+
+        Weinberg, 1972, pp 421-424; Weedman, 1986, pp 65-67; Peebles, 1993, pp 325-327.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
[…]
