### 2025-06-25 6f86eb1
remove some slow tests which dont do much
