### 2025-06-25 8996d73
remove use of numba
