+++ autogalaxy/profiles/mass/dark/gnfw.py
@@ -1,84 +1,9 @@
-import inspect
 import numpy as np
 from typing import Tuple
 
 import autoarray as aa
 
 from autogalaxy.profiles.mass.dark.abstract import AbstractgNFW
-
-
-def jit_integrand(integrand_function):
-    from scipy import LowLevelCallable
-
-    from numba import cfunc
-    from numba.types import intc, CPointer, float64
-
-    jitted_function = aa.util.numba.jit(nopython=True, cache=True)(integrand_function)
-    no_args = len(inspect.getfullargspec(integrand_function).args)
-
-    wrapped = None
-
-    if no_args == 4:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(xx[0], xx[1], xx[2], xx[3])
-
-    elif no_args == 5:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(xx[0], xx[1], xx[2], xx[3], xx[4])
-
-    elif no_args == 6:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(xx[0], xx[1], xx[2], xx[3], xx[4], xx[5])
-
-    elif no_args == 7:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(xx[0], xx[1], xx[2], xx[3], xx[4], xx[5], xx[6])
-
-    elif no_args == 8:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(
-                xx[0], xx[1], xx[2], xx[3], xx[4], xx[5], xx[6], xx[7]
-            )
-
-    elif no_args == 9:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(
-                xx[0], xx[1], xx[2], xx[3], xx[4], xx[5], xx[6], xx[7], xx[8]
-            )
-
-    elif no_args == 10:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(
-                xx[0], xx[1], xx[2], xx[3], xx[4], xx[5], xx[6], xx[7], xx[8], xx[9]
-            )
-
-    elif no_args == 11:
-        # noinspection PyUnusedLocal
-        def wrapped(n, xx):
-            return jitted_function(
-                xx[0],
-                xx[1],
-                xx[2],
-                xx[3],
-                xx[4],
-                xx[5],
-                xx[6],
-                xx[7],
-                xx[8],
-                xx[9],
-                xx[10],
-            )
-
-    cf = cfunc(float64(intc, CPointer(float64)))
-
-    return LowLevelCallable(cf(wrapped).ctypes)
 
 
 class gNFW(AbstractgNFW):
@@ -109,7 +34,6 @@ class gNFW(AbstractgNFW):
 
         """
 
-        @jit_integrand
         def surface_density_integrand(x, kappa_radius, scale_radius, inner_slope):
             return (
                 (3 - inner_slope)
@@ -254,7 +178,6 @@ class gNFW(AbstractgNFW):
         from scipy import special
         from scipy.integrate import quad
 
-        @jit_integrand
         def deflection_integrand(x, kappa_radius, scale_radius, inner_slope):
             return (x + kappa_radius / scale_radius) ** (inner_slope - 3) * (
                 (1 - np.sqrt(1 - x**2)) / x
