### 2025-06-25 72944ee
remove use of numba
