### 2025-06-26 cdf4137
another attempt to fix Aymmeric
