+++ autogalaxy/imaging/model/plotter_interface.py
@@ -1,5 +1,7 @@
-from pathlib import Path
-from typing import List
+from __future__ import annotations
+from typing import List, TYPE_CHECKING
+if TYPE_CHECKING:
+    from pathlib import Path
 
 from autoconf.fitsable import hdu_list_for_output_from
 
