+++ autogalaxy/analysis/analysis/analysis.py
@@ -1,6 +1,6 @@
 import json
 import logging
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple
 from os import path
 import os
 import time
