### 2025-07-02 #235 @canorve
- [X] Package has a README.rst file in the root directory.

    - [X] Code style.
    - [X] Launch binder.
    - [X] JOSS paper.

7 hours

---

First of all, I apologize for the delay in completing this review.

PyAutoGalaxy is a well-documented and scientifically relevant tool for
galaxy morphology modeling. Its strengths include comprehensive tutorials,
research-focused functionality, and adherence to modern Python packaging
standards. The project has strong potential for continued development,
and I believe it will see broad adoption within the astronomical community.

Below are my comments, most of which I believe are minor.

##### API documentation

API documentation is missing for the following modules:

 * `autogalaxy.plot.Visuals1D`: [link](https://pyautogalaxy.readthedocs.io/en/latest/api/_autosummary/autogalaxy.plot.Visuals1D.html)
 * `autogalaxy.plot.Visuals2D`: [link](https://pyautogalaxy.readthedocs.io/en/latest/api/_autosummary/autogalaxy.plot.Visuals2D.html)

Some class methods are not documented. For example:

* `convergence_func`, `convergence_func_gaussian`, `coord_func_f`, `coord_func_f_jit`, `coord_func_g`, `coord_func_g_jit`, `decompose_convergence_via_mge`, `deflection_func_sph`, `deflections_2d_via_potential_2d_from`, and others in:

  [autogalaxy.profiles.mass.NFWTruncatedMCRLudlowSph](https://pyautogalaxy.readthedocs.io/en/latest/api/_autosummary/autogalaxy.profiles.mass.NFWTruncatedMCRLudlowSph.html)

* `concentration_func`, `contour_list_from`, `convergence_1d_from`, and others in:

  [autogalaxy.profiles.mass.NFWMCRScatterLudlowSph](https://pyautogalaxy.readthedocs.io/en/latest/api/_autosummary/autogalaxy.profiles.mass.NFWMCRScatterLudlowSph.html)

Providing documentation for these methods would enhance clarity and usability, especially for new users exploring the API.

##### README

The top-to-bottom order suggested in the template is not followed.

A statement of need is present in the documentation and in the paper, but it is not clearly stated in the README.

No installation instructions are provided directly in the README; only a link to the installation guide is included.

Additional setup requirements for using the package are documented, but not mentioned in the README.

Citation information is available in the documentation but not included in the README.

##### Oversampling

Why does PyAutoGalaxy use averaging rather than integration when applying oversampling?
[Reference notebook](https://github.com/Jammy2211/autogalaxy_workspace/blob/release/notebooks/guides/advanced/over_sampling.ipynb)

In particular, when the light profile is curved, the ideal approach is to integrate the profile
over the pixel area. Since this is computationally expensive, oversampling offers an
efficient alternative, as noted in [this Astropy discussion](https://github.com/astropy/astropy/issues/11179#issuecomment-749518969).

However, according to the linked notebook, the oversampled values are averaged
rather than summed.

It is possible that I am misunderstanding either the implementation or the concept of
oversampling in this context, but I would appreciate clarification on why averaging is
used instead of integration. I raise this point because I am particularly interested in
using your routines to simulate Sérsic galaxy profiles.

##### Installation and functionability

First of all, why are the following dependencies not included in the installation requirements?

pip install autoconf
pip install autofit
pip install autoarray

Including them would simplify the installation process for new users.

The installation from source guide ([link](https://pyautogalaxy.readthedocs.io/en/latest/installation/source.html)) does not include the command:

I was unable to install the submitted version of the package and encountered the following error:

However, I was able to successfully download and install the package from the main branch of the repository.

The following command from the installation documentation did not work for me:

```
python3 -m pytest
```

It returned the error:

```
/PyAutoGalaxy-main/pyautogal_venv_2/bin/python3: No module named pytest
```

Instead, I was able to run the tests successfully using the `pytest` command alone.

However, for the package installed from the main branch, the tests did not run successfully in their entirety:

Finally, the code generally follows the recommended formatting guidelines. However, when
running `flake8`, I observed that many lines exceed the recommended maximum line length:

The tests were run on the following system:

```
❯ uname -a
Linux saturnv 6.12.10-76061203-generic #202412060638~1748542656~22.04~663e4dc SMP PREEMPT_DYNAMIC Thu M x86_64 x86_64 x86_64 GNU/Linux
```
