+++ test_autogalaxy/conftest.py
@@ -1,5 +1,6 @@
 import jax.numpy as jnp
 
+
 def pytest_configure():
     _ = jnp.sum(jnp.array([0.0]))  # Force backend init
 
