+++ autogalaxy/imaging/plot/fit_imaging_plotters.py
@@ -11,7 +11,6 @@ from autogalaxy.imaging.fit_imaging import FitImaging
 from autogalaxy.plot.abstract_plotters import Plotter
 from autogalaxy.plot.mat_plot.two_d import MatPlot2D
 from autogalaxy.plot.visuals.two_d import Visuals2D
-from autogalaxy.plot.include.two_d import Include2D
 
 
 class FitImagingPlotter(Plotter):
@@ -20,7 +19,6 @@ class FitImagingPlotter(Plotter):
         fit: FitImaging,
         mat_plot_2d: MatPlot2D = None,
         visuals_2d: Visuals2D = None,
-        include_2d: Include2D = None,
         residuals_symmetric_cmap: bool = True,
     ):
         """
@@ -32,8 +30,7 @@ class FitImagingPlotter(Plotter):
         but a user can manually input values into `MatPlot2d` to customize the figure's appearance.
 
         Overlaid on the figure are visuals, contained in the `Visuals2D` object. Attributes may be extracted from
-        the `FitImaging` and plotted via the visuals object, if the corresponding entry is `True` in the `Include2D`
-        object or the `config/visualize/include.ini` file.
+        the `FitImaging` and plotted via the visuals object.
 
         Parameters
         ----------
@@ -43,15 +40,11 @@ class FitImagingPlotter(Plotter):
             Contains objects which wrap the matplotlib function calls that make the plot.
         visuals_2d
             Contains visuals that can be overlaid on the plot.
-        include_2d
-            Specifies which attributes of the `FitImaging` are extracted and plotted as visuals.
         residuals_symmetric_cmap
             If true, the `residual_map` and `normalized_residual_map` are plotted with a symmetric color map such
             that `abs(vmin) = abs(vmax)`.
         """
-        super().__init__(
-            mat_plot_2d=mat_plot_2d, include_2d=include_2d, visuals_2d=visuals_2d
-        )
+        super().__init__(mat_plot_2d=mat_plot_2d, visuals_2d=visuals_2d)
 
         self.fit = fit
 
@@ -59,7 +52,6 @@ class FitImagingPlotter(Plotter):
             fit=self.fit,
             get_visuals_2d=self.get_visuals_2d,
             mat_plot_2d=self.mat_plot_2d,
-            include_2d=self.include_2d,
             visuals_2d=self.visuals_2d,
             residuals_symmetric_cmap=residuals_symmetric_cmap,
         )
@@ -84,7 +76,6 @@ class FitImagingPlotter(Plotter):
             inversion=self.fit.inversion,
             mat_plot_2d=self.mat_plot_2d,
             visuals_2d=self.get_visuals_2d(),
-            include_2d=self.include_2d,
         )
 
     @property
