### 2025-07-18 c967d95
ellipse fixes
