+++ autogalaxy/ellipse/model/plotter_interface.py
@@ -93,16 +93,16 @@ class PlotterInterfaceEllipse(PlotterInterface):
             fit_list=fit_list, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
         )
 
-        try:
-            fit_plotter.figures_2d(data=should_plot("data"))
-        except ValueError:
-            print(fit_plotter.fit_list[0].ellipse.major_axis)
-            print(fit_plotter.fit_list[0].ellipse.ell_comps)
+        fit_plotter.figures_2d(
+            data=should_plot("data"),
+            ellipse_residuals=should_plot("ellipse_residuals"),
+        )
 
         if should_plot("data_no_ellipse"):
             fit_plotter.figures_2d(
                 data=True,
                 disable_data_contours=True,
+
             )
 
         fit_plotter.mat_plot_2d.use_log10 = True
