+++ autogalaxy/ellipse/plot/fit_ellipse_plotters.py
@@ -1,13 +1,11 @@
-import matplotlib.pyplot as plt
 import numpy as np
-from typing import List
-
-from autoconf import conf
+import math
+from typing import List, Optional
 
 import autoarray as aa
 from autoarray import plot as aplt
 
-from autoarray.plot.auto_labels import AutoLabels
+from autogalaxy.ellipse.plot import fit_ellipse_plot_util
 
 from autogalaxy.ellipse.fit_ellipse import FitEllipse
 from autogalaxy.plot.abstract_plotters import Plotter
@@ -18,6 +16,7 @@ from autogalaxy.plot.mat_plot.two_d import MatPlot2D
 from autogalaxy.plot.visuals.two_d import Visuals2D
 from autogalaxy.plot.include.two_d import Include2D
 
+from autogalaxy.util import error_util
 
 class FitEllipsePlotter(Plotter):
     def __init__(
@@ -51,6 +50,8 @@ class FitEllipsePlotter(Plotter):
         self,
         data: bool = False,
         disable_data_contours: bool = False,
+        ellipse_residuals: bool = False,
+        for_subplot : bool = False,
         suffix: str = "",
     ):
         """
@@ -71,6 +72,7 @@ class FitEllipsePlotter(Plotter):
         filename_tag = ""
 
         if data:
+
             self.mat_plot_2d.contour = aplt.Contour(
                 manual_levels=np.sort(
                     [float(np.mean(fit.data_interp)) for fit in self.fit_list]
@@ -93,7 +95,7 @@ class FitEllipsePlotter(Plotter):
                 ellipse_list.append(aa.Grid2DIrregular.from_yx_1d(y=y, x=x))
 
             visuals_2d = self.get_visuals_2d() + Visuals2D(
-                positions=ellipse_list, lines=ellipse_list
+                lines=ellipse_list
             )
 
             self.mat_plot_2d.plot_array(
@@ -107,3 +109,139 @@ class FitEllipsePlotter(Plotter):
 
             if disable_data_contours:
                 self.mat_plot_2d.contour = contour_original
+
+        if ellipse_residuals:
+
+            try:
+                colors = self.mat_plot_2d.grid_plot.config_dict["c"]
+            except KeyError:
+                colors = "k"
+
+            fit_ellipse_plot_util.plot_ellipse_residuals(
+                array=self.fit_list[0].dataset.data.native,
+                fit_list=self.fit_list,
+                colors=colors,
+                output=self.mat_plot_2d.output,
+                for_subplot=for_subplot,
+            )
+
+    def subplot_fit_ellipse(self):
+
+        """
+        Standard subplot of the attributes of the plotter's `FitEllipse` object.
+        """
+
+        self.open_subplot_figure(number_subplots=2)
+
+        self.figures_2d(data=True)
+        self.figures_2d(ellipse_residuals=True, for_subplot=True)
+
+        self.mat_plot_2d.output.subplot_to_figure(auto_filename="subplot_fit_ellipse")
+        self.close_subplot_figure()
+
+class FitEllipsePDFPlotter(Plotter):
+    def __init__(
+        self,
+        fit_pdf_list: List[FitEllipse],
+        mat_plot_1d: MatPlot1D = MatPlot1D(),
+        visuals_1d: Visuals1D = Visuals1D(),
+        include_1d: Include1D = Include1D(),
+        mat_plot_2d: MatPlot2D = MatPlot2D(),
+        visuals_2d: Visuals2D = Visuals2D(),
+        include_2d: Include2D = Include2D(),
+        sigma: Optional[float] = 3.0,
+    ):
+        super().__init__(
+            mat_plot_1d=mat_plot_1d,
+            visuals_1d=visuals_1d,
+            include_1d=include_1d,
+            mat_plot_2d=mat_plot_2d,
[…]

[…]
