+++ autogalaxy/config/visualize/plots.yaml
@@ -49,6 +49,7 @@ fit_interferometer:                        # Settings for plots of fits to inter
   fits_dirty_images: false                 # output dirty_images.fits showing the dirty image, noise-map, model-data, resiual-map, normalized residual map and chi-squared map?
 
 fit_ellipse:                               # Settings for plots of ellipse fitting fits (e.g. FitEllipse)
+  subplot_fit_ellipse : true               # Plot subplot of all fit quantities for ellipse fits (e.g. the model data, residual-map, etc.)?
   data : true                              # Plot the data of the ellipse fit?
   data_no_ellipse: true                    # Plot the data without the black data ellipses, which obscure noisy data?
   ellipse_residuals: true                  # Plot the residuals of the ellipse fit?
