### 2025-07-21 07c2bcd
image refactor progress
