+++ autogalaxy/galaxy/plot/galaxy_plotters.py
@@ -87,20 +87,8 @@ class GalaxyPlotter(Plotter):
         self._mass_plotter = MassPlotter(
             mass_obj=self.galaxy,
             grid=self.grid,
-            get_visuals_2d=self.get_visuals_2d,
             mat_plot_2d=self.mat_plot_2d,
             visuals_2d=self.visuals_2d,
-        )
-
-    def get_visuals_1d_light(self) -> Visuals1D:
-        return self.get_1d.via_light_obj_from(light_obj=self.galaxy)
-
-    def get_visuals_1d_mass(self) -> Visuals1D:
-        return self.get_1d.via_mass_obj_from(mass_obj=self.galaxy, grid=self.grid)
-
-    def get_visuals_2d(self) -> Visuals2D:
-        return self.get_2d.via_light_mass_obj_from(
-            light_mass_obj=self.galaxy, grid=self.grid
         )
 
     def light_profile_plotter_from(
@@ -124,22 +112,23 @@ class GalaxyPlotter(Plotter):
         from autogalaxy.profiles.plot.light_profile_plotters import LightProfilePlotter
 
         if not one_d_only:
+
             return LightProfilePlotter(
                 light_profile=light_profile,
                 grid=self.grid,
                 mat_plot_2d=self.mat_plot_2d,
-                visuals_2d=self.get_2d.via_light_obj_from(
-                    light_obj=light_profile, grid=self.grid
-                ),
+                visuals_2d=self.visuals_2d,
                 mat_plot_1d=self.mat_plot_1d,
-                visuals_1d=self.get_1d.via_light_obj_from(light_obj=light_profile),
+                visuals_1d=self.visuals_1d
+                + Visuals1D().add_half_light_radius(light_obj=light_profile),
             )
 
         return LightProfilePlotter(
             light_profile=light_profile,
             grid=self.grid,
             mat_plot_1d=self.mat_plot_1d,
-            visuals_1d=self.get_1d.via_light_obj_from(light_obj=light_profile),
+            visuals_1d=self.visuals_1d
+            + Visuals1D().add_half_light_radius(light_obj=light_profile),
         )
 
     def mass_profile_plotter_from(
@@ -165,11 +154,10 @@ class GalaxyPlotter(Plotter):
                 mass_profile=mass_profile,
                 grid=self.grid,
                 mat_plot_2d=self.mat_plot_2d,
-                visuals_2d=self.get_2d.via_mass_obj_from(
-                    mass_obj=mass_profile, grid=self.grid
-                ),
+                visuals_2d=self._mass_plotter.visuals_2d_with_critical_curves,
                 mat_plot_1d=self.mat_plot_1d,
-                visuals_1d=self.get_1d.via_mass_obj_from(
+                visuals_1d=self.visuals_1d
+                + Visuals1D().add_einstein_radius(
                     mass_obj=mass_profile, grid=self.grid
                 ),
             )
@@ -178,9 +166,8 @@ class GalaxyPlotter(Plotter):
             mass_profile=mass_profile,
             grid=self.grid,
             mat_plot_1d=self.mat_plot_1d,
-            visuals_1d=self.get_1d.via_mass_obj_from(
-                mass_obj=mass_profile, grid=self.grid
-            ),
+            visuals_1d=self.visuals_1d
+            + Visuals1D().add_einstein_radius(mass_obj=mass_profile, grid=self.grid),
         )
 
     @property
@@ -246,7 +233,8 @@ class GalaxyPlotter(Plotter):
             self.mat_plot_1d.plot_yx(
                 y=image_1d,
                 x=image_1d.grid_radial,
-                visuals_1d=self.get_visuals_1d_light(),
[…]
