+++ autogalaxy/plot/get_visuals/one_d.py
@@ -1,218 +1,0 @@
-import logging
-from typing import List, Union
-
-import autoarray as aa
-import autoarray.plot as aplt
-
-from autogalaxy.galaxy.galaxy import Galaxy
-from autogalaxy.plot.visuals.one_d import Visuals1D
-from autogalaxy.profiles.light.abstract import LightProfile
-from autogalaxy.profiles.mass.abstract.abstract import MassProfile
-
-from autogalaxy.util import error_util
-
-logger = logging.getLogger(__name__)
-
-
-class GetVisuals1D(aplt.GetVisuals1D):
-    def __init__(self, visuals: Visuals1D):
-        """
-        Class which gets 1D attributes and adds them to a `Visuals1D` objects, such that they are plotted on 1D figures.
-
-        The `GetVisuals1D` class adds new visuals to a pre-existing `Visuals1D` object that is passed to its `__init__`
-        method. This only adds a new entry if the visual are not already in this object.
-
-        Parameters
-        ----------
-        visuals
-            The pre-existing visuals of the plotter which new visuals are added too via the `GetVisuals1D` class.
-        """
-        super().__init__(visuals=visuals)
-
-    def via_light_obj_from(self, light_obj: Union[LightProfile, Galaxy]) -> Visuals1D:
-        """
-        From an object with light profiles (e.g. a `LightProfile`, `Galaxy`) get its attributes that can be plotted
-        and return them  in a `Visuals1D` object.
-
-        Only attributes not already in `self.visuals` are extracted for plotting.
-
-        From a light object the following 1D attributes can be extracted for plotting:
-
-        - half_light_radius: the radius containing 50% of the light objects total integrated luminosity.
-
-        Parameters
-        ----------
-        light_obj
-            The light object (e.g. a `LightProfile`, `Galaxy`) whose attributes are extracted for plotting.
-
-        Returns
-        -------
-        Visuals1D
-            The collection of attributes that can be plotted by a `Plotter` object.
-        """
-
-        half_light_radius = self.get(
-            "half_light_radius", value=light_obj.half_light_radius
-        )
-
-        return self.visuals + self.visuals.__class__(
-            half_light_radius=half_light_radius
-        )
-
-    def via_light_obj_list_from(
-        self, light_obj_list: Union[List[LightProfile], List[Galaxy]], low_limit: float
-    ) -> Visuals1D:
-        """
-        From a list of objects with light profiles (e.g. a `LightProfile`, `Galaxy`) get its attributes that can be
-        plotted and return them  in a `Visuals1D` object.
-
-        Only attributes not already in `self.visuals` are extracted for plotting.
-
-        This function iterates over all light objects in the list and averages over each attribute's values to estimate
-        the mean value of the attribute and its error, both of which can then be plotted. This is typically used
-        to plot 1D errors on a quantity that are estimated via a Probability Density Function.
-
-        From a light object lust the following 1D attributes can be extracted for plotting:
-
-        - half_light_radius: the radius containing 50% of the light objects total integrated luminosity.
-
-        Parameters
-        ----------
-        light_obj_list
-            The list of light objects (e.g. a `LightProfile`, `Galaxy`) whose mean attributes and error estimates are
-            extracted for plotting.
-        low_limit
-            The value of sigma to which errors are estimated (e.g. 1.0 will estimate errors at the ~0.32 and ~0.68
-            intervals of the probability distribution.
-
-        Returns
-        -------
-        Visuals1D
-            The mean value and errors of each attribute that are plotted in 1D by a `Plotter` object.
-        """
-
-        half_light_radius_list = [
-            light_profile.half_light_radius for light_profile in light_obj_list
-        ]
-
-        if None in half_light_radius_list:
[…]
