### 2025-07-22 #235 @Jammy2211
Thanks for the review!

I'll get to work on it, but with holidays will likely be a bit slower to respond than normal. Nevertheless, nothing looks to difficult and getting the API docs up to scratch has been on my to do list for a while :).
