### 2025-07-23 7937028
ellipse does not trim fits output
