### 2025-07-24 10c5c3a
quirks removed
