+++ autogalaxy/ellipse/model/plotter_interface.py
@@ -46,8 +46,8 @@ class PlotterInterfaceEllipse(PlotterInterface):
             dataset_plotter.subplot_dataset()
 
         image_list = [
-            dataset.data.native_for_fits,
-            dataset.noise_map.native_for_fits,
+            dataset.data.native,
+            dataset.noise_map.native,
         ]
 
         hdu_list = hdu_list_for_output_from(
