+++ autogalaxy/galaxy/galaxy.py
@@ -229,10 +229,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
             > 0
         ):
             return sum(self.image_2d_list_from(grid=grid, operated_only=operated_only))
-
-        return np.zeros((grid.shape[0],))
-
-    #        return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def image_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -264,9 +261,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 image_1d_list.append(light_profile.image_1d_from(grid=grid_radial))
 
             return sum(image_1d_list)
-        return np.zeros((grid.shape[0],))
 
-    #   return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_vector_yx
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -294,9 +290,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                     self.cls_list_from(cls=MassProfile),
                 )
             )
-        return np.zeros((grid.shape[0], 2))
 
-    #        return jnp.zeros((grid.shape[0], 2))
+        return jnp.zeros((grid.shape[0], 2))
 
     @aa.grid_dec.to_array
     def convergence_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -324,9 +319,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 )
             )
 
-        return np.zeros((grid.shape[0],))
-
-    #        return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_grid
     def traced_grid_2d_from(self, grid: aa.type.Grid2DLike) -> aa.type.Grid2DLike:
@@ -376,9 +369,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
 
             return sum(convergence_1d_list)
 
-        return np.zeros((grid.shape[0],))
-
-    #     return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_array
     def potential_2d_from(self, grid: aa.type.Grid2DLike, **kwargs) -> np.ndarray:
@@ -405,9 +396,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                     self.cls_list_from(cls=MassProfile),
                 )
             )
-        return np.zeros((grid.shape[0],))
-
-    #        return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @aa.grid_dec.to_projected
     def potential_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
@@ -439,9 +428,8 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections):
                 )
 
             return sum(potential_1d_list)
-        return np.zeros((grid.shape[0],))
 
-    #        return jnp.zeros((grid.shape[0],))
+        return jnp.zeros((grid.shape[0],))
 
     @property
     def half_light_radius(self):
