### 2025-07-30 f3d748c
Update subplot_fit_ellipse to use log_10 for data, and remove Fill option from mat_plot_2d
