### 2025-07-30 b030cf8
Add EllipseMultipoleRelative class
