+++ autogalaxy/ellipse/ellipse/ellipse_multipole.py
@@ -1,5 +1,6 @@
 import numpy as np
 from typing import Tuple
+from autogalaxy.convert import multipole_comps_from, multipole_k_m_and_phi_m_from
 
 
 from autogalaxy.ellipse.ellipse.ellipse import Ellipse
@@ -62,6 +63,56 @@ class EllipseMultipole:
         radial = np.add(
             self.multipole_comps[1] * np.cos(self.m * (angles - ellipse.angle_radians)),
             self.multipole_comps[0] * np.sin(self.m * (angles - ellipse.angle_radians)),
+        )
+
+        x = points[:, 1] + (radial * np.cos(angles))
+        y = points[:, 0] + (radial * np.sin(angles))
+
+        return np.stack(arrays=(y, x), axis=-1)
+
+class EllipseMultipoleRelative(EllipseMultipole):
+    def __init__(
+        self,
+        m=4,
+        input_multipole_comps: Tuple[float, float] = (0.0, 0.0),
+        major_axis=1.,
+    ):
+
+        k, phi = multipole_k_m_and_phi_m_from(multipole_comps=input_multipole_comps, m=m)
+        k_adjusted = k*major_axis
+
+        adjusted_multipole_comps = multipole_comps_from(k_adjusted, phi, m)
+
+        super().__init__(m, adjusted_multipole_comps)
+
+        self.adjusted_multipole_comps = adjusted_multipole_comps
+        self.m = m
+
+    def points_perturbed_from(
+        self, pixel_scale, points, ellipse: Ellipse, n_i: int = 0
+    ) -> np.ndarray:
+        """
+        Returns the (y,x) coordinates of the input points, which are perturbed by the multipole of the ellipse.
+
+        Parameters
+        ----------
+        pixel_scale
+            The pixel scale of the data that the ellipse is fitted to and interpolated over.
+        points
+            The (y,x) coordinates of the ellipse that are perturbed by the multipole.
+        ellipse
+            The ellipse that is perturbed by the multipole, which is used to compute the angles of the ellipse.
+
+        Returns
+        -------
+        The (y,x) coordinates of the input points, which are perturbed by the multipole.
+        """
+
+        angles = ellipse.angles_from_x0_from(pixel_scale=pixel_scale, n_i=n_i)
+
+        radial = np.add(
+            self.adjusted_multipole_comps[1] * np.cos(self.m * (angles - ellipse.angle_radians)),
+            self.adjusted_multipole_comps[0] * np.sin(self.m * (angles - ellipse.angle_radians)),
         )
 
         x = points[:, 1] + (radial * np.cos(angles))
