### 2025-07-30 50fcfef
Fix declaration of input_multipole_comps
