### 2025-07-30 e8fc317
测试好了计算黑森矩阵，椭率引入得很正确
