+++ autogalaxy/profiles/mass/total/dPIE_lentool.py
@@ -4,14 +4,20 @@ import numpy as np
 import autoarray as aa
 from autogalaxy.profiles.mass.abstract.abstract import MassProfile
 
-class PIEMD(MassProfile): 
+from typing import Tuple
+import numpy as np
+
+import autoarray as aa
+from autogalaxy.profiles.mass.abstract.abstract import MassProfile
+
+class PIEMD(MassProfile):
 
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
         ell_comps: Tuple[float, float] = (0.0, 0.0),
         ra: float = 0.1,
-        b0: float = 1.0,
+        b0: float = 0.1,
     ):
         super().__init__(centre=centre, ell_comps=ell_comps)
 
@@ -19,19 +25,18 @@ class PIEMD(MassProfile):
         self.b0 = b0
 
     def _ci05(self,x, y, eps, rc):
-        print('x:',x,'y',y)
         # Calculate intermediate values
         if eps < 1e-10:
             eps = 1e-10
         sqe = np.sqrt(eps)
-        cx1 = (1.0 - eps) / (1.0 + eps)
+        axis_ratio = (1.0 - eps) / (1.0 + eps)
         cxro = (1.0 + eps) * (1.0 + eps)
         cyro = (1.0 - eps) * (1.0 - eps)
         rem2 = x * x / cxro + y * y / cyro
         
         # Define complex numbers
-        zci = np.complex128(complex(0.0, -0.5 * (1.0 - eps * eps) / sqe))
-        znum = np.complex128(cx1 * x + 1j * (2.0 * sqe * np.sqrt(rc * rc + rem2) - y / cx1))
+        zci = np.complex128(complex(0.0, -0.5 * (1.0 - eps * eps) / sqe))  # 标量复数
+        znum = np.complex128(axis_ratio * x + 1j * (2.0 * sqe * np.sqrt(rc * rc + rem2) - y / axis_ratio))
         zden = np.complex128(x + 1j * (2.0 * rc * sqe - y))
             
         # Perform complex division: zis = znum / zden
@@ -74,19 +79,103 @@ class PIEMD(MassProfile):
         """
         ellip = self._ellip()
 
+        factor = self.b0
+
         zis = self._ci05(
             x=grid[:, 1], y=grid[:, 0], eps=ellip, rc=self.ra
         )
 
         # This is in axes aligned to the major/minor axis
-        deflection_x = self.b0 * zis.real
-        deflection_y = self.b0 * zis.imag
+        deflection_x = zis.real
+        deflection_y = zis.imag
 
         # And here we convert back to the real axes
         return self.rotated_grid_from_reference_frame_from(
-            grid=np.multiply(1.0, np.vstack((deflection_y, deflection_x)).T), **kwargs
+            grid=np.multiply(factor, np.vstack((deflection_y, deflection_x)).T), **kwargs
         )
 
+    def _mdci05(self, x, y, eps, rc, b0):
+        # Prevent division by zero by setting a minimum value for eps
+        if eps < 1e-10:
+            eps = 1e-10
+
+        # Calculate intermediate values
+        sqe = np.sqrt(eps)  # Square root of ellipticity
+        cx1 = (1.0 - eps) / (1.0 + eps)  # Scaling factor for x
+        cx1inv = 1.0 / cx1  # Inverse of cx1
+        cxro = (1.0 + eps) * (1.0 + eps)  # Denominator for x^2 term in rem^2
+        cyro = (1.0 - eps) * (1.0 - eps)  # Denominator for y^2 term in rem^2
[…]
