### 2025-08-06 3e127ac
mergered from the local code and renamed the file name to pseudo_isothernal.py
