+++ autogalaxy/profiles/mass/total/dPIE_lentool.py
@@ -1,449 +1,0 @@
-from typing import Tuple
-import numpy as np
-
-import autoarray as aa
-from autogalaxy.profiles.mass.abstract.abstract import MassProfile
-
-from typing import Tuple
-import numpy as np
-
-import autoarray as aa
-from autogalaxy.profiles.mass.abstract.abstract import MassProfile
-
-class PIEMD(MassProfile):
-
-    def __init__(
-        self,
-        centre: Tuple[float, float] = (0.0, 0.0),
-        ell_comps: Tuple[float, float] = (0.0, 0.0),
-        ra: float = 0.1,
-        b0: float = 0.1,
-    ):
-        super().__init__(centre=centre, ell_comps=ell_comps)
-
-        self.ra = ra
-        self.b0 = b0
-
-    def _ci05(self,x, y, eps, rc):
-        # Calculate intermediate values
-        if eps < 1e-10:
-            eps = 1e-10
-        sqe = np.sqrt(eps)
-        axis_ratio = (1.0 - eps) / (1.0 + eps)
-        cxro = (1.0 + eps) * (1.0 + eps)
-        cyro = (1.0 - eps) * (1.0 - eps)
-        rem2 = x * x / cxro + y * y / cyro
-        
-        # Define complex numbers
-        zci = np.complex128(complex(0.0, -0.5 * (1.0 - eps * eps) / sqe))  # 标量复数
-        znum = np.complex128(axis_ratio * x + 1j * (2.0 * sqe * np.sqrt(rc * rc + rem2) - y / axis_ratio))
-        zden = np.complex128(x + 1j * (2.0 * rc * sqe - y))
-            
-        # Perform complex division: zis = znum / zden
-        norm = zden.real * zden.real + zden.imag * zden.imag  # |zden|^2
-        if np.any(norm < 1e-10):
-            raise ValueError("zden magnitude too small, risk of division by zero")
-        zis_re = (znum.real * zden.real + znum.imag * zden.imag) / norm
-        zis_im = (znum.imag * zden.real - znum.real * zden.imag) / norm
-        zis = np.complex128(zis_re + 1j * zis_im)
-        
-        # Calculate ln(zis) = ln(|zis|) + i*Arg(zis)
-        zis_mag = np.abs(zis)
-        if np.any(zis_mag < 1e-10):
-            raise ValueError("zis magnitude too small, risk of log(0)")
-        zis_re = np.log(zis_mag)
-        zis_im = np.angle(zis)
-        zis = np.complex128(zis_re + 1j * zis_im)
-        
-        # Calculate zres = zci * ln(zis)
-        zres = zci * zis
-        
-        return zres
-    
-    def _ellip(self):
-        ellip = np.sqrt(self.ell_comps[0] ** 2 + self.ell_comps[1] ** 2)
-        MAX_ELLIP = 0.99999
-        return min(ellip, MAX_ELLIP)
-    
-    @aa.grid_dec.to_vector_yx
-    @aa.grid_dec.transform
-    @aa.grid_dec.relocate_to_radial_minimum
-    def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike, **kwargs):
-        """
-        Calculate the deflection angles on a grid of (y,x) arc-second coordinates.
-
-        Parameters
-        ----------
-        grid
-            The grid of (y,x) arc-second coordinates the deflection angles are computed on.
-        """
-        ellip = self._ellip()
-
-        factor = self.b0
-
-        zis = self._ci05(
-            x=grid[:, 1], y=grid[:, 0], eps=ellip, rc=self.ra
-        )
-
-        # This is in axes aligned to the major/minor axis
[…]
