+++ autogalaxy/profiles/mass/total/pseudo_isothermal.py
@@ -1,0 +1,861 @@
+from typing import Tuple
+import numpy as np
+
+import autoarray as aa
+from autogalaxy.profiles.mass.abstract.abstract import MassProfile
+
+# Within this profile family, PIEMD, dPIEMD, and dPIEMDSph are directly ported from Lenstool's C code.
+# The dPIEP and dPIEPSph profiles are modified from the original `dPIE` implementation in autogalaxy by Jackson O'Donnell.
+# This profile family was integrated and thoroughly annotated by Hengkai Ding.
+
+def _ci05(x, y, eps, rcore):
+    """
+    Returns the first derivatives of the lens potential as complex number I'* = (∂ψ/∂x + i ∂ψ/∂y) / E0 for PIEMD at any positions (x,y), 
+    see Kassiola & Kovner(1993) Eq. 4.1.2, which is the integral of Eq. 2.3.8.
+    Note here b0(or called E0) is out of the `_ci05`.
+
+    Parameters
+    ----------
+    eps
+        The ellipticity of the corresponding profiles.
+    rcore
+        The inner core radius.
+    Returns
+    -------
+    complex
+        The value of the I'* term.
+    """
+    if eps < 1e-10:
+        eps = 1e-10
+    sqe = np.sqrt(eps)
+    axis_ratio = (1.0 - eps) / (1.0 + eps)
+    cxro = (1.0 + eps) * (1.0 + eps)
+    cyro = (1.0 - eps) * (1.0 - eps)
+    rem2 = x * x / cxro + y * y / cyro
+    ##### I'* = zres = zci * ln(zis) = zci * ln(znum / zden), see Eq. 4.1.2 #####
+
+    # Define intermediate complex variables
+    zci = np.complex128(complex(0.0, -0.5 * (1.0 - eps * eps) / sqe))
+    znum = np.complex128(axis_ratio * x + 1j * (2.0 * sqe * np.sqrt(rcore * rcore + rem2) - y / axis_ratio))
+    zden = np.complex128(x + 1j * (2.0 * rcore * sqe - y))
+        
+    # zis = znum / zden = (a+bi)/(c+di) = [(ac+bd)+(bc-ad i)] / (c^2+d^2)
+    norm = zden.real * zden.real + zden.imag * zden.imag  # |zden|^2
+    zis_re = (znum.real * zden.real + znum.imag * zden.imag) / norm
+    zis_im = (znum.imag * zden.real - znum.real * zden.imag) / norm
+    zis = np.complex128(zis_re + 1j * zis_im)
+    
+    # ln(zis) = ln(|zis|) + i*Arg(zis)
+    zis_mag = np.abs(zis)
+    zis_re = np.log(zis_mag)
+    zis_im = np.angle(zis)
+    zis = np.complex128(zis_re + 1j * zis_im)
+    
+    # I'* = zres = zci * ln(zis)
+    zres = zci * zis
+    
+    return zres
+
+def _ci05f(x, y, eps, rcore, rcut):
+    """
+    Returns the first derivatives of the lens potential as complex number I'* = (∂ψ/∂x + i ∂ψ/∂y) / (b0 * ra / (rs - ra)) for dPIEMD at any positions (x,y), 
+    which is the integral of Eq. 2.3.8 in  Kassiola & Kovner(1993). 
+    
+    Note here (b0 * ra / (rs - ra)) is out of the `_ci05f`. The only difference of integral of Eq. 2.3.8 between dPIEMD and PIEMD is the \\kappa:
+    \\kappa(r_{em})_{dPIEMD} = rs / (rs - ra) * (\\kappa_{PIEMD,ra} - \\kappa_{PIEMD,rs}).
[…]
