### 2025-08-08 8399338
Refactor EllipseMultipoleRelative to EllipseMultipoleScaled and add docs to make more intuitive
