### 2025-08-08 f9d47ab
Add option to disable the data contours in a fit ellipse subplot
