+++ autogalaxy/ellipse/plot/fit_ellipse_plotters.py
@@ -124,7 +124,7 @@ class FitEllipsePlotter(Plotter):
                 for_subplot=for_subplot,
             )
 
-    def subplot_fit_ellipse(self):
+    def subplot_fit_ellipse(self, disable_data_contours: bool = False):
         """
         Standard subplot of the attributes of the plotter's `FitEllipse` object.
         """
@@ -132,7 +132,7 @@ class FitEllipsePlotter(Plotter):
         self.open_subplot_figure(number_subplots=2)
 
         self.mat_plot_2d.use_log10 = True
-        self.figures_2d(data=True)
+        self.figures_2d(data=True, disable_data_contours=disable_data_contours)
         self.figures_2d(ellipse_residuals=True, for_subplot=True)
 
         self.mat_plot_2d.output.subplot_to_figure(auto_filename="subplot_fit_ellipse")
