+++ autogalaxy/profiles/mass/dark/abstract.py
@@ -1,3 +1,4 @@
+import jax.numpy as jnp
 import numpy as np
 from typing import Tuple
 
@@ -129,55 +130,80 @@ class AbstractgNFW(MassProfile, DarkProfile, MassProfileMGE):
         amplitude_list *= np.sqrt(2.0 * np.pi) * sigma_list
         return amplitude_list, sigma_list
 
-    def coord_func_f(self, grid_radius):
+    def coord_func_f(self, grid_radius: jnp.ndarray) -> jnp.ndarray:
+        """
+        Given an array `grid_radius` and a work array `f`, fill f[i] with
+
+          • (1 / sqrt(r_i^2 − 1)) * arccos(1 / r_i)   if r_i > 1
+          • (1 / sqrt(1 − r_i^2)) * arccosh(1 / r_i)  if r_i < 1
+          • leave f[i] unchanged if r_i == 1 (you can adjust this if you want a different convention)
+
+        This version uses only pure JAX ops, so it can be jitted or grad’ed without
+        any Python control flow on tracer values.
+        """
         if isinstance(grid_radius, float) or isinstance(grid_radius, complex):
-            grid_radius = np.array([grid_radius])
+            grid_radius = jnp.array([grid_radius])
 
-        return self.coord_func_f_jit(
-            grid_radius=np.array(grid_radius),
-            f=np.ones(shape=grid_radius.shape[0], dtype="complex64"),
-        )
+        f = jnp.ones(shape=grid_radius.shape[0], dtype="complex64")
 
-    @staticmethod
-    def coord_func_f_jit(grid_radius, f):
-        for index in range(f.shape[0]):
-            if np.real(grid_radius[index]) > 1.0:
-                f[index] = (
-                    1.0 / np.sqrt(np.square(grid_radius[index]) - 1.0)
-                ) * np.arccos(np.divide(1.0, grid_radius[index]))
-            elif np.real(grid_radius[index]) < 1.0:
-                f[index] = (
-                    1.0 / np.sqrt(1.0 - np.square(grid_radius[index]))
-                ) * np.arccosh(np.divide(1.0, grid_radius[index]))
+        # compute both branches
+        r = grid_radius
+        inv_r = 1.0 / r
 
-        return f
+        # branch for r > 1
+        out_gt = (1.0 / jnp.sqrt(r**2 - 1.0)) * jnp.arccos(inv_r)
 
-    def coord_func_g(self, grid_radius):
-        if isinstance(grid_radius, float) or isinstance(grid_radius, complex):
-            grid_radius = np.array([grid_radius])
+        # branch for r < 1
+        out_lt = (1.0 / jnp.sqrt(1.0 - r**2)) * jnp.arccosh(inv_r)
 
+        # combine: if r>1 pick out_gt, elif r<1 pick out_lt, else keep original f
+        return jnp.where(r > 1.0, out_gt, jnp.where(r < 1.0, out_lt, f))
+
+    def coord_func_g(self, grid_radius: jnp.ndarray) -> jnp.ndarray:
+        """
+        Vectorized version of the original looped `coord_func_g_jit`.
+
+        Applies conditional logic element-wise:
+          - if r > 1: g = (1 - f_r) / (r^2 - 1)
+          - if r < 1: g = (f_r - 1) / (1 - r^2)
+          - else:     g = 1/3
+
+        Parameters
+        ----------
+        grid_radius : jnp.ndarray
+            The input grid radius values.
+        f_r : jnp.ndarray
+            Precomputed values from `coord_func_f`.
+        g : jnp.ndarray
+            Output array (will be overwritten).
+
+        Returns
+        -------
+        jnp.ndarray
+            The updated `g` array.
+        """
+        # Convert single values to JAX arrays
[…]

[…]
