### 2025-10-29 #235 @hamogu
Thank you @erik-whiting for agreeing to take on reviewing. Please let me know if you need any support and what I can do to help you along. Please see https://github.com/pyOpenSci/software-submission/issues/235#issuecomment-2803589783 for instructions and links on where to find the review template etc.
Let me know if there is anything I can do to help you!
