### 2025-11-03 #235 @erik-whiting
Hey just wanted to update everyone and throw out a quick question:

I have started my review but I'm having a hard time installing the package. Both the conda and pip guides are failing for me, I've tried a few workarounds on my own but nothing is working so far. It could be something on my end but I'm just curious, is a Windows installation supported? Also, I haven't yet tried building from source.

Just updating y'all that the review is ongoing. Hoping to figure out my installation issues soon.
