### 2025-11-03 #235 @Jammy2211
Are you installing version `2025.1.18.7` or the new version (we did a PyAutoGalaxy-JAX release (2 years in the making) last week and are still cleaning up installation issues).
