### 2025-11-22 #235 @hamogu
Is there anything I can help with to move this forward? Should it work on windows? If not, should I try to find a non-windows reviewer or should I find a way to provide @erik-whiting with (remote) access to a non-windows machine?
