### 2025-11-22 #235 @erik-whiting
> Is there anything I can help with to move this forward? Should it work on windows? If not, should I try to find a non-windows reviewer or should I find a way to provide @erik-whiting with (remote) access to a non-windows machine?

hey, so sorry. I recently completed a JOSS review and in my head it was this one, so I totally thought this was over. I will get back on this review tomorrow (Sunday), my apologies. For the record, I was able to get past the installation issue so I just need to finish the a main part of the review
