### 2025-11-23 #235 @erik-whiting
- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

Hello! Sorry for the delay. Here are my notes:

There are no community guidelines. Yes, there’s a [CONTRIBUTING.md](http://contributing.md/), but no community standards to be found

Installation instructions are not found in the [README.md](http://readme.md/) but they are linked to from there, so I’m not sure how to count this check

No citation information in the README

On the overview page, you have this quote: “...using PyPI to pip install PyAutoGalaxy into your Python distribution.” This led me to believe I could run pip install PyAutoGalaxy, which doesn’t work. I suggest rewording this.

---

I think these minor things are easy enough to fix and I'm not sure the lack of installation instructions in the README should be a blocking issue since you link to the documentation. I also don't think the note about rewording the overview page should be blocking. Really, the only thing you need to add is citation information and a community guide, at least from my perspective.
