### 2026-01-09 #235 @hamogu
@Jammy2211 Happy new year! Both reviewer's reviews are in and both suggest only pretty small changes. I agree that you don't have to tick every box on PyOpenSci's template (e.g. not having installation instructions in the README but a link to them instead is fine with me). However, other suggestions and questions are certainly worth considering.

Just to make sure that you don't wait for me to do something while I wait for you to do something I want to clarify where we are in the process. To me, it seems that we (me and the reviewers) should just wait for you at this point until you tell us that you either implemented the suggestions or let us know why you disagree.

Again, I apologize for how long this has taken so far, so I certainly don't want to hurry you, but just let us know when you think we should have a final look!
