### 2026-01-22 #235 @Jammy2211
> There are no community guidelines. Yes, there’s a [CONTRIBUTING.md](http://contributing.md/), but no community standards to be found

Added the following to CONTRIBUTING.md

```
Community Guidelines
We strive to maintain a welcoming, respectful, and inclusive community. All contributors—whether opening issues, submitting pull requests, reviewing code, or participating in discussions—are expected to follow these guidelines.

Be respectful and considerate of others.
Assume good intent and be patient, especially with newcomers.
Keep feedback constructive and focused on the work, not the person.
Communicate clearly and professionally.
Respect maintainers’ time and decisions.
Harassment, discrimination, or abusive behavior of any kind will not be tolerated.

This project follows our [Code of Conduct](https://github.com/Jammy2211/PyAutoGalaxy/blob/main/CODE_OF_CONDUCT.md). By participating, you agree to uphold it.

Reporting concerns
If you experience or witness behavior that violates these guidelines or the Code of Conduct, please contact the project maintainers privately.
```

> Installation instructions are not found in the [README.md](http://readme.md/) but they are linked to from there, so I’m not sure how to count this check

As discussed, sounds like this is fine.

> No citation information in the README

Added to README

> On the overview page, you have this quote: “...using PyPI to pip install PyAutoGalaxy into your Python distribution.” This led me to believe I could run pip install PyAutoGalaxy, which doesn’t work. I suggest rewording this.

Fixed, thanks for the good spot!
