### 2026-03-21 #235 @hamogu
I looked back over @canorve's review from 2025 and @erik-whiting. I think the latter report picked up on everything that was still open from the first one, so I think this is ready to accept.

Thanks @canorve and @erik-whiting for the review and I apologize to @Jammy2211 that it has taken so long. We aim to be much faster with PyOpenSci reviews, but we clearly have a lot of room to get better.
