### 2026-03-21 #235 @hamogu
- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] Change the status tag of the issue to `6/pyOS-approved6 🚀🚀🚀`.  (You can remove all other labels; If the the package is moving on to JOSS then add a JOSS label)
- [ ] Update the header of this issue with the date accepted, the version accepted, and the accepted version DOI (from Zenodo) to the header of this issue.
- [x] Invite the package maintainer(s) and both reviewers to the pyOpenSci Slack.

---
