### 2026-03-21 #235 @hamogu
There are a few more tasks to wrap this up - I posted a checklist above. The current version is of course no longer the one that you officially submitted.
@Jammy2211 It seems fair to list your most recent version (2025.2.26.4) as the accepted version, since that has all the changes in it that came out of the review process. Do you agree?
Let me now and I'll put that in so that the script that builds the pyopensci website can render the appropriate information; I have to put that in before I set the status to "approved" or te script my fail to build the info for the website.

I also want to invade you and all reviewers and other maintainers if this package to join the PyOpenSci slack if you aren't on it yet. It's a very good place to ask for advice on packaging questions if you ever have one. (Plus same General Open Source discussion). Just let me know which email address to invite.

And, of course, PyOoenSci would also be very happy to put out a blog post that describes you packages as an example of what packages can do, but don't feel pressured - it's an offer, not a requirement.
