### 2026-03-25 #235 @Jammy2211
> @Jammy2211 It seems fair to list your most recent version (2025.2.26.4) as the accepted version, since that has all the changes in it that came out of the review process. Do you agree?

Yes, this is fine.

> I also want to invade you and all reviewers and other maintainers if this package to join the PyOpenSci slack if you aren't on it yet. It's a very good place to ask for advice on packaging questions if you ever have one. (Plus same General Open Source discussion). Just let me know which email address to invite.

Yes please, if you can add me (email on my GitHub user pagE) thatd be great.

> And, of course, PyOoenSci would also be very happy to put out a blog post that describes you packages as an example of what packages can do, but don't feel pressured - it's an offer, not a requirement.

Yes, I am happy to do that and it would juist require adopting the main page of the readthedocs.

What is the interface between PyOpenSci and astropy, I recall something about this being a process to get listed here https://www.astropy.org/affiliated/, is that possible?

Also, I would be keen for my lensing package PyAutoLens, which I think has a PyOpenSci sbumission somewhere, to begin the review process!
