### 2026-03-25 #235 @hamogu
@Jammy2211 Do you have a zenodo integrate set up so that each release gets a permanent DOI that we can link to? See https://help.zenodo.org/docs/github/enable-repository/
