### 2026-03-25 #235 @hamogu
Once we are through all the steps here (DOI, review survey, badge on your readme, post-review survey) and I set this to "accepted" it will automatically show up in the astropy affiliated listing (with 1-2 days delay because the scripts that build the underlying json data run nightly).
