### 2026-03-25 #235 @hamogu
As for PyAutoLens: You submitted that under the old Astropy system, I have it in a spreadsheet there. Once this issue is closed (PyOpenSci has a policy to do only one package per author at the same time because it's so hard to find reviewers), would you mind opening a new issue here? I promise that we'll be faster.
