### 2026-04-28 #235 @Jammy2211
Yep there is a zendo: https://zenodo.org/account/settings/github/repository/Jammy2211/PyAutoGalaxy

I cant see a zenodo release for 2025.2.26.4, I guess the integration broke for a while. Shall we just use the most recent (I've done some reelases post 2025):

https://doi.org/10.5281/zenodo.19552258

Post survey review done.

i will open a PyAutoLens isue once were done here.
